% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{asmat}
\alias{asmat}
\title{Coerces non-\code{sparseMatrix} objects as \code{matrix}}
\usage{
asmat(obj)
}
\arguments{
\item{obj}{The object under consideration.}
}
\value{
Returns a \code{matrix} or a \code{sparseMatrix}.
}
\description{
This function coerse non-\code{sparseMatrix} objects as
  \code{matrix} and keeps \code{sparseMatrix} unchanged.
}
