% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{beta.product}
\alias{beta.product}
\title{Computes vector products}
\usage{
beta.product(beta, beta.obs.hat)
}
\arguments{
\item{beta}{Bootstrap \code{beta.obs} estimator.}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel}.}
}
\value{
Returns an \eqn{n \times n} matrix.
    \item{beta.prod}{An \eqn{n \times n} matrix.}
}
\description{
This function computes the product of the two vectors. This is
  used in the \code{sigma.summation} function that computes the
  asymptotic varianace estimator.
}
\details{
Denote \eqn{\bm{\beta}} and \eqn{\hat{\bm{\beta}}_{\rm obs}} as
  the \eqn{n \times 1} vectors \code{beta} and \code{beta.obs.hat}
  respectively. This function returns the \eqn{n \times n} matrix
  \eqn{\widetilde{\bm{\beta}}\widetilde{\bm{\beta}}'} where
  \eqn{\widetilde{\bm{\beta}}} is defined as
  \eqn{\widetilde{\bm{\beta}} \equiv \bm{\beta} -
  \widehat{\bm{\beta}}_{\rm obs}}.
}
