% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{beta.r.compute}
\alias{beta.r.compute}
\title{Computes \eqn{\widehat{\bm{\beta}}^r_n}}
\usage{
beta.r.compute(
  n,
  lpmodel,
  beta.obs.hat,
  beta.tgt,
  beta.n,
  beta.star,
  omega.i,
  indicator,
  solver
)
}
\arguments{
\item{lpmodel}{A list of objects that are used in inference of linear 
programming problems. The list of objects required in the \code{dkqs} 
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{omega.i}{The matrix \eqn{\widehat{\bm{\Omega}}^i_n}, i.e. the 
regularized matrix for 
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{indicator}{A binary variable that equals to 1 for \eqn{d \geq p} and 
equals to 0 for \eqn{d < p}.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the restricted estimator 
  \eqn{\widehat{\bm{\beta}}^r_n}.
}
