% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{beta.star.qp}
\alias{beta.star.qp}
\title{Computes the starred components of \eqn{\widehat{\bm{\beta}}}}
\usage{
beta.star.qp(
  data,
  lpmodel,
  beta.tgt,
  weight.mat,
  beta.obs.hat,
  beta.sigma,
  solver
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{weight.mat}{Weighting matrix.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}
}
\value{
Returns the vector \eqn{\widehat{\bm{\beta}}_n^\star}.
  \item{beta.star}{The vector \eqn{\widehat{\bm{\beta}}_n^\star}.}
}
\description{
This function computes the vector
  \eqn{\widehat{\bm{\beta}}_n^\star} for the case where \eqn{d<p} in the
  FSST test.
}
\details{
This corresponding to solving problem (3) of Torgovitsky (2020).
}
