% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{beta_bs}
\alias{beta_bs}
\title{Computes the test statistics via bootstrapping}
\usage{
beta_bs(
  df,
  bs_seed,
  bs_num,
  J,
  s_star,
  A_obs,
  A_tgt,
  func_obs,
  beta_obs_hat,
  beta_tgt,
  tau,
  N
)
}
\arguments{
\item{df}{The dataframe that contains the sample data.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_num}{The total number of bootstraps.}

\item{J}{The number of distinct nonzero values in vector \eqn{\bm{y}}.}

\item{s_star}{The value of \eqn{s^\ast} in the cone-tightening procedure.}

\item{A_obs}{The observed matrix in the linear program.}

\item{A_tgt}{The "target matrix" in the linear program.}

\item{func_obs}{The function that generates the required 
\eqn{\beta_{\mathrm{obs}}}.}

\item{beta_tgt}{The value of \eqn{t} in the null hypothesis.}

\item{tau}{The RHS vector fo the linear constraints.}

\item{beat_obs_hat}{The value of \eqn{\hat{\beta}_{\mathrm{obs}}} based on
the function supplied by the user.}
}
\value{
Returns the list of test statistics obtained from bootstrap, i.e.
   \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.
}
\description{
This function computes the test statistics via bootstrapping.
}
