% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{beta_bs_parallel}
\alias{beta_bs_parallel}
\title{Computes the bootstrap test statistics with parallelization}
\usage{
beta_bs_parallel(
  data,
  R,
  J,
  s_star,
  A_obs,
  A_tgt,
  func_obs,
  beta_obs_hat,
  beta_tgt,
  tau,
  n,
  solver,
  cores,
  progress
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{R}{The total number of bootstraps \eqn{B} to be conducted.}

\item{J}{The number of distinct nonzero values in vector \eqn{\bm{y}}.}

\item{s_star}{The value of 
\eqn{\hat{s}^\ast \equiv A_{\mathrm{obs}}\hat{\bm{x}}_n^\ast} 
in the cone-tightening procedure.}

\item{A_obs}{The "observed matrix" in the inference \eqn{A_{\mathrm{obs}}}.}

\item{A_tgt}{The "target matrix" in the inference \eqn{A_{\mathrm{tgt}}}.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta_obs_hat}{The value of \eqn{\hat{\beta}_{\mathrm{obs}}} based on
the function supplied by the user.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{tau}{The value of tau chosen by the user.}

\item{n}{The number of observations in the dataframe.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop. 
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the inference procedure. If it is set as \code{TRUE}, the 
messages are displayed throughout the procedure. Otherwise, the messages
will not be displayed.}
}
\value{
Returns the list of estimates from bootstrap:
  \item{T_bs}{A list of bootstrap test statistics 
     \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.}
 \item{beta_bs_bar_set}{A list of \eqn{\tau_n}-tightened recentered 
    bootstrap estimates \eqn{\bar{\beta}^\ast_{\mathrm{obs},n,b}}.}
 \item{pb}{Progress bar object.}
}
\description{
This function computes the bootstrap test statistics with 
   parallelized for-loops.
}
