% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{bisec.eval}
\alias{bisec.eval}
\title{Evaluation of test statistic and check if the point has been evaluated}
\usage{
bisec.eval(f, farg, pt, pvals, rngstate, para.match)
}
\arguments{
\item{f}{The function that represents a testing procedure.}

\item{farg}{The list of arguments to be passed to the function of testing
procedure.}

\item{pt}{The point to be evaluated in the bisection method.}

\item{pvals}{The data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{rngstate}{The current RNG state obtained from \code{.Random.seed}.}

\item{para.match}{The list of parameters to be matched.}
}
\value{
Returns the \eqn{p}-value of the point considered and an updated
   data frame that contains the points and the \eqn{p}-values.
   \item{pval}{The \eqn{p}-value that corresponds to the test point.}
   \item{pvals}{The updated data frame that contains the \eqn{p}-values that
     correspond to each set of tuning parameters and each test point.}
}
\description{
This function is used in the \code{\link[lpinfer]{invertci}}
procedure. It checks if the \eqn{p}-value for the point considered has
already been evaluated in previous iterations or provided by the user. The
function will compute the \eqn{p}-value if it is not provided. Otherwise,
it will use the previous data.
}
