% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{bisec.print}
\alias{bisec.print}
\title{Print messages in bisection procedure and store results}
\usage{
bisec.print(procedure, alpha, returnlist, a, b, progress, dp, df_bis)
}
\arguments{
\item{procedure}{The variable indicating whether the function is evaluating
the end-points or first mid-point, or is iterating through the bisection
procedure.}

\item{alpha}{The significance level(s). This can be a vector.}

\item{returnlist}{The list of information obtained from running
\code{\link[lpinfer]{bisec.eval}}.}

\item{a}{The lower bound of the current interval. This is \code{NULL} if
the initial end-points are being evaluated.}

\item{b}{The upper bound of the current interval. This is \code{NULL} if
the initial end-points are being evaluated.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the procedure of constructing confidence interval. If
it is set as \code{TRUE}, the messages are displayed throughout the
procedure. Otherwise, the messages will not be displayed.}

\item{dp}{The number of decimal places in the output.}

\item{df_bis}{A data frame storing the information from the bisection
method.}
}
\value{
Return the updated data frame that stores the information for the
   the iteration.
   \item{df_bis}{The data frame storing the information for the bisection
      method in each iteration.}
}
\description{
This function prints the information in
  \code{\link[lpinfer]{ci.bisection}} and store the result for each
  iteration.
}
