% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bs.assign}
\alias{bs.assign}
\title{Auxiliary function in the bootstrap replications}
\usage{
bs.assign(
  R,
  R.eval,
  R.succ,
  maxR,
  any.list,
  lpmodel,
  data = NULL,
  m = NULL,
  replace = TRUE
)
}
\arguments{
\item{R}{The number of bootstrap replications.}

\item{R.eval}{The number of bootstrap replications that has been evaluated.}

\item{R.succ}{The number of successful bootstrap replications.}

\item{maxR}{The maximum number of bootstrap replications.}

\item{any.list}{An object returned from the
\code{\link[lpinfer]{lpmodel.anylist}} function, which indicates
whether at least one of the components in \code{lpmodel} is a \code{list}.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{data}{An \code{data frame} or a \code{matrix}.}

\item{m}{The size of each subsample.}

\item{replace}{A boolean variable to indicate whether the function samples
the data with or without replacement.}
}
\value{
Returns the following three objects.
  \item{i0}{The starting index to be evaluated.}
  \item{i1}{The last index to be evaluated.}
  \item{bs.list}{The list of bootstrap replications in the \code{lpmodel}
    object passed from the user.}
}
\description{
This function is used in the bootstrap replications of the
  procedures. It returns the indices to be considered in the
  bootstrap replications and the list to be passed to the
  \code{\link[lpinfer]{lpmodel.anylist}} function.
}
