% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.Ab}
\alias{check.Ab}
\title{Check function: constraint matrix and the corresponding rhs vector}
\usage{
check.Ab(A, b, Aname, bname)
}
\arguments{
\item{A}{The constraint matrix.}

\item{b}{The corresponding rhs vector.}

\item{Aname}{The name of the constraint matrix.}

\item{bname}{The name of the corresponding rhs vector.}
}
\value{
Returns the updated matrix and rhs vector or prints a
   stop message.
   \item{A}{The updated constraint matrix.}
   \item{b}{The updated rhs vector.}
}
\description{
This function checks the constraint matrix \eqn{\bm{A}} and
   the corresponding vector \eqn{\bm{\beta}}.
}
