% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.boolean}
\alias{check.boolean}
\title{Check function: boolean variable}
\usage{
check.boolean(x, name.var)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name.var}{Name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the variable is boolean. If not,
  an error message is displayed.
}
