% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.cores}
\alias{check.cores}
\title{Check function: check the number of cores}
\usage{
check.cores(x)
}
\arguments{
\item{cores}{Number of cores.}
}
\value{
Returns the updated number of cores.
  \item{cores}{Updated number of cores}
}
\description{
This function checks if the number of cores specified is a
   positive integer. If not, set it as one.
}
