% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.dataframe}
\alias{check.dataframe}
\title{Check function: data}
\usage{
check.dataframe(data)
}
\arguments{
\item{data}{A \code{data} object to be checked.}
}
\value{
Returns the object as a \code{data.frame}.
}
\description{
This function force the data input as a \code{data.frame}.
}
