% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.datafunction}
\alias{check.datafunction}
\title{Check function: passing data to function}
\usage{
check.datafunction(data, f, lpmodel.comp)
}
\arguments{
\item{data}{The data object in the class \code{data.frame}.}

\item{f}{A user-defined function in \code{lpmodel}.}

\item{lpmodel.comp}{The name of the \code{lpmodel} object.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether there will be any error if an
  object of class \code{data.frame} is passed to the user-defined function.
}
\details{
An error message will be displayed.
}
