% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.func}
\alias{check.func}
\title{Check function: function}
\usage{
check.func(f, A, data, name.var, name.A, mat.type)
}
\arguments{
\item{f}{The function to be tested.}

\item{A}{A matrix that contains the information about the dimension for the
output of f.}

\item{data}{A data frame.}

\item{name.var}{The name of the variable.}

\item{name.A}{The name of the matrix \code{A}.}

\item{mat.type}{The type of the matrix to be checked.}
}
\value{
Returns the output of the function in the format of \code{numeric}
  and the correct dimension.
  \itemize{
     \item If \code{mat.type} is \code{col}, then a column vector is
     returned.
     \item If \code{mat.type} is \code{matrix}, then a square matrix is
     returned.}
}
\description{
This function checks whether the class of the function
  provided by the user is \code{function}, and whether the output
  produced by the function is \code{numeric} with the correct dimension.
  If not, an error message is displayed.
}
