% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.lpobjects}
\alias{check.lpobjects}
\title{Check function: matrices and vectors in \code{lpmodel}}
\usage{
check.lpobjects(data, mat, mat.name, mat.cat, R)
}
\arguments{
\item{data}{Data used in the tests.}

\item{mat}{The matrix object in \code{lpmodel}.}

\item{mat.name}{Name of the matrix object in \code{lpmodel}.}

\item{mat.cat}{Category of the matrix object.}

\item{R}{Number of bootstrap replications.}
}
\value{
Returns two objects:
   \item{mat}{The updated object in \code{lpmodel}.}
   \item{sample}{A sample object that is used to compare the dimension.}
}
\description{
This function checks if the matrix objects in \code{lpmodel}
   are in the correct format.
}
\details{
See the details section for the function \code{check.lpmodel}
  for more details on the strings for each category.
}
