% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.matrix}
\alias{check.matrix}
\title{Check function: matrix}
\usage{
check.matrix(mat, mat.name, mat.cat, inside.list)
}
\arguments{
\item{mat}{The matrix object in \code{lpmodel}.}

\item{mat.name}{The name of the matrix object in \code{lpmodel}.}

\item{mat.cat}{The category of the matrix object.}

\item{inside.list}{An indicator variables of whether the object that is
being checked is inside a list or not.}
}
\value{
Nothing is returned.
}
\description{
This function checks if the matrix objects in the
   \code{lpmodel} object are in the correct format.
}
