% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.nonnegaetive}
\alias{check.nonnegaetive}
\title{Check function: nonnegative number}
\usage{
check.nonnegaetive(x, name.var)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name.var}{Name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether variable satisfies the following
  requirements:
  \itemize{
    \item{The class of the variable is \code{numeric}.}
    \item{The length of the variable is 1.}
    \item{The variable is nonnegative.}
  }
}
