% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.numeric}
\alias{check.numeric}
\title{Check function: numeric}
\usage{
check.numeric(x, name.var)
}
\arguments{
\item{x}{The variable to be checked.}

\item{name.var}{The name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the class of the variable
  is \code{numeric} and has length 1. If not, an error message is displayed.
}
