% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.numrange}
\alias{check.numrange}
\title{Check function: range of a variable}
\usage{
check.numrange(x, name.var, left.type, left, right.type, right)
}
\arguments{
\item{x}{The variable to be checked.}

\item{name.var}{The name of the variable.}

\item{left.type}{The type of the left interval (\code{open} or
\code{closed}).}

\item{left}{The value of lower bound.}

\item{right.type}{The type of the right interval (\code{open} or
\code{closed}).}

\item{right}{the value of the upper bound.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the variable is within a
  certain interval. If not, an error message is displayed.
}
