% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.positive}
\alias{check.positive}
\title{Check function: positive number}
\usage{
check.positive(x, name.var)
}
\arguments{
\item{x}{The variable to be checked.}

\item{name.var}{The name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether variable satisfies the following
  requirements:
  \itemize{
    \item{The class of the variable is \code{numeric}.}
    \item{The length of the variable is 1.}
    \item{The variable is strictly positive.}
  }
  If not, an error message is displayed.
}
