% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.positiveinteger}
\alias{check.positiveinteger}
\title{Check function: positive integer}
\usage{
check.positiveinteger(x, name.var)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name.var}{Name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the class of the variable 
  is \code{numeric}, has length 1 and is a positive integer. If not, an
  error message is displayed.
}
