% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.samplesize}
\alias{check.samplesize}
\title{Check function: sample size \code{n} if \code{data} is \code{NULL}}
\usage{
check.samplesize(x, name.var)
}
\arguments{
\item{x}{The variable to be checked.}

\item{name.var}{The name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks the sample size \code{n} if \code{data}
  is \code{NULL}. This function is only used if \code{data} is \code{NULL}.
  Typically, when \code{data} is \code{NULL}, it refers to the case where
  the bootstrap replications are all passed to the \code{lpmodel} object.
}
