% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.solver}
\alias{check.solver}
\title{Check function: solvers}
\usage{
check.solver(x, name.var, norm = 2, qc = FALSE)
}
\arguments{
\item{x}{The variable to be checked.}

\item{name.var}{The name of the variable.}

\item{norm}{The norm used (if applicable).}

\item{qc}{An indicator of whether the problem consists of a quadratic
constraint.}
}
\value{
Returns the function name that corresponds to the solver for the
  problem.
   \item{solver}{The function for the solver used.}
   \item{solver.name}{The name of the solver used in lower case.}
}
\description{
This function checks the solver used is supported and is
  appropriate for the testing function. If not, an error message is
  displayed.
}
