% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_Ab}
\alias{check_Ab}
\title{Check function: Constraint matrix and the corresponding RHS vector}
\usage{
check_Ab(A, b, Aname, bname)
}
\arguments{
\item{A}{Constraint matrix.}

\item{b}{Corresponding vector.}

\item{Aname}{Variable name of constraint matrix.}

\item{bname}{Varaible name of the corresponding vector.}
}
\value{
Returns the updated matrix and vector or print stop message.
   \item{A_updated}{Updated constraint matrix.}
   \item{b_updated}{Updated rhs vector.}
}
\description{
This function checks the constraint matrix \eqn{\bm{A}} and 
   the corresponding vector \eqn{\bm{\beta}}.
}
