% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_boolean}
\alias{check_boolean}
\title{Check function: boolean variable}
\usage{
check_boolean(x, name_var)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name_var}{Name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the variable is boolean. If not, 
  an error message is displayed.
}
