% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_func}
\alias{check_func}
\title{Check function: function}
\usage{
check_func(f, A, df, name_var, name_A, mat_type)
}
\arguments{
\item{f}{Function to be tested}

\item{A}{Matrix that will be contains dimensional information for the 
output of f}

\item{df}{Data frame that will be passed to the function}

\item{name_var}{Name of the variable.}

\item{name_A}{Name of matrix}

\item{mat_type}{Type of the matrix to be checked}
}
\value{
Returns the output of the function in the format of \code{numeric}
  and the correct dimension. If \code{mat_type} is \code{col}, then a 
  column vector is returned. If \code{mat_type} is \code{matrix}, then a
  square matrix is returned.
}
\description{
This function checks the class of the function provided by 
  the user is \code{function}, and whether the output produced by the 
  function is \code{numeric} and has the correct dimension. If not, an 
  error message is displayed.
}
