% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_norm}
\alias{check_norm}
\title{Check function: norm}
\usage{
check_norm(x, name_var)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name_var}{Name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the the norm used in the
  problem is an L1 or L2 norm. If not, an error message is displayed.
}
