% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_numeric}
\alias{check_numeric}
\title{Check function: numeric}
\usage{
check_numeric(x, name_var)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name_var}{Name of the variable.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the class of the variable 
  is \code{numeric} and has length 1.
}
