% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_numrange}
\alias{check_numrange}
\title{Check function: number within a range}
\usage{
check_numrange(x, name_var, left_type, left, right_type, right)
}
\arguments{
\item{x}{Variable to be checked.}

\item{name_var}{Name of the variable.}

\item{left_type}{Type of the left interval (\code{open} or \code{closed})}

\item{left}{Value of lower bound}

\item{right_type}{Type of the right interval (\code{open} or \code{closed})}

\item{right}{Value of the upper bound}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the variable is within a 
  certain interval. If not, an error message is displayed.
}
