% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.bs}
\alias{chorussell.bs}
\title{Bootstrap procedure for the \code{\link[lpinfer]{chorussell}} procedure}
\usage{
chorussell.bs(
  data,
  lpmodel,
  beta.tgt,
  R,
  maxR,
  kappa,
  norm,
  n,
  estimate,
  solver,
  progress
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{maxR}{The maximum number of bootstrap replications to be considered in
case there are some errors.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{estimate}{A boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns a list of output that are obtained from the Cho-Russell
  procedure:
  \item{ub.bs}{The list of upper bounds from bootstrap data.}
  \item{lb.bs}{The list of lower bounds from bootstrap data.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{error.list}{A list of error messages.}
  \item{R.eval}{The number of bootstrap replications that have been
    conducted.}
  \item{R.succ}{The number of successful bootstrap replications.}
}
\description{
This function carries out the bootstrap procedure of the
  \code{\link[lpinfer]{chorussell}} procedure. This function supports
  parallel programming via the \code{future.apply} package.
}
