% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.eval}
\alias{chorussell.eval}
\title{Computes the required object in the \code{\link[lpinfer]{chorussell}}
procedure}
\usage{
chorussell.eval(
  beta.tgt,
  lb.can1,
  lb.can2,
  ub.can1,
  ub.can2,
  n,
  R,
  ci,
  alpha,
  tol,
  ub,
  lb,
  logical.ub,
  logical.lb,
  remove.const,
  kappa,
  progress,
  df.lb = NULL,
  df.ub = NULL
)
}
\arguments{
\item{beta.tgt}{The value to be tested.}

\item{lb.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb}\right)}.}

\item{lb.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb} -
\Delta \right)}.}

\item{ub.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub}\right)}.}

\item{ub.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub} +
\Delta \right)}.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{R}{The number of bootstrap replications.}

\item{ci}{A boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{alpha}{The significance level. This can be a vector.}

\item{tol}{The tolerance level in the bisection procedure.}

\item{ub}{The sample upper bound.}

\item{lb}{The sample lower bound.}

\item{logical.ub}{The logical upper bound.}

\item{logical.lb}{The logical lower bound.}

\item{remove.const}{A boolean variable. This determine whether the
constraints are to be removed.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Depending on \code{ci}, this function either returns the
  \eqn{p}-value or the \eqn{(1-\alpha)}-confidence interval.
}
\description{
This function computes the required object in the
  \code{\link[lpinfer]{chorussell}} procedure. If \code{ci} is \code{TRUE},
  this function returns the \eqn{(1-\alpha)}-confidence interval. Otherwise,
  it returns the \eqn{p}-value.
}
