% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.lp.fn}
\alias{chorussell.lp.fn}
\title{Computes whether the candidate bounds satisfy the constraints in the
\code{\link[lpinfer]{chorussell}} procedure when the sample estimates are
non-infinite}
\usage{
chorussell.lp.fn(
  x,
  lb.can1,
  lb.can2,
  ub.can1,
  ub.can2,
  ub.can,
  alpha,
  pbar,
  ci,
  kappa,
  k,
  progress
)
}
\arguments{
\item{x}{A candidate lower bound.}

\item{lb.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb}\right)}.}

\item{lb.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb} -
\Delta \right)}.}

\item{ub.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub}\right)}.}

\item{ub.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub} +
\Delta \right)}.}

\item{ub.can}{All candidates of the upper bounds.}

\item{alpha}{The significance level. This can be a vector.}

\item{pbar}{The progress bar object.}

\item{ci}{A boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{k}{Iteration number.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns a data frame.
  \item{df}{A data frame that contains the pairs of feasible lower bounds
    and upper bounds, as well as their sum.}
}
\description{
This function carries out one set of computation in the
  \code{\link[lpinfer]{chorussell}} procedure to check whether the set of
  values satisfy the constraints in the \code{\link[lpinfer]{chorussell}}
  procedure. In particular, given one candidate of the lower bound, it
  checks all possible pairs with the upper bounds and determine whether they
  satisfy the constraints.
}
