% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.lp.fn.unbd}
\alias{chorussell.lp.fn.unbd}
\title{Computes whether the candidate bounds satisfy the constraints in the
Cho-Russell procedure when one of the sample estimates of the bounds is
infinite}
\usage{
chorussell.lp.fn.unbd(
  x,
  lb.can1,
  lb.can2,
  ub.can1,
  ub.can2,
  lb.can,
  ub.can,
  lb,
  ub,
  alpha,
  pbar,
  ci,
  k,
  progress
)
}
\arguments{
\item{x}{A candidate lower bound.}

\item{lb.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb}\right)}.}

\item{lb.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb} -
\Delta \right)}.}

\item{ub.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub}\right)}.}

\item{ub.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub} +
\Delta \right)}.}

\item{ub.can}{All candidates of the upper bounds.}

\item{alpha}{Significance level.}

\item{pbar}{Progress bar object.}

\item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns a data frame.
  \item{df}{A data frame that contains the pairs of feasible lower bounds
    and upper bounds, as well as their sum.}
}
\description{
This function computes whether the candidate bounds satisfy
  the constraints in the Cho-Russell procedure when one of the sample
  estimates of the bounds is infinite. If the upper bound is infinite,
  then the third column of the data frame will store the negative of the
  lower bound so the largest candidate for the lower bound will be chosen
  in order to minimize the interval length.
}
