% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.lp}
\alias{chorussell.lp}
\title{Computes the \eqn{(1-\alpha)}-confidence interval in the Cho-Russell
procedure}
\usage{
chorussell.lp(
  lb.can1,
  lb.can2,
  ub.can1,
  ub.can2,
  n,
  R,
  alpha,
  ub,
  lb,
  remove.const,
  ci,
  kappa,
  k = 0,
  progress
)
}
\arguments{
\item{lb.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb}\right)}.}

\item{lb.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm lb} - \hat{\theta}_{\rm lb} -
\Delta \right)}.}

\item{ub.can1}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub}\right)}.}

\item{ub.can2}{The vector of values that corresponds to
\eqn{\sqrt{n}\left(\hat{\theta}^b_{\rm ub} - \hat{\theta}_{\rm ub} +
\Delta \right)}.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{R}{Number of bootstrap replications.}

\item{alpha}{Significance level. This can be a vector.}

\item{remove.const}{Determine whether the constraints are to be removed.}

\item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{kappa}{Tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the following list of objects:
  \item{bd}{A vector that represents the \eqn{(1-\alpha)}-confidence
    interval.}
  \item{unique}{Indicator variable of whether the solution is unique.}
}
\description{
This function computes the \eqn{(1-\alpha)}-confidence
interval in the Cho-Russell procedure by solving the optimization problem.
}
