% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell}
\alias{chorussell}
\title{Conducts inference using the Cho-Russell procedure}
\usage{
chorussell(
  data = NULL,
  lpmodel,
  beta.tgt,
  n = NULL,
  R = 100,
  Rmulti = 1.25,
  kappa = 1e-05,
  norm = 2,
  estimate = TRUE,
  solver = NULL,
  ci = FALSE,
  alpha = 0.05,
  tol = 1e-04,
  progress = TRUE,
  remove.const = TRUE
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test that consists of
the following objects:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{R}{Number of bootstrap replications.}

\item{Rmulti}{Multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{kappa}{Tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{estimate}{Boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{alpha}{Significance level.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{remove.const}{Determine whether the constraints are to be removed.}
}
\value{
Returns the following list of output:
  \item{ub}{Upper bound from original data.}
  \item{lb}{Lower bound from original data.}
  \item{ub.bs}{List of upper bounds from bootstrap data.}
  \item{lb.bs}{List of lower bounds from bootstrap data.}
  \item{test.logical}{Indicator variable for whether the computation has
    been conducted. If \code{test.logical} is 1, it refers to the case
    where \code{beta.tgt} is inside the logical bound. If
    \code{test.logical} is 0, it refers to the case where
    \code{beta.tgt} is outside the logical bound.}
  \item{logical.lb}{Logical lower bound.}
  \item{logical.ub}{Logical upper bound.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{Number of successful bootstrap replications.}
  \item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
    \eqn{(1-\alpha)}-confidence interval is returned.}
  \item{pval}{\eqn{p}-value (if \code{ci} is set as \code{FALSE}).}
  \item{c.ub}{Upper bound of the \eqn{(1-\alpha)}-confidence interval.}
  \item{c.lb}{Lower bound of the \eqn{(1-\alpha)}-confidence interval.}
  \item{alpha}{Significance level.}
  \item{iter}{Total number of iterations (if \code{ci} is \code{FALSE}.)}
  \item{unique}{Whether the solution is unique.}
}
\description{
This function conducts inference using the procedure
  proposed by Cho and Russell (2019).
}
\details{
See the details section of \code{\link[lpinfer]{estbounds}} for a
  list of strings acceptable for the option \code{norm}.
}
