% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.simp.fn}
\alias{chorussell.simp.fn}
\title{Checks one candidate in \code{\link[lpinfer]{chorussell}}}
\usage{
chorussell.simp.fn(x, can1, pbar, bd, progress)
}
\arguments{
\item{x}{One candidate of the upper or lower bound.}

\item{can1}{This refers to either \code{lb.can1} or \code{ub.can1}.}

\item{pbar}{The progress bar object.}

\item{bd}{A string indicating which set of candidates that the function is
trying to simplify.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the decision to drop or not to drop the candidate. If the
  decision is to drop the candidate bound, \code{NULL} is returned.
  Otherwise, the bound is returned back via the \code{future_lapply}
  function.
}
\description{
This function checks one candidate for the
  \eqn{(1-\alpha)}-confidence interval to see whether it should be dropped
  from the list of considerations. For details, please see the description
  of the \code{\link[lpinfer]{chorussell.simp}} function.
}
