% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{ci.inout}
\alias{ci.inout}
\title{Determine whether a point is inside the confidence interval or not}
\usage{
ci.inout(pval, alpha, type)
}
\arguments{
\item{pval}{The \eqn{p}-value of the test statistic.}

\item{alpha}{The significance level(s). This can be a vector.}

\item{type}{The type of the confidence interval. Set \code{type} as 1 for the
upper bound of the confidence interval and set \code{type} as -1 for the
lower bound of the confidence interval.}
}
\value{
Returns whether the part of the interval to be selected in the
   next iteration of the bisection method.
   \item{part}{A string indicating whether the left or right segment of the
   interval is chosen.}
}
\description{
This function determines whether the bisection method is going
   to be updated by choosing the left segment or the right segment of the
   interval as the updated interval for the next iteration.
}
