% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtable.R
\name{construct.cv.table}
\alias{construct.cv.table}
\title{General function to create a table of critical values}
\usage{
construct.cv.table(
  param,
  param.name,
  ts.sample.list,
  ts.bs.list,
  levels = c(0.99, 0.95, 0.9),
  digits = 5
)
}
\arguments{
\item{param}{Tuning parameter of the testing procedure.}

\item{param.name}{Name of the tuning parameter.}

\item{ts.sample.list}{A vector or sample test statistics.}

\item{ts.bs.list}{A vector or a list of bootstrap test statistics.}

\item{levels}{Confidence levels.}

\item{digits}{Number of digits in the test statistics.}
}
\value{
Returns a table of critical values.
  \item{cv.table}{Table of critical values.}
}
\description{
This function generates a table of critical values for the
  testing methods. By default, the table includes the sample test statistic,
  and critical values at the 90\%, 95\% and 99\% level. The table can be
  retrieved from \code{cv.table} in the output of the functions.
}
