% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{decimal_places}
\alias{decimal_places}
\title{Auxiliary function: Count decimal places}
\usage{
decimal_places(x)
}
\arguments{
\item{x}{Real number.}
}
\value{
Returns the number of decimal places of the number.
   \item{dp}{Number of decimal places of the number \code{x}.}
}
\description{
This function returns the number of decimal places of a real 
   snumber.
}
