% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs.bs.fn}
\alias{dkqs.bs.fn}
\title{Carries out one bootstrap replication for the DKQS procedure}
\usage{
dkqs.bs.fn(
  x,
  data,
  lpmodel,
  beta.obs.hat,
  beta.tgt,
  s.star.list,
  tau.list,
  solver,
  n
)
}
\arguments{
\item{x}{This is either the list of indices that represent the bootstrap
replications, or the list of bootstrap components of the \code{lpmodel}
object passed from the user.}

\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{beta.obs.hat}{The value of sample \eqn{\hat{\bm{\beta}}_{\mathrm{obs}}}
from the \code{lpmodel} object.}

\item{beta.tgt}{Value of beta to be tested.}

\item{s.star.list}{The list of values of
\eqn{\hat{\bm{s}}^\star \equiv \bm{A}_{\mathrm{obs}}\hat{\bm{x}}_n^\star}
in the cone-tightening procedure for each \eqn{\tau}.}

\item{tau.list}{The list of feasible parameters \eqn{\tau}.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}
}
\value{
Returns a list of output that are obtained from the DKQS
  procedure:
  \item{Ts}{Bootstrap test statistic.}
  \item{beta}{Bootstrap estimator.}
  \item{param}{List of problematic parameters in the DKQS test.}
  \item{msg}{Error message (if applicable).}
}
\description{
This function carries out the one bootstrap replication of the
  DKQS procedure This function is used in the \code{dkqs.bs} function via
  the \code{future_lapply} command.
}
