% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs.bs}
\alias{dkqs.bs}
\title{Bootstrap procedure for the DKQS procedure}
\usage{
dkqs.bs(
  data,
  lpmodel,
  beta.tgt,
  R,
  maxR,
  s.star.list,
  tau.list,
  solver,
  progress,
  n
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{maxR}{The maximum number of bootstrap replications to be considered in
case there are some errors.}

\item{s.star.list}{The list of values of
\eqn{\hat{\bm{s}}^\star \equiv \bm{A}_{\mathrm{obs}}\hat{\bm{x}}_n^\star}
in the cone-tightening procedure for each \eqn{\tau}.}

\item{tau.list}{The list of feasible parameters \eqn{\tau}.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}
}
\value{
Returns the list of estimates from bootstrap:
  \item{T.bs}{A list of bootstrap test statistics
     \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.}
 \item{beta.bs.bar.list}{A list of \eqn{\tau_n}-tightened recentered
    bootstrap estimates \eqn{\bar{\bm{\beta}}^\star_{\mathrm{obs},n,b}}.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{The number of successful bootstrap replications.}
}
\description{
This function carries out the bootstrap procedure of the
  \code{\link[lpinfer]{dkqs}} procedure This function supports parallel
  programming via the \code{furrr} package.
}
