% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs.check}
\alias{dkqs.check}
\title{Checks and updates the input in \code{\link[lpinfer]{dkqs}}}
\usage{
dkqs.check(data, lpmodel, beta.tgt, R, Rmulti, tau, n, solver, progress)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{Rmulti}{The multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{tau}{The value of the tuning parameter \eqn{\tau} in the DKQS
procedure. This can be a vector.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the updated parameters back to the
\code{\link[lpinfer]{dkqs}} function. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{test.logical}}
      \item{\code{lb}}
      \item{\code{ub}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{dkqs}} function. If there is any invalid input,
   the function will be terminated and error messages will be printed.
}
