% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs.infeasible.tau}
\alias{dkqs.infeasible.tau}
\title{Function to print the infeasible taus}
\usage{
dkqs.infeasible.tau(tau.infeasible)
}
\arguments{
\item{tau.infeasible}{A vector of infeasible taus.}
}
\value{
Nothing is returned.
}
\description{
This function is used to print the list of infeasible taus
  in the \code{print} and \code{summary} functions for \code{dkqs}.
}
