% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs_cone_check}
\alias{dkqs_cone_check}
\title{Checks and update the input}
\usage{
dkqs_cone_check(
  df,
  A_obs,
  A_tgt,
  func_obs,
  beta_tgt,
  bs_seed,
  bs_num,
  p_sig,
  tau_input,
  lpsolver,
  qpsolver
)
}
\arguments{
\item{df}{The dataframe that contains the sample data.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the value 
of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_num}{The total number of bootstraps \eqn{B}.}

\item{p_sig}{The number of decimal places in the \eqn{p}-value.}

\item{tau_input}{The value of tau chosen by the user.}

\item{lpsolver}{The name of the linear programming solver package to 
be used to obtain the solutions to the linear programs. The linear 
programming solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}' and `\code{lpsolveAPI}'.}

\item{qpsolver}{The name of the quadratic programming solver package to 
be used to obtain the solutions to the quadratic programs. The quadratic 
programming solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}' and `\code{osqp}'.}
}
\description{
This function checks and updates the input of the user. If 
   there is any invalid input, this function will be terminated and generates 
   appropriate error messages.
}
