% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{dkqs_cone_logicalb}
\alias{dkqs_cone_logicalb}
\title{Compute the logical upper and lower bounds for dkqs_cone}
\usage{
dkqs_cone_logicalb(f, farg)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing 
procedure.}
}
\value{
Returns the logical upper and lower bounds for dkqs_cone.
   \item{lb0}{Logical lower bound for \code{dkqs_cone}.}
   \item{ub0}{Logical upper bound for \code{dkqs_cone}.}
}
\description{
This function computes the logical upper and lower bounds for
   the test \code{dkqs_cone}.
}
