% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{dkqs_logicalb}
\alias{dkqs_logicalb}
\title{Compute the logical upper and lower bounds for dkqs}
\usage{
dkqs_logicalb(f, farg)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing 
procedure.}
}
\value{
Returns the logical upper and lower bounds for dkqs.
   \item{lb0}{Logical lower bound for \code{dkqs}.}
   \item{ub0}{Logical upper bound for \code{dkqs}.}
}
\description{
This function computes the logical upper and lower bounds for
   the test \code{dkqs}.
}
