% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{error.id.match}
\alias{error.id.match}
\title{Matches the id of the error messages}
\usage{
error.id.match(error.list, df.error)
}
\arguments{
\item{error.list}{List of error messages.}

\item{df.error}{Consolidated data frame of error messages.}
}
\value{
Returns an updated list of error messages.
  \item{df.error}{Updated data frame of error messages.}
}
\description{
This function is used after the bootstrap replications. This
  function is used to match the id of the error messages, given the list
  of error messages returned from the \code{future_lapply} function.
}
