% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds.check}
\alias{estbounds.check}
\title{Checks and updates the input in \code{estbounds}}
\usage{
estbounds.check(data, lpmodel, kappa, norm, solver, estimate, progress)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test that consists of
the following objects:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{kappa}{Tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{estimate}{Boolean variable to indicate whether the bounds should be
estimated or not.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the updated parameters and objects back to the function
\code{estbounds}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
      \item{\code{kappa}}
   }
}
\description{
This function checks and updates the input from the user for
   the function \code{estbounds}. If there is any invalid input, this
   function will terminate the procedure and generate appropriate error
   messages.
}
