% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds2.L2}
\alias{estbounds2.L2}
\title{Estimates the bounds with shape contraints (Stage 2 with \eqn{\ell^2}-norm)}
\usage{
estbounds2.L2(data, firststepsoln, lpmodel, modelsense, kappa, solver)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{firststepsoln}{List of solutions to the first step problem.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{A.shp}}
  \item{\code{beta.obs}}
  \item{\code{beta.shp}}
}}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{kappa}{Parameter used in the second step of the two-step procedure
for obtaining the solution subject to the shape constraints. It can be
any nonnegative number.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the second step of the two-step procedure.
 \item{objval}{Optimal value calculated from the optimizer.}
 \item{x}{Optimal point calculated from the optimizer.}
}
\description{
This function evaluates the solution to stage 2 of the
   two-step procedure obtaining the estimated bound. \eqn{\ell^2}-norm
   is used in the constraint
}
