% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds2_L2}
\alias{estbounds2_L2}
\title{Solves the second step in the two-step procedure with L2-norm}
\usage{
estbounds2_L2(firststepsoln, A_obs, beta_obs, modelsense, kappa, solver)
}
\arguments{
\item{firststepsoln}{List of solutions to the first step problem.}

\item{beta_obs}{Observed value of parameter of interest.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{kappa}{Parameter used in the second step of the two-step procedure 
for obtaining the solution subjected to the shape constraints.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the second step of the two-step procedure.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function evaluates the solution to stage 1 of the
   linear program in obtaining the estimated bound. \eqn{\ell^2}-norm is  
   used in the objective function.
}
