% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds_check}
\alias{estbounds_check}
\title{Checks and updates the input of the function \code{estbounds}}
\usage{
estbounds_check(
  df,
  func_obs,
  A_obs,
  A_tgt,
  beta_tgt,
  A_shp_eq,
  A_shp_ineq,
  beta_shp_eq,
  beta_shp_ineq,
  kappa,
  lnorm,
  solver,
  estimate,
  progress
)
}
\arguments{
\item{df}{The data being used in the inference.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{A_shp_eq}{Matrix representing equality shape constraints.}

\item{A_shp_ineq}{Matrix representing inequality shape constraints.}

\item{beta_shp_eq}{RHS vector in equality shape constraints.}

\item{beta_shp_ineq}{RHS vector in inequality shape constraints.}

\item{kappa}{Parameter used in the second step of the two-step procedure 
for obtaining the solution subjected to the shape constraints.}

\item{lnorm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{estimate}{Boolean variable to indicate whether the estimated 
problem should be considered.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the inference procedure. If it is set as \code{TRUE}, the 
messages are displayed throughout the procedure. Otherwise, the messages
will not be displayed.}
}
\value{
Returns the updated value of the parameters back to the function 
   \code{estbounds}. The following information are updated:
   \itemize{
      \item{\code{df}}
      \item{\code{A_obs}}
      \item{\code{A_tgt}}
      \item{\code{beta_obs}}
      \item{\code{beta_tgt}}
      \item{\code{A_shp_eq}}
      \item{\code{beta_shp_eq}}
      \item{\code{A_shp_ineq}}
      \item{\code{beta_shp_ineq}}
      \item{\code{solver}}
   }
}
\description{
This function checks and updates the input from the user for 
   the function \code{estbounds}. If there is any invalid input, this  
   function will terminate the procedure and generate appropriate error 
   messages.
}
