% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds_original}
\alias{estbounds_original}
\title{Computes the true bounds with shape contraints}
\usage{
estbounds_original(
  A_obs,
  A_tgt,
  beta_tgt,
  beta_obs,
  A_shp_eq,
  A_shp_ineq,
  beta_shp_eq,
  beta_shp_ineq,
  original_sense,
  solver
)
}
\arguments{
\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{beta_obs}{Observed value of parameter of interest.}

\item{A_shp_eq}{Matrix representing equality shape constraints.}

\item{A_shp_ineq}{Matrix representing inequality shape constraints.}

\item{beta_shp_eq}{RHS vector in equality shape constraints.}

\item{beta_shp_ineq}{RHS vector in inequality shape constraints.}

\item{original_sense}{Sense of the contraints to compute the true bound.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the linear program.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
The function computes the true bound subject to the shape 
   constraints without approximation.
}
