% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.bs.fn}
\alias{fsst.beta.bs.fn}
\title{Computes one bootstrap estimates for \code{beta.obs}.}
\usage{
fsst.beta.bs.fn(x, data, lpmodel)
}
\arguments{
\item{x}{This is either the list of indices that represent the bootstrap
replications, or the list of bootstrap components of the \code{lpmodel}
object passed from the user.}

\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{beta}{Bootstrap estimator.}
  \item{msg}{Error message (if applicable).}
}
\description{
This function carries out the one bootstrap replication for
  getting \code{beta.obs} in the FSST test. This function is used in the
  \code{fsst.beta.bs} function via the \code{future_lapply} command.
}
