% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.bs}
\alias{fsst.beta.bs}
\title{Computing the bootstrap betas}
\usage{
fsst.beta.bs(n, data, beta.obs.hat, lpmodel, R, maxR, progress, df.error, iseq)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}} based on the given
information in \code{lpmodel}.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{R}{Number of bootstraps chosen by the users.}

\item{maxR}{Maximum number of bootstrap replications in case error occured.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}

\item{df.error}{Table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{iseq}{The sequences of betas to iterate over}
}
\value{
Returns the bootstrap estimators.
  \item{beta.obs.bs}{List of bootstrap betas.}
  \item{R.succ}{Number of successful bootstrap replications.}
  \item{df.error}{Updated table showing the id of the bootstrap
      replication(s) with error(s) and the corresponding error message(s).}
}
\description{
This function computes the bootstrap betas
}
