% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.star.bs.fn}
\alias{fsst.beta.star.bs.fn}
\title{Computes one bootstrap estimates of \code{beta.star} and \code{x.star}}
\usage{
fsst.beta.star.bs.fn(
  beta.obs.bs,
  data,
  lpmodel,
  beta.tgt,
  weight.mat,
  sigma.beta.obs,
  solver,
  pbar,
  progress,
  n.bs,
  eval.count
)
}
\arguments{
\item{beta.obs.bs}{The bootstrap estimates of \code{beta.obs}.}

\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{weight.mat}{The weighting matrix for the \code{\link[lpinfer]{fsst}}
procedure.}

\item{sigma.beta.obs}{An estimator of the asymptotic variance for
\code{beta.obs}.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{pbar}{The progress bar object.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{n.bs}{The total number of replications to be conducted in this
procedure.}

\item{eval.count}{The count for the number of times the \code{future_lapply}
function has been called. If this object is zero, it means that the
\code{future_lapply} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_lapply} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{beta.star}{The bootstrap estimates of \code{beta.star}.}
  \item{x.star}{The bootstrap estimates of \code{x.star}.}
  \item{msg}{An error message (if applicable).}
}
\description{
This function computes one bootstrap estimate for
  \code{beta.star} and \code{x.star}. This function is used
  in the \code{\link[lpinfer]{fsst.beta.star.bs}} function via the
  \code{future_lapply} command.
}
