% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.star.bs}
\alias{fsst.beta.star.bs}
\title{Computes the bootstrap estimates of the starred version of \code{beta.obs}}
\usage{
fsst.beta.star.bs(
  data,
  lpmodel,
  beta.n,
  beta.n.bs,
  beta.tgt,
  weight.mat,
  beta.obs.hat,
  beta.obs.bs,
  R,
  sigma.beta.obs,
  solver,
  df.error,
  p,
  d
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.n}{Sample \eqn{\widehat{\bm{\beta}}_n} vector that is defined as
\eqn{\widehat{\bm{\beta}}_n \equiv (\widehat{\bm{\beta}}_{{\rm obs},n},
\bm{\beta}_{{\rm shp},n}, \bm{\beta}_{{\rm tgt}})'}.}

\item{beta.n.bs}{The bootstrap estimates of \code{beta.n}.}

\item{beta.tgt}{Value of beta to be tested.}

\item{weight.mat}{Weighting matrix.}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel}.}

\item{beta.obs.bs}{Bootstrap estimates of \code{beta.obs}.}

\item{R}{Number of bootstrap replications.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{df.error}{Table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{p}{Length of the beta vector.}

\item{d}{Number of columns of the \code{A} matrix.}
}
\value{
Return the following list of objects:
  \item{beta.star}{This corresponds to \eqn{\widehat{\bm{\beta}}^\star}.}
  \item{beta.star.bs}{This corresponds to the bootstrap estimates of
    \eqn{\widehat{\bm{\beta}}^\star}.}
  \item{x.star}{This corresponds to \eqn{\bm{x}^\star}, which is the
    solution to the quadratic program \code{beta.star.qp}.}
  \item{x.star.bs}{This corresponds to the bootstrap estimates of
    \eqn{\bm{x}^\star}.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{Number of new errors.}
  \item{error.id}{Indices of the bootstrap replications that give errors.}
}
\description{
This function computes the bootstrap estimates of
  \eqn{\widehat{\bm{\beta}}^\star_n}.
}
