% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.star}
\alias{fsst.beta.star}
\title{Compute bootstrap components of \eqn{\widehat{\bm{\beta}}^\star}}
\usage{
fsst.beta.star(
  data,
  lpmodel,
  beta.n,
  beta.n.bs,
  beta.tgt,
  weight.mat,
  beta.obs.hat,
  beta.obs.bs,
  R,
  sigma.beta.obs,
  solver,
  df.error,
  p,
  d
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{weight.mat}{Weighting matrix}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}} based on the given
information in \code{lpmodel}.}

\item{R}{Number of bootstraps chosen by the users.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{p}{Length of the beta vector.}

\item{d}{Number of columns of the A matrix.}
}
\value{
Return the following list of objects:
  \item{beta.star}{This corresponds to \eqn{\widehat{\bm{\beta}}^\star}.}
  \item{beta.star.bs}{This corresponds to the bootstrap counterparts of 
    \eqn{\widehat{\bm{\beta}}^\star}.}
  \item{x.star}{This corresponds to \eqn{\bm{x}^\star}, which is the
    solution to the quadratic program '\code{beta.star.qp}'.}
  \item{x.star.bs}{This corresponds to the bootstrap counterparts of 
    \eqn{\bm{x}^\star}.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{Number of new errors.}
  \item{error.id}{Indices of the bootstrap replications that give errors.}
}
\description{
This function computes the bootstrap components of
  \eqn{\widehat{\bm{\beta}}^\star}.
}
