% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.check}
\alias{fsst.check}
\title{Checks and updates the input in \code{fsst}}
\usage{
fsst.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  Rmulti,
  lambda,
  rho,
  n,
  weight.matrix,
  solver,
  progress,
  sqrtm.method,
  sqrtm.kmax,
  sqrtm.tol
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{Rmulti}{The multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{lambda}{Parameter used to obtain the restricted estimator
\eqn{\widehat{\bm{\beta}}^r_n}. A data-driven parameter \code{lambda} can
be included if \code{NA} is included as part of the vector for
\code{lambda}. For instance, if \code{lambda} is set as \code{c(0.1, NA)},
then both 0.1 and the data-driven \code{lambda} will be applied in the
\code{\link[lpinfer]{fsst}} procedure. The default is to use the
data-driven \code{lambda}.}

\item{rho}{Parameter used in the studentization of matrices.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{weight.matrix}{The option used in the weighting matrix. There are
three options available:
\itemize{
  \item{\code{identity} --- identity matrix}
  \item{\code{diag} --- the diagonal matrix that takes the diagonal
     elements of the inverse of the variance matrix}
  \item{\code{avar} --- inverse of the variance matrix}
}}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{sqrtm.method}{The method used to obtain the matrix square root in 
the \code{\link[lpinfer]{fsst}} procedure. There are two options available:
\itemize{
  \item{\code{"pracma"} --- the \code{\link[pracma]{sqrtm}}
  function from the \code{pracma} package will be used. 
  Note that if this method does not converge, the 
  \code{expm::sqrtm} function will be used automatically.}
  \item{\code{"expm"} --- the \code{\link[expm]{sqrtm}}
  function from the \code{expm} package will be used.}
}}

\item{sqrtm.kmax}{The maximum number of iterations used to compute the 
matrix square root. This is passed to the \code{kmax} argument of 
the \code{\link[pracma]{sqrtm}} function if 
\code{sqrtm.method = "pracma"}. Otherwise, this argument is ignored.}

\item{sqrtm.tol}{The absolute tolerance used to compute the matrix square
root. This is passed to the \code{tol} argument of 
the \code{\link[pracma]{sqrtm}} function if 
\code{sqrtm.method = "pracma"}. Otherwise, this argument is ignored.}
}
\value{
Returns the updated parameters back to the function \code{fsst}.
  The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{test.logical}}
      \item{\code{logical.lb}}
      \item{\code{logical.ub}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{fsst}} function. If there is any invalid input,
   the function will be terminated and error messages will be printed.
}
