% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.cone.bs}
\alias{fsst.cone.bs}
\title{Computing bootstrap components of `Cone`}
\usage{
fsst.cone.bs(
  n,
  omega.i,
  beta.n,
  beta.star,
  lpmodel,
  lambda,
  indicator,
  beta.star.bs,
  beta.r,
  beta.star.list,
  solver,
  cores,
  progress,
  length.lambda,
  lambda.i
)
}
\arguments{
\item{omega.i}{The matrix \eqn{\widehat{\bm{\Omega}}^i_n}, i.e. the
regularized matrix for
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{lambda}{Parameter used in the cone bootstrap sattistics}

\item{indicator}{A binary variable that equals to 1 for \eqn{d \geq p} and
equals to 0 for \eqn{d < p}.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop.
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}

\item{length.lambda}{The length of the `\code{lambda}` vector.}

\item{lambda.i}{The current value of `\code{lambda}` being considered}
}
\value{
A list of bootstrap Range statistics.
  \item{range.n.list}{A list of bootstrap Range statistics
    \eqn{\{\mathrm{Cone}_{n,b}\}^B_{b=1}}}
}
\description{
This function computes the bootstrap components of the
  `Cone` test statistics.
}
