% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtable.R
\name{fsst.cv.table}
\alias{fsst.cv.table}
\title{Wrapper for the \code{\link[lpinfer]{cv.table}} function for the
\code{\link[lpinfer]{fsst}} procedure}
\usage{
fsst.cv.table(
  param,
  param.name,
  cone.n.list,
  range.n.list,
  cone.bs.list,
  range.bs.list,
  T.bs.list,
  levels = c(0.99, 0.95, 0.9),
  digits = 5
)
}
\arguments{
\item{param}{The tuning parameter of the testing procedure.}

\item{param.name}{The name of the tuning parameter.}

\item{cone.n.list}{A list of sample cone test statistics.}

\item{range.n.list}{A list of sample range test statistics.}

\item{cone.bs.list}{A vector or a list of bootstrap cone test statistics.}

\item{range.bs.list}{A vector or a list of bootstrap range test statistics.}

\item{T.bs.list}{A vector or a list of bootstrap test statistics.}

\item{levels}{The confidence levels.}

\item{digits}{The number of digits in the test statistics.}
}
\value{
Returns a table of critical values.
  \item{cv.table}{A table of critical values.}
}
\description{
This is a wrapper of the \code{\link[lpinfer]{cv.table}}
  function in order to produce a table that consists of the critical values
  for the test statistics, cone component and the range component for the
  \code{\link[lpinfer]{fsst}} procedure.
}
