% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.label.lambda}
\alias{fsst.label.lambda}
\title{Indicates the data-driven \code{lambda} in the output}
\usage{
fsst.label.lambda(lambdas, lambda.data)
}
\arguments{
\item{lambdas}{The vector of \code{lambda}.}

\item{lambda.data}{The data-driven lambda.}
}
\value{
Returns the updated vector of \code{lambda} where the data-driven
  \code{lambda} is labelled with a star and the corresponding message
  \item{lambdas}{The updated \code{lambda}.}
  \item{msg}{An indicative message.}
}
\description{
This function labels the data-driven \code{lambda} in the
  \code{print} or \code{summary} output and returns an indicative message.
  This only affects the output messages, not the objects returned.
}
