% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.lambda}
\alias{fsst.lambda}
\title{Data-driven choice of \code{lambda} in the FSST procedure}
\usage{
fsst.lambda(
  n,
  omega.i,
  beta.n,
  beta.star,
  lpmodel,
  R.succ,
  beta.star.list,
  solver,
  progress,
  df.error,
  p,
  d
)
}
\arguments{
\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{omega.i}{The matrix \eqn{\widehat{\bm{\Omega}}^i_n}, i.e. the
regularized matrix for
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{beta.n}{Sample \eqn{\widehat{\bm{\beta}}_n} vector that is defined as
\eqn{\widehat{\bm{\beta}}_n \equiv (\widehat{\bm{\beta}}_{{\rm obs},n},
\bm{\beta}_{{\rm shp},n}, \bm{\beta}_{{\rm tgt}})'}.}

\item{beta.star}{This corresponds to the starred version of the
\code{beta.n} vector, i.e. \eqn{\widehat{\bm{\beta}}^\star_n}.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.star.list}{This corresponds to the bootstrap estimates of
\code{beta.star}.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{df.error}{Table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{p}{Length of the beta vector.}

\item{d}{Number of columns of the \code{A} matrix.}
}
\value{
Returns the data-driven \code{lambda} and the error messages.
  \item{lambda}{Data-driven \code{lambda}}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{Number of new errors.}
  \item{error.id}{Indices of the bootstrap replications that give errors.}
  \item{R.succ}{Number of successful bootstrap replications.}
}
\description{
This function provides a data-driven choice of \code{lambda}
  in the FSST procedure.
}
\details{
\eqn{\alpha_n} is set as \eqn{1} if the number of observations in
  the data set is less than 16.
}
