% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs.fn}
\alias{fsst.range.bs.fn}
\title{Computes one bootstrap estimate for the range component of the test
statistic}
\usage{
fsst.range.bs.fn(
  beta.x.star,
  n,
  lpmodel,
  beta.obs.hat,
  x.star,
  weight.mat,
  pbar,
  progress,
  n.bs,
  eval.count
)
}
\arguments{
\item{beta.x.star}{A list that combines the bootstrap estimates of
\code{beta.star} and \code{x.star}.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel}.}

\item{x.star}{Optimal value from \code{beta.star.qp}.}

\item{weight.mat}{Weighting matrix.}

\item{pbar}{Progress bar object.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{n.bs}{Total number of replications to be conducted in this procedure.}

\item{eval.count}{Count for the number of times the \code{future_lapply}
function has been called. If this object is zero, it means that the
\code{future_lapply} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_lapply} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Return the following list of objects:
  \item{Ts}{Range test statistic.}
  \item{msg}{Error message (if applicable).}
}
\description{
This function computes the one bootstrap estimate for the
  range component of the test statistic.
}
