% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs}
\alias{fsst.range.bs}
\title{Computing bootstrap components of `Range`}
\usage{
fsst.range.bs(
  n,
  lpmodel,
  beta.obs.hat,
  beta.obs.bs,
  x.star,
  x.star.bs,
  weight.mat,
  R,
  cores,
  progress
)
}
\arguments{
\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.obs.hat}{Estimator of \eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}}
based on the given information in \code{lpmodel}.}

\item{beta.obs.bs}{List of bootstrap components of \code{beta.obs}.}

\item{x.star}{Optimal value from `\code{beta.star.qp}`.}

\item{x.star.bs}{Bootstrap components of `\code{x.star}`.}

\item{weight.mat}{Weighting matrix}

\item{R}{Number of bootstraps chosen by the users.}

\item{cores}{Number of cores to be used in the parallelized for-loop.
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
A list of bootstrap Range statistics.
  \item{range.n.list}{A list of bootstrap Range statistics
    \eqn{\{\mathrm{Range}_{n,b}\}^B_{b=1}}}
}
\description{
This function computes the bootstrap components of the
  `Range` test statistics.
}
