% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.lp}
\alias{fsst.range.lp}
\title{Computes the solution to the range problem}
\usage{
fsst.range.lp(n, omega.e, beta.n, beta.star, solver)
}
\arguments{
\item{omega.e}{The matrix \eqn{\widehat{\bm{\Omega}}^e_n}, i.e. the 
regularized matrix for 
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the range problem.
}
