% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range}
\alias{fsst.range}
\title{Computes the range component of the test statistics}
\usage{
fsst.range(n, beta.obs.hat, x.star, lpmodel, weight.mat.root)
}
\arguments{
\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{beta.obs.hat}{The sample estimator
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel} (and \code{data} if applicable).}

\item{x.star}{The optimal value from \code{beta.star.qp}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{weight.mat.root}{The matrix square root of \code{weight.mat}.}
}
\value{
Returns the range test statistic.
 \item{range}{The sample range test statistic.}
}
\description{
This function computes the range component of the test
  statistics in the \code{\link[lpinfer]{fsst}} procedure.
}
