% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{full.beta.bs}
\alias{full.beta.bs}
\title{Construct the full beta vector in the \code{\link[lpinfer]{fsst}} procedure}
\usage{
full.beta.bs(lpmodel, beta.tgt, beta.obs.bs, R)
}
\arguments{
\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{beta.obs.bs}{The bootstrap estimates of \code{beta.obs}.}

\item{R}{The number of bootstrap replications.}
}
\value{
List of full beta vectors
  \item{beta.bs}{The list of full beta vector
  \eqn{\{\widehat{\bm{\beta}}_{n,b}\}^B_{b=1}}.}
}
\description{
This function concatenate the \code{beta.obs}, \code{beta.shp}
  and \code{beta.tgt} components to form the full \code{beta} vector. In
  particular, the full vector is defined as
  \eqn{\widehat{\bm{\beta}}_n} vector that is defined as
  \eqn{\widehat{\bm{\beta}}_n \equiv (\widehat{\bm{\beta}}_{{\rm obs},n},
  \bm{\beta}_{{\rm shp},n}, \bm{\beta}_{{\rm tgt}})'}.
}
