% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{gurobi.optim}
\alias{gurobi.optim}
\title{LP and QP solver by Gurobi}
\usage{
gurobi.optim(
  Af,
  bf,
  nf,
  A,
  rhs,
  sense,
  modelsense,
  lb,
  qc = NULL,
  weight = diag(ncol(bf))
)
}
\arguments{
\item{Af}{The matrix that is involved in the objective function.}

\item{bf}{The vector that is involved in the objective function.}

\item{nf}{The number of observations in the dataframe.}

\item{A}{The constraint matrix.}

\item{rhs}{The RHS vector for the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{lb}{The lower lound vector.}

\item{qc}{List of quadratic constraint(s). There can be multiple quadratic
constraints. Each constraint has to be a list.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the quadratic or linear
   program using the `\code{Gurobi}' package. This function can have linear
   and/or quadratic constraints.
}
