% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{gurobi_optim}
\alias{gurobi_optim}
\title{Gurobi solver for quadratic or linear programs}
\usage{
gurobi_optim(obj2, obj1, obj0, A, rhs, sense, modelsense, lb)
}
\arguments{
\item{obj2}{The matrix that corresponds to the second-order term in the
quadratic program.}

\item{obj1}{The vector that corresponds to the first-order term in the
quadratic or linear program.}

\item{obj0}{The constant term in the quadratic or linear program.}

\item{A}{The constraint matrix.}

\item{rhs}{The RHS vector fo the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{lb}{The lower lound vector.}
}
\value{
Returns the solution to the quadratic or linear program.
}
\description{
This function computes the solution to the quadratic or linear
   program using Gurobi.
}
