% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{infeasible.betatgt.warning}
\alias{infeasible.betatgt.warning}
\title{Display warning message for infeasible 'beta.tgt'}
\usage{
infeasible.betatgt.warning()
}
\value{
Nothing is returned.
}
\description{
This function displays the warning message for infeasible
  'beta.tgt'.
}
