% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{infeasible.summary.betatgt}
\alias{infeasible.summary.betatgt}
\title{Wrapper for '\code{infeasible.msg.betatgt}'}
\usage{
infeasible.summary.betatgt()
}
\value{
Nothing is returned.
}
\description{
This function is a wrapper for '\code{infeasible.msg.betatgt}'
  to print the reason for the p-value to be zero and the p-value. It is
  used in the summary functions.
}
