% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{invertci}
\alias{invertci}
\title{Constructs confidence interval}
\usage{
invertci(
  f,
  farg = list(),
  alpha = 0.05,
  init.lb = NULL,
  init.ub = NULL,
  tol = 1e-04,
  max.iter = 20,
  pvals = NULL,
  dp = 5,
  progress = TRUE
)
}
\arguments{
\item{f}{The function that represents a testing procedure.}

\item{farg}{The list of arguments to be passed to the function of testing
procedure.}

\item{alpha}{The significance level(s). This can be a vector.}

\item{init.lb}{The initial brackets to search for the lower bound. This is
not required if the \code{\link[lpinfer]{chorussell}} is used.}

\item{init.ub}{The initial brackets to search for the upper bound. This is
not required if the \code{\link[lpinfer]{chorussell}} is used.}

\item{tol}{The tolerance level in the bisection method.}

\item{max.iter}{The maximum number of iterations in the bisection method.}

\item{pvals}{The data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{dp}{The number of decimal places in the output.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the procedure of constructing confidence interval. If
it is set as \code{TRUE}, the messages are displayed throughout the
procedure. Otherwise, the messages will not be displayed.}
}
\value{
Returns the confidence interval and a data frame that contains the
   points being tested in the procedure.
   \item{pvals}{The data frame that consists of the points and the
      corresponding \eqn{p}-values that have been tested in constructing
      the confidence intervals.}
   \item{df_ub}{The data frame storing the information for the bisection
      method in each iteration when evaluating the upper bound of the
      confidence interval.}
   \item{df_lb}{The data frame storing the information for the bisection
      method in each iteration when evaluating the lower bound of the
      confidence interval.}
   \item{alpha}{The significance levels.}
   \item{tol}{The tolerance level in the bisection method.}
   \item{iter}{The total number of iterations taken.}
   \item{call}{The matched call.}
   \item{para.name}{The name of the tuning parameters involved.}
   \item{para.vals}{The values of the tuning parameters involved.}
   \item{ci}{The confidence intervals constructed.}
}
\description{
This function constructs the confidence interval using the
   bisection method.
}
\details{
The number of decimal places displayed in the messages (if
   \code{progress} is set as \code{TRUE}) is equal to the number of decimal
   places in the variable \code{tol}.
}
\section{Example}{

\preformatted{
  source("./example/dgp_missingdata.R") # Change directory if necessary
  J <- 5
  N <- 1000
  data <- missingdata_draw(J = J, n = N, seed = 1, prob.obs = .5)
  lpm <- missingdata_lpm(J = J, info = "full", data = data)
  farg <- list(data = data,
               lpmodel = lpm,
               R = 100,
               phi = 2/3,
               solver = "gurobi",
               progress = FALSE)
  invertci(f = subsample,
           farg = farg,
           alpha = .05)
}
}

\section{More examples}{

  More examples can be found in the \code{invertci_example.R} file
  under the \code{example} subdirectory of the installation directory for
  the \code{lpinfer} package.
}

