% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{invertci.show.param}
\alias{invertci.show.param}
\title{Print the parameters used in the bisection method}
\usage{
invertci.show.param(para.name, para.vals, j)
}
\arguments{
\item{para.name}{The name(s) of the parameter(s).}

\item{para.vals}{The value(s) of the parameter(s).}

\item{j}{The row number for the parameters being used.}
}
\value{
Nothing is returned.
}
\description{
This function is used to print the tuning parameters that are
  relevant to the current set of iterations in the
  \code{\link[lpinfer]{invertci}} function.
}
