% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{limsolve.optim}
\alias{limsolve.optim}
\title{LP and QP solver by limSolve}
\usage{
limsolve.optim(Af, bf, nf, A, rhs, sense, modelsense, lb)
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to linear and quadratic 
   programs using the `\code{limSolve}' package.
}
