% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{limsolve_optim}
\alias{limsolve_optim}
\title{LP and QP solver by limSolve}
\usage{
limsolve_optim(Af, bf, nf, A, rhs, sense, modelsense, lb)
}
\arguments{
\item{Af}{The matrix that is involved in the objective function.}

\item{bf}{The vector that is involved in the objective function.}

\item{nf}{The number of observations in the dataframe.}

\item{A}{The constraint matrix.}

\item{rhs}{The RHS vector for the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{lb}{The lower lound vector.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to linear and quadratic 
   programs using the `\code{limSolve}' package.
}
