% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lpinfer.seed}
\alias{lpinfer.seed}
\title{Gets the current seed with RNG kind "L'Ecuyer-CMRG"}
\usage{
lpinfer.seed()
}
\value{
Returns a \code{seed}.
  \item{seed}{Seed with RNG kind "L'Ecuyer-CMRG".}
}
\description{
This function gets the current seed with RNG kind
  "L'Ecuyer-CMRG".
}
