% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{lpm.print}
\alias{lpm.print}
\title{Print the \code{lpmodel} or \code{lpmodel.natural} object}
\usage{
lpm.print(x, lpm.string, data = NULL, ...)
}
\arguments{
\item{x}{The \code{lpmodel} object or \code{lpmodel.natural} object.}

\item{lpm.string}{The string that contains the name of the variables
available in the \code{lpmodel} object.}

\item{data}{An \code{data frame} or a \code{matrix}.}
}
\value{
Nothing is returned.
}
\description{
This function prints the details of the components that are
   contained in the \code{lpmodel} or \code{lpmodel.natural} object.
}
