% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{lpmodel.beta.eval}
\alias{lpmodel.beta.eval}
\title{Evaluates the point estimate and asymptotic variance of \code{beta.obs}}
\usage{
lpmodel.beta.eval(data, obj, i)
}
\arguments{
\item{data}{Data frame}

\item{obj}{Object in lpmodel}

\item{i}{Index}
}
\value{
Returns the point estimate and the asymptotic variane of the
   \code{beta.obs} object.
   \item{beta.obs}{Point estimate of \eqn{\widehat{\bm{\beta}}_{\rm obs}}}
   \item{omega}{Asymptotic variance of \eqn{\widehat{\bm{\beta}}_{\rm obs}}}
}
\description{
This function returns the matrix or vector depending on the
  class of the variable in the \code{lpmodel} object. In the design of
  the \code{lpinfer} module, objects in \code{lpmodel} can have three
  different types of classes:
  \itemize{
     \item{\code{function} --- If the object is a function, then the
        object that is evaluated by the function with the data will be
        returned.}
     \item{\code{list} --- If the object is a list, then the \eqn{i}th
        element will be returned.}
  }
}
