% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{lpmodel.eval}
\alias{lpmodel.eval}
\title{Evaluates an object inside \code{lpmodel}}
\usage{
lpmodel.eval(data, obj, i)
}
\arguments{
\item{data}{Data frame.}

\item{obj}{Object in lpmodel.}

\item{i}{Index.}
}
\value{
Returns an object at iteration \code{i}.
'
}
\description{
This function returns the matrix or vector depending on the
  class of the variable in the \code{lpmodel} object. In the design of
  the \code{lpinfer} module, objects in \code{lpmodel} can have three
  different types of classes:
  \itemize{
     \item{\code{function} --- If the object is a function, then the
        object that is evaluated by the function with the data will be
        returned.}
     \item{\code{list} --- If the object is a list, then the \eqn{i}th
        element will be returned.}
     \item{\code{matrix} or \code{numeric} --- If the object is a matrix
        or in a vector, then it will be directly returned.}
  }
}
