% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{lpmodel.extractlist}
\alias{lpmodel.extractlist}
\title{Extracts the bootstrap replications of the \code{lpmodel} object}
\usage{
lpmodel.extractlist(obj, len)
}
\arguments{
\item{obj}{A component in the \code{lpmodel} object.}

\item{len}{The length of the list.}
}
\value{
Returns one object.
  \item{result}{This is either a \code{list} or \code{NULL}.}
}
\description{
This function extracts the bootstrap replications in the
  \code{lpmodel} object if the object is a \code{list}. Otherwise, this
  function returns \code{NULL}.
}
