% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{lpmodel}
\alias{lpmodel}
\title{Define a \code{lpmodel} object}
\usage{
lpmodel(
  A.obs = NULL,
  A.shp = NULL,
  A.tgt = NULL,
  beta.obs = NULL,
  beta.shp = NULL
)
}
\arguments{
\item{A.obs}{A matrix, list or function}

\item{A.shp}{A matrix, list or function}

\item{A.tgt}{A matrix, list or function}

\item{beta.obs}{A vector, list or function}

\item{beta.shp}{A vector, list or function}
}
\value{
Returns a list of \code{lpmodel} objects in the \code{lpmodel}
   class.
}
\description{
This function defines the objects required in the
   \code{lpinfer} module in the \code{lpmodel} class.
}
