% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{lpsolveapi.optim}
\alias{lpsolveapi.optim}
\title{lpSolveAPI solver for linear programs}
\usage{
lpsolveapi.optim(Af, bf, nf, A, rhs, sense, modelsense, lb)
}
\value{
Returns the optimal objective value and the corresponding argument
  to the linear program.
 \item{x}{Optimal point calculated from the linear program}
 \item{larg}{Arguments for the linear program.}
}
\description{
This function computes the solution to the linear program
   using the `\code{lpsolveAPI}' package.
}
\details{
The package `\code{lpSolveAPI}' cannot be used to solve quadratic 
  programs.
}
