% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{many_qpci}
\alias{many_qpci}
\title{Wrapper function for \code{qpci}}
\usage{
many_qpci(
  f,
  farg,
  alphas = c(0.05),
  lb0 = NULL,
  lb1 = NULL,
  ub0 = NULL,
  ub1 = NULL,
  tol = NULL,
  max_iter = 10,
  df_ci = NULL,
  progress_one = FALSE,
  progress_many = FALSE
)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing 
procedure.}

\item{alphas}{The list of significance levels to be used in constructing
the confidence intervals.}

\item{lb0}{Logical lower bound for the confidence interval.}

\item{lb1}{Maximum possible lower bound for the confidence interval.}

\item{ub0}{Logical upper bound for the confidence interval.}

\item{ub1}{Minimum possible upper bound for the confidence interval.}

\item{tol}{Tolerance level in the bisection method.}

\item{max_iter}{Maximum number of iterations in the bisection method.}

\item{df_ci}{Data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{progress_one}{The boolean variable for whether the result messages 
should be displayed in running the function \code{qpci}. If it is set as 
\code{TRUE}, the messages are displayed throughout the procedure. 
Otherwise, the messages will not be displayed.}

\item{progress_many}{The boolean variable for whether the result messages 
should be displayed in running the function \qpci{many_qpci}, i.e. the 
current function. If it is set as \code{TRUE}, the messages are displayed 
throughout the procedure. Otherwise, the essages will not be displayed.}
}
\value{
Returns a list of confidence intervals.
}
\description{
This function serves a wrapper function for computing 
   confidence interval with many significance levels.
}
