% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{mincriterion.check}
\alias{mincriterion.check}
\title{Checks and updates the input of the function \code{mincriterion}}
\usage{
mincriterion.check(data, lpmodel, norm, solver)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{A.shp}}
  \item{\code{beta.obs}}
  \item{\code{beta.shp}}
}}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the updated parameters back to the function
\code{mincriterion}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
   }
}
\description{
This function checks and updates the input from the user for
   the function \code{mincriterion}. If there is any invalid input, this
   function will terminate the procedure and generate appropriate error
   messages.
}
