% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{mincriterion}
\alias{mincriterion}
\title{First-stage of the estimation procedure for \code{estbounds}}
\usage{
mincriterion(
  data,
  func_obs,
  A_obs,
  A_tgt,
  A_shp_eq,
  A_shp_ineq,
  beta_tgt,
  beta_shp_eq,
  beta_shp_ineq,
  norm,
  solver
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{A_obs}{The "observed matrix" in the inference \eqn{A_{\mathrm{obs}}}.}

\item{A_tgt}{The "target matrix" in the inference \eqn{A_{\mathrm{tgt}}}.}

\item{A_shp_eq}{Matrix representing equality shape constraints.}

\item{A_shp_ineq}{Matrix representing inequality shape constraints.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{beta_shp_eq}{RHS vector in equality shape constraints.}

\item{beta_shp_ineq}{RHS vector in inequality shape constraints.}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the first step of the two-step procedure 
   and argument for the linear program.
 \item{objval}{Optimal value calculated from the optimizer.}
 \item{x}{Optimal point calculated from the optimizer.}
 \item{larg}{Arguments for the estimation program.}
 \item{norm}{Norm used in the estimation problem.}
 \item{solver}{The solver used in the estimation problem}
 \item{call}{The details of the function that has been called.}
}
\description{
This function evaluates the solution to stage 1 of the 
   two-step procedure obtaining the estimated bound. This function can
   be used to evaluate both the estimation problem with the 1-norm or 
   the 2-norm.
}
