% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{objective.function}
\alias{objective.function}
\title{Computes the coefficient terms of the objective functions}
\usage{
objective.function(A, b, n, weight = NULL)
}
\arguments{
\item{A}{The matrix \eqn{\bm{A}}.}

\item{b}{The column vector \eqn{\bm{\beta}}.}

\item{n}{The sample size \eqn{n}.}

\item{weight}{The weighting matrix.}
}
\value{
Returns the following three quantities: \code{obj2} is the
  coefficient of the quadratic term, \code{obj1} is the coefficient of the
  linear term and \code{obj0} is the constant term. More explicitly, their
  form are given as follows:
  \item{obj2}{The coefficient for the second-order term. It is
    returned as \code{NULL} for linear programs and
    \eqn{n\bm{A}'\bm{W}\bm{A}} for quadratic programs.}
  \item{obj1}{The coefficient term of the linear term. For quadratic
    programs, it is returned as \eqn{-2n\bm{\beta}'\bm{W}\bm{A}}.}
  \item{obj0}{The constant term of the linear program. For quadratic
    programs, it is returned as \eqn{n\bm{\beta}'\bm{W}\bm{\beta}}.}
}
\description{
This function computes the matrices in the objective functions
   for linear programs. This function takes matrix \eqn{\bm{A}} and
   \eqn{\bm{\beta}} as input and computes the coefficients of the objective
   function.
}
\details{
\itemize{
  \item{\strong{Quadratic programs} ---
     Given inputs \eqn{\bm{A} \in \mathbf{R}^{m\times m}},
     \eqn{\bm{b} \in \mathbf{R}^m} and
     \eqn{\bm{W} \in \mathbf{R}^{m\times m}}, the equation of the objective
     function of the quadratic program can be written as
     \deqn{n (\bm{A}\bm{x} -\bm{\beta})'\bm{W}(\bm{A}\bm{x}-\bm{\beta})
     = n\bm{x}'\bm{A}'\bm{W}\bm{A}\bm{x} - 2n\bm{\beta}'\bm{W}\bm{A}\bm{x} +
     n\bm{\beta}'\bm{W}\bm{\beta}.}
     If the \eqn{\bm{W}} matrix is not specified, then it will be taken as
     an identity matrix of order \eqn{m}.}
  \item{\strong{Linear programs} ---
     For all linear problems that are considered in this code, one of
     \eqn{\bm{A}} and \eqn{\bm{b}} is \code{NULL} or is a zero vector. The
     term that is nonzero and nonnull will be multiplied by \eqn{n} and
     used as \code{obj1}.}
}
}
