% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{osqp_optim}
\alias{osqp_optim}
\title{osqp solver for quadratic programs}
\usage{
osqp_optim(obj2, obj1, obj0, A, rhs, sense, modelsense, lb)
}
\arguments{
\item{obj2}{The matrix that corresponds to the second-order term in the
quadratic program. This input will be set to \code{NULL} if the problem
considered is a linear program.}

\item{obj1}{The vector that corresponds to the first-order term in the
quadratic or linear program.}

\item{obj0}{The constant term in the quadratic or linear program.}

\item{A}{The constraint matrix.}

\item{rhs}{The RHS vector for the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{lb}{The lower lound vector.}
}
\value{
Returns the optimal objective value and the corresponding argument
  to the linear program.
}
\description{
This function computes the solution to the quadratic program
   using the `\code{osqp}' package.
}
\details{
The package `\code{osqp}' cannot be used to solve linear programs. 
  Hence, the argument \code{obj2} cannot be \code{NULL}.
}
