% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{para.comb}
\alias{para.comb}
\title{Combine lists in parallel loops}
\usage{
para.comb(x, ...)
}
\arguments{
\item{x}{List of objects returned.}

\item{...}{Additional arguments.}
}
\description{
This function is used to combine the list of objects in the
  parallelized bootstrap replications.
}
\details{
This is used in the parallelized bootstrap replications.
}
