% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{post.bs}
\alias{post.bs}
\title{Auxiliary function for the post-bootstrap procedure}
\usage{
post.bs(
  test.return,
  i0,
  i1,
  R.eval,
  T.list = NULL,
  beta.list = NULL,
  error.list = NULL,
  list.param = NULL,
  error.param = NULL,
  ub.list = NULL,
  lb.list = NULL
)
}
\arguments{
\item{test.return}{The list of information returned from the test function.}

\item{i0}{The starting index to be evaluated.}

\item{i1}{The last index to be evaluated.}

\item{R.eval}{The number of bootstrap replications that has been evaluated.}

\item{T.list}{The list of bootstrap test statistics.}

\item{beta.list}{The list of the bootstrap replications of the
\code{beta.obs} component.}

\item{error.list}{The list of error messages.}

\item{error.param}{The list of parameters that lead to errors (if applicable).}

\item{ub.list}{The list of upper bounds (if applicable).}

\item{lb.list}{The list of lower bounds (if applicable).}

\item{param.list}{The list of parameters (if applicable).}
}
\value{
Returns the following three objects.
  \item{R.succ}{The number of successful bootstrap replications.}
  \item{R.eval}{The number of bootstrap replications that has been evaluated.}
  \item{T.list}{The list of test statistics.}
  \item{beta.list}{The list of the bootstrap replications of the
    \code{beta.obs} component.}
  \item{error.list}{The list of error messages (if applicable).}
  \item{error.param}{The list of parameters that lead to errors (if
    applicable).}
  \item{ub.list}{The list of upper messages (if applicable).}
  \item{lb.list}{The list of lower messages (if applicable).}
}
\description{
This function is used after the bootstrap replications to
  extract the relevant and updated information.
}
