% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{print.chorussell}
\alias{print.chorussell}
\title{Print results from \code{\link[lpinfer]{chorussell}}}
\usage{
print.chorussell(x, ...)
}
\arguments{
\item{x}{The output objects returned from \code{\link[lpinfer]{chorussell}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function either prints the \eqn{p}-values or a
  \eqn{(1-\alpha)} confidence interval from
  \code{\link[lpinfer]{chorussell}}.
}
