% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{print.dkqs}
\alias{print.dkqs}
\title{Print results from \code{dkqs}}
\usage{
\method{print}{dkqs}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{dkqs}.}

\item{...}{Additional arguments.}
}
\value{
Print the basic set of results from \code{dkqs}.
}
\description{
This function uses the print method on the return list of the
   function \code{dkqs}.
}
