% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{print.dkqs_cone}
\alias{print.dkqs_cone}
\title{Print results from \code{dkqs_cone}}
\usage{
\method{print}{dkqs_cone}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{dkqs_cone}.}

\item{...}{Additional arguments.}
}
\value{
Print the basic set of results from \code{dkqs_cone}.
}
\description{
This function uses the print method on the return list of the
   function \code{dkqs_cone}.
}
