% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{print.estbounds}
\alias{print.estbounds}
\title{Print results from \code{\link[lpinfer]{estbounds}}}
\usage{
print.estbounds(x, ...)
}
\arguments{
\item{x}{The output objects returned from \code{\link[lpinfer]{estbounds}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function uses the print method on the return list of the
   function \code{\link[lpinfer]{estbounds}}.
}
