% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{print.fsst}
\alias{print.fsst}
\title{Print results from \code{fsst}}
\usage{
print.fsst(x, ...)
}
\arguments{
\item{x}{Object returned from \code{fsst}.}

\item{...}{Additional arguments.}
}
\description{
This function uses the print method on the return list of the
   function \code{fsst}.
}
\details{
The following information are printed
  \itemize{
    \item{Test statistic (Max of Cone and Range)}
    \item{Test statistic (Cone)}
    \item{Test statistic (Range)}
    \item{\eqn{p}-value}
    \item{Solver used}
    \item{Number of cores used}
  }
}
