% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{print.invertci_multiple}
\alias{print.invertci_multiple}
\title{Print results from \code{\link[lpinfer]{invertci}} with multiple
significance levels}
\usage{
\method{print}{invertci_multiple}(x, alphas = NULL, ...)
}
\arguments{
\item{x}{The output object returned from \code{\link[lpinfer]{invertci}}.}

\item{alphas}{A list of alphas that the user would like to print.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned
}
\description{
This function uses the print method on the return list of the
   function \code{\link[lpinfer]{invertci}}.
}
