% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{print.invertci_single}
\alias{print.invertci_single}
\title{Print results from \code{invertci} with a single alpha}
\usage{
print.invertci_single(x, ...)
}
\arguments{
\item{x}{Object returned from \code{invertci}.}

\item{...}{Additional arguments.}
}
\value{
Print the basic set of results from \code{invertci}.
}
\description{
This function uses the print method on the return list of the
   function \code{invertci}.
}
