% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{print.lpmodel.natural}
\alias{print.lpmodel.natural}
\title{Print the \code{lpmodel.natural} object}
\usage{
\method{print}{lpmodel.natural}(x, data = NULL, ...)
}
\arguments{
\item{x}{An \code{lpmodel.natural} object.}

\item{data}{An \code{data frame} or a \code{matrix}.}

\item{...}{Some additional arguments passed to the \code{lpm.print}
function.}
}
\value{
Nothing is returned.
}
\description{
This function is a wrapper of the
  \code{\link[lpinfer]{lpm.print}} function and prints the details of the
  components in the \code{lpmodel.natural} object.
}
