% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{print.lpmodel}
\alias{print.lpmodel}
\title{Print the \code{lpmodel} object}
\usage{
\method{print}{lpmodel}(x, data = NULL, ...)
}
\arguments{
\item{x}{The \code{lpmodel} object.}

\item{data}{Data used in the tests.}
}
\value{
Nothing is returned.
}
\description{
This function is a wrapper of the \code{lpm.print} function
   and prints the details of the components in the \code{lpmodel} object.
}
