% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{print.subsample}
\alias{print.subsample}
\title{Print results from the \code{\link[lpinfer]{subsample}} procedure}
\usage{
print.subsample(x, ...)
}
\arguments{
\item{x}{An output object returned from \code{\link[lpinfer]{subsample}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints the \eqn{p}-values from
the \code{\link[lpinfer]{subsample}} procedure.
}
