% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{pval}
\alias{pval}
\title{Calculates the \eqn{p}-value}
\usage{
pval(T.bs, T.n)
}
\arguments{
\item{T.n}{The sample test statistics.}

\item{T_bs}{The list of bootstrap estimates of the test statistic.}
}
\value{
Returns the \eqn{p}-value:
  \item{p}{The \eqn{p}-value.}
}
\description{
This function computes the \eqn{p}-value of the test based on
   the bootstrap estimates.
}
