% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{qpci_check}
\alias{qpci_check}
\title{Checks and updates the input of the function \code{qpci}}
\usage{
qpci_check(f, farg, alpha, lb0, lb1, ub0, ub1, tol, max_iter, df_ci, noisy)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing 
procedure.}

\item{alpha}{Significance level of the test.}

\item{lb0}{Logical lower bound for the confidence interval.}

\item{lb1}{Maximum possible lower bound for the confidence interval.}

\item{ub0}{Logical upper bound for the confidence interval.}

\item{ub1}{Minimum possible upper bound for the confidence interval.}

\item{tol}{Tolerance level in the bisection method.}

\item{max_iter}{Maximum number of iterations in the bisection method.}

\item{df_ci}{Dataframe that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{noisy}{The boolean variable for whether the result messages should
be displayed in the procedure of constructing confidence interval. If 
it is set as \code{TRUE}, the messages are displayed throughout the 
procedure. Otherwise, the messages will not be displayed.}
}
\value{
Returns the updated value of the parameters back to the function 
   \code{qpci} in the correct format.
   \item{df_ci}{Data frame that stores the points that has been evaluated
      and the corresponding \eqn{p}-values.}
   \item{lb0}{Logical lower bound for the confidence interval.}
   \item{lb1}{Maximum possible lower bound for the confidence interval.}
   \item{ub0}{Logical upper bound for the confidence interval.}
   \item{ub1}{Minimum possible upper bound for the confidence interval.}
}
\description{
This function checks and updates the input from the user for 
   the function \code{qpci}. If there is any invalid input, this function 
   will terminate the procedure and generate appropriate error messages.
}
