% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{quan.stat}
\alias{quan.stat}
\title{Function that computes the basic quantiles}
\usage{
quan.stat(stat, quan = c(0.9, 0.95, 0.99))
}
\arguments{
\item{stat}{Test statistics}

\item{quan}{Quantiles}
}
\value{
Return the quantile of the test statistics in the order of the 
  `\code{quan}` vector.
  \item{stat.quan}{Quantile of the test statistics in the order of the
  `\code{quan}` vector.}
}
\description{
This function is used to evaluate the test statistics at 
  different standard quantiles. By default, it evaluates the test 
  statistics at the quantiles - 90%, 95% and 99%.
}
