% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{rcplex.optim}
\alias{rcplex.optim}
\title{LP and QP solver by Rcplex}
\usage{
rcplex.optim(Af, bf, nf, A, rhs, sense, modelsense, lb)
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the linear and quadratic
   programs using the `\code{Rcplex}' package.
}
