% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Simulated data}
\format{
A data frame with 1,000 rows and 2 columns.
\describe{
  \item{Y}{Multivariate discrete outcome variable that takes value of 0 to 1
  with increment 0.1.}
  \item{D}{Binary treatment where \eqn{D_i=1} means that \eqn{Y_i} is
  observed.}
}
}
\source{
Simulated
}
\usage{
sampledata
}
\description{
Simulated data
}
\keyword{datasets}
