% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampledata_generation}
\alias{sampledata_generation}
\title{Simulated data}
\format{A data frame with 1,000 rows and 2 columns.
\describe{
  \item{Y}{multivariate discrete outcome variable taht takes value of 0 to 1
  with step size 0.1}
  \item{D}{binary treatment where \eqn{D_i=1} means \eqn{Y_i} is observed}
}}
\source{
Simulated
}
\usage{
sampledata_generation()
}
\description{
Simulated data
}
