% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{sigma.est}
\alias{sigma.est}
\title{Computing bootstrap and variance estimators}
\usage{
\method{sigma}{est}(n, data, beta.obs.hat, beta.obs.bs, progress)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}} based on the given
information in \code{lpmodel}.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}

\item{df.error}{Table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}
}
\value{
Returns the bootstrap estimators and the estimator of the
  asymptotic variance.
    \item{sigma.mat}{Estimator of the asympototic variance.}
    \item{beta.obs.bs}{List of bootstrap betas.}
  \item{R.succ}{Number of successful bootstrap replications.}
    \item{df.error}{Updated table showing the id of the bootstrap
      replication(s) with error(s) and the corresponding error message(s).}
}
\description{
Based on the function \code{beta.obs} from the object
  \code{lpmodel}, this function computes the bootstrap estimates and
  the asympototic variance estimator of the boostrap estimator. The
  asymptotic variance estimator is computed as
    \deqn{\frac{n}{B} \sum^B_{i=1} \left(\widehat{\bm{\beta}}_b -
    \widehat{\bm{\beta}}\right)  \left(\widehat{\bm{\beta}}_b -
    \widehat{\bm{\beta}}\right)'}
  for some suitable estimators \eqn{\widehat{\bm{\beta}}} and the
  corresponding bootstrap estimators
  \eqn{\{\widehat{\bm{\beta}}_b\}^B_{b=1}}.
}
