% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{sigma.summation.parallel}
\alias{sigma.summation.parallel}
\title{Computing the variance estimator using parallel programming}
\usage{
\method{sigma}{summation.parallel}(n, beta.bs.list, cores, progress, ind.times)
}
\arguments{
\item{beta.bs.list}{List of bootstrap estimators
\eqn{\{\widehat{\bm{\beta}}_b\}^B_{b=1}}.}

\item{cores}{Number of cores to be used in the parallelized for-loop. 
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{progress}{The boolean variable for whether the result messages 
should be displayed in the inference procedure. If it is set as 
\code{TRUE}, the messages are displayed throughout the procedure. 
Otherwise, the messages will not be displayed.}

\item{ind.times}{Indicator variable showing whether this is called for the 
first or the second time}
}
\value{
Returns the bootstrap estimators and the estimator of the 
  asymptotic variance.
    \item{sigma.mat}{Estimator of the asympototic variance}
}
\description{
Based on the bootstrap estimators
  \eqn{\{\widehat{\bm{\beta}}_b\}^B_{b=1}}, this function computes the 
  asympototic variance estimator of the boostrap estimator, i.e.
    \deqn{\frac{n}{B} \sum^B_{i=1} \left(\widehat{\bm{\beta}}_b - 
    \widehat{\bm{\beta}}\right)  \left(\widehat{\bm{\beta}}_b - 
    \widehat{\bm{\beta}}\right)'.}
}
