% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{smatrixconvert}
\alias{smatrixconvert}
\title{Coerces a \code{sparseMatrix} as a \code{matrix}}
\usage{
smatrixconvert(mat)
}
\arguments{
\item{mat}{The matrix that is used in the \code{model} component.}
}
\value{
Returns a matrix in the \code{matrix} form.
}
\description{
This function coerces a \code{sparseMatrix} as a 
  \code{matrix}. This function is used specifically in the 
  \code{\link[lpinfer]{gurobi.optim}} function to ensure that the 
  arguments are accepted by the \code{gurobi} function.
}
