% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsf.R
\name{standard.form}
\alias{standard.form}
\title{Obtain standard form of linear program}
\usage{
standard.form(A, b, sense, lb = NULL, obj = NULL)
}
\arguments{
\item{A}{The matrix that represents the constraints.}

\item{b}{The rhs vector that represents the constraints.}

\item{sense}{The sense of the constraints.}

\item{lb}{The lower bound of the variables.}

\item{obj}{The objective function.}
}
\value{
Returns the following four objects:
  \item{A}{The matrix representing the constraints in standard form.}
  \item{b}{The rhs vector representing the constraints in standard form.}
  \item{lb}{The lower bound of the variables.}
  \item{obj}{The objective function.}
}
\description{
This function is mainly a wrapper of the \code{simplex}
   function from the \code{boot} package. It returns the standard form
   of the constraints matrix of a linear program.
}
