% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.bs.fn}
\alias{subsample.bs.fn}
\title{Carries out one bootstrap replication for the subsampling test}
\usage{
subsample.bs.fn(
  x,
  data,
  lpmodel,
  beta.tgt,
  norm,
  m,
  solver,
  replace,
  n,
  pbar,
  eval.count,
  n.bs,
  any.list,
  progress,
  omega.hat
)
}
\arguments{
\item{x}{This is either the list of indices that represent the bootstrap
replications, or the list of bootstrap components of the \code{lpmodel}
object passed from the user.}

\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{m}{The size of each subsample.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{replace}{A boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{pbar}{The progress bar object.}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}

\item{n.bs}{The total number of replications to be conducted in this
procedure.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{omega.hat}{The estimator of the asymptotic variance.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{Ts}{The bootstrap test statistic.}
  \item{beta}{The bootstrap estimate of \code{beta.obs}.}
  \item{msg}{An error message (if applicable).}
}
\description{
This function carries out the one bootstrap replication of the
  subsampling test. This function is used in the
  \code{\link[lpinfer]{subsample.bs}} function via the
  \code{future_map} function.
}
