% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.bs}
\alias{subsample.bs}
\title{Bootstrap procedure for the subsampling test}
\usage{
subsample.bs(
  data,
  i1,
  lpmodel,
  beta.tgt,
  norm,
  solver,
  replace,
  progress,
  m,
  n,
  omega.hat,
  df.error,
  eval.count,
  error.id,
  seed.list
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{replace}{Boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{m}{Size of each subsample.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{omega.hat}{Estimator of the asymptotic variance.}

\item{error.id}{List of ID that corresponds to problematic bootstrap draws.}

\item{seed.list}{List of RNG states and the corresponding iterations.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{T.sub}{Bootstrap test statistics from the subsampling procedure.}
  \item{beta.sub}{Bootstrap estimators for the \code{beta} component.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.eval}{Number of bootstrap replications that have been conducted.}
  \item{R.succ}{Number of successful bootstrap replications.}
}
\description{
This function carries out the bootstrap procedure of the
  subsampling test. This function supports parallel programming via the
  \code{future.apply} package.
}
