% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.check}
\alias{subsample.check}
\title{Checks and updates the input in \code{\link[lpinfer]{subsample}}}
\usage{
subsample.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  Rmulti,
  solver,
  norm,
  phi,
  n,
  replace,
  progress
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{Rmulti}{The multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{phi}{The tuning parameter for the subsampling test. The size of each
subsample is \eqn{n^\phi} where \eqn{\phi \in [0,1]}.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{replace}{A boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the updated parameters and objects back to the function
\code{subsample}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
      \item{\code{test.logical}}
      \item{\code{logical.lb}}
      \item{\code{logical.ub}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{subsample}} function. If there is any invalid input,
   the function will be terminated and error messages will be printed.
}
