% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.prob}
\alias{subsample.prob}
\title{Formulates and solves the \code{\link[lpinfer]{subsample}} problem}
\usage{
subsample.prob(
  data,
  lpmodel,
  beta.tgt,
  norm,
  solver,
  n,
  beta.obs.hat,
  omega.hat
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object used in the test.}

\item{beta.tgt}{The value to be tested.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{beta.obs.hat}{The value of sample \eqn{\hat{\bm{\beta}}_{\mathrm{obs}}}
from the \code{lpmodel} object.}

\item{omega.hat}{The estimator of the asymptotic variance.}
}
\value{
Returns the following list of outputs:
  \item{status}{The status of the optimization problem.}
  \item{x}{The optimal point.}
  \item{objval}{The optimal value.}
  \item{larg}{The list of arguments passed to the optimizer.}
  \item{beta}{The beta vector \eqn{\widehat{\bm{\beta}}_{\mathrm{obs}}}
    used in the optimization problem that is obtained from the
    \code{beta.obs} component of the \code{lpmodel} object.}
  \item{omega}{The Omega matrix \eqn{\widehat{\bm{\Omega}}_n} used in the
    optimization problem that is obtained from the \code{beta.obs}
    component of the \code{lpmodel} object.}
}
\description{
This function formulates and solves the linear or quadratic
  program in the \code{\link[lpinfer]{subsample}} procedure. If the user
  chooses a 1-norm, this function solves a linear program. If the user
  chooses a 2-norm, this function solves a quadratic program.
}
