% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Conducts inference using the subsampling procedure}
\usage{
subsample(
  data = NULL,
  lpmodel,
  beta.tgt,
  R = 100,
  Rmulti = 1.25,
  norm = 2,
  phi = 2/3,
  n = NULL,
  replace = FALSE,
  solver = NULL,
  progress = TRUE,
  previous.output = NA
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object used in the test.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{Rmulti}{The multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{phi}{The tuning parameter for the subsampling test. The size of each
subsample is \eqn{n^\phi} where \eqn{\phi \in [0,1]}.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{replace}{A boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{previous.output}{The list of outputs obtained from the previous
evaluation of the procedure. This is only used in the
\code{\link[lpinfer]{fsst}} test and when \code{previous.out} contains
the studentized matrix \code{omega.i}. The rest of the arguments
are ignored.}
}
\value{
Returns a list of output calculated from the function:
  \item{pval}{The \eqn{p}-value.}
  \item{T.n}{The test statistic \eqn{T_n}.}
  \item{T.bs}{The list of bootstrap estimates of the test statistics
    from the subsampling procedure.}
  \item{solver}{The solver used.}
  \item{cv.table}{A table of critical values.}
  \item{call}{The matched call.}
  \item{phi}{The \eqn{\phi} parameter used.}
  \item{norm}{The norm used.}
  \item{subsample.size}{The size of subsample}
  \item{test.logical}{An indicator variable for whether the computation has
    been conducted. If \code{test.logical} is 1, it refers to the case
    where \code{beta.tgt} is inside the logical bound. If
    \code{test.logical} is 0, it refers to the case where
    \code{beta.tgt} is outside the logical bound.}
  \item{logical.lb}{The logical lower bound.}
  \item{logical.ub}{The logical upper bound.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{The number of successful bootstrap replications.}
}
\description{
This function conducts inference and returns the
  \eqn{p}-value using the subsampling procedure.
}
\details{
There are three possible combinations for the parameters
\code{phi} and \code{replace}:
\itemize{
  \item{If \code{replace} is set as \code{FALSE}, it refers to the
    subsampling procedure. In this case, \code{phi} has to be inside the
    interval \eqn{(0, 1)}.}
  \item{If \code{replace} is set as \code{TRUE} and \code{phi} is set as 1,
    then it refers to the bootstrap procedure.}
  \item{If \code{replace} is set as \code{TRUE} and \code{phi} is inside the
    interval \eqn{(0, 1)}, then it refers to the \eqn{m} out of \eqn{n}
    bootstrap procedure, where \eqn{m} is the size of the subsample and
    \eqn{n} is the total number of observations.}
}
The following components are required in the \code{lpmodel} for the
subsampling procedure:
\itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
}
}
\section{Example}{

\preformatted{
  source("./example/dgp_missingdata.R") # Change directory if necessary
  J <- 5
  N <- 1000
  data <- missingdata_draw(J = J, n = N, seed = 1, prob.obs = .5)
  lpm <- missingdata_lpm(J = J, info = "full", data = data)
  subsample(data = data,
           lpmodel = lpm,
           beta.tgt = .2,
           R = 100,
           phi = 2/3,
           solver = "gurobi")
}
}

\section{More examples}{

  More examples can be found in the \code{subsample_example.R} file
  under the \code{example} subdirectory of the installation directory for
  the \code{lpinfer} package.
}

