% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_manycores}
\alias{subsample_manycores}
\title{Subsampling procedure with parallel programming}
\usage{
subsample_manycores(
  data,
  R,
  func_obs,
  func_var,
  A_obs,
  A_shp,
  A_tgt,
  beta_shp,
  beta_tgt,
  norm,
  solver,
  progress,
  m
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{R}{The total number of bootstraps \eqn{B} to be conducted.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{func_var}{Function that generates the asymptotic variance 
matrix of the estimator \eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{A_obs}{The "observed matrix" in the inference \eqn{A_{\mathrm{obs}}}.}

\item{A_tgt}{The "target matrix" in the inference \eqn{A_{\mathrm{tgt}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the inference procedure. If it is set as \code{TRUE}, the 
messages are displayed throughout the procedure. Otherwise, the messages
will not be displayed.}

\item{m}{subsample_onecore}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{T_sub}{List of test statistic from the subsampling procedure.}
  \item{pb}{Progress bar object.}
}
\description{
This function computes the list of test statistics that are 
  obtained from the subsampling procedure using parallel programming.
}
