% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_prob}
\alias{subsample_prob}
\title{Formulate and solve the subsampling problem}
\usage{
subsample_prob(
  df,
  func_obs,
  func_var,
  A_obs,
  A_shp,
  A_tgt,
  beta_shp,
  beta_tgt,
  lnorm,
  solver
)
}
\arguments{
\item{df}{The data being used in the inference.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{func_var}{Function that generates the asymptotic variance 
matrix of the estimator \eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{A_obs}{The "observed matrix" in the inference \eqn{A_{\mathrm{obs}}}.}

\item{A_tgt}{The "target matrix" in the inference \eqn{A_{\mathrm{tgt}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{lnorm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns a list of output that are obtained from the optimizer:
  \item{x}{Optimal point calculated from the optimizer.}
  \item{objval}{Optimal value calculated from the optimizer.}
  \item{larg}{List of arguments passed to the optimizer.}
  \item{beta}{The beta vector \eqn{\widehat{\bm{\beta}}_{\mathrm{obs}}}
    used in the optimization problem that is calculated from the 
    \code{func_obs} function with data \code{df}.}
  \item{omega}{The Omega matrix \eqn{\widehat{\bm{\Omega}}_n} used in the 
    optimization problem that is calculated from the \code{func_obs} 
    function with data \code{df}.}
}
\description{
Based on the sample data given by the data frame \code{df}, 
  this function formulates and solves linear/quadratic program in the 
  subsampling procedure.
}
