% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.bisection.print}
\alias{summary.bisection.print}
\title{Print results in constructing bounds in bisection method}
\usage{
\method{summary}{bisection.print}(df_bis, i, ...)
}
\arguments{
\item{df_bis}{A data frame storing the information from the bisection
method.}

\item{i}{The row number that we want to print for \code{df_bis}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function is used to display the message when constructing
   the bounds and used in \code{\link[lpinfer]{summary.invertci}} to print
   the results in each step of the bisection method.
}
