% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{summary.chorussell}
\alias{summary.chorussell}
\title{Summary of results from \code{chorussell}}
\usage{
\method{summary}{chorussell}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{chorussell}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{chorussell}.
}
