% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{summary.dkqs}
\alias{summary.dkqs}
\title{Summary of results from \code{\link[lpinfer]{dkqs}}}
\usage{
\method{summary}{dkqs}(x, ...)
}
\arguments{
\item{x}{An output object returned from \code{\link[lpinfer]{dkqs}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{\link[lpinfer]{dkqs}}.
}
