% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{summary.dkqs_cone}
\alias{summary.dkqs_cone}
\title{Summary of results from \code{dkqs_cone}}
\usage{
\method{summary}{dkqs_cone}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{dkqs_cone}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{dkqs_cone}.
}
\description{
This function uses the print method on the return list of the
   function \code{dkqs_cone}.
}
