% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{summary.estbounds}
\alias{summary.estbounds}
\title{Summary of results from \code{estbounds}}
\usage{
\method{summary}{estbounds}(x, ...)
}
\arguments{
\item{x}{Objects returned from \code{estbounds}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{estbounds}.
}
