% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{summary.fsst}
\alias{summary.fsst}
\title{Summary of results from \code{fsst}}
\usage{
\method{summary}{fsst}(x, ...)
}
\arguments{
\item{x}{Objects returned from \code{fsst}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{fsst}.
}
