% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.invertci}
\alias{summary.invertci}
\title{Summary of results from \code{\link[lpinfer]{invertci}}}
\usage{
\method{summary}{invertci}(x, alphas = NULL, ...)
}
\arguments{
\item{x}{The output object returned from \code{\link[lpinfer]{invertci}}.}

\item{alphas}{A list of alphas that the user would like to print.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function summarizes the results for
  \code{\link[lpinfer]{invertci}}.
}
