% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.invertci_multiple}
\alias{summary.invertci_multiple}
\title{Summary of results from \code{\link[lpinfer]{invertci}} for multiple
significance levels}
\usage{
summary.invertci_multiple(x, alphas, msg.bound, ...)
}
\arguments{
\item{x}{The output object returned from \code{\link[lpinfer]{invertci}}.}

\item{alphas}{A list of alphas that the user would like to print.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function summarizes the results for
  \code{\link[lpinfer]{invertci}}.
}
