% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{summary.lpmodel.natural}
\alias{summary.lpmodel.natural}
\title{Summary of the \code{lpmodel.natural} object}
\usage{
summary.lpmodel.natural(x, ...)
}
\arguments{
\item{x}{The \code{lpmodel.natural} object.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function is a wrapper of the \code{print.lpmodel.natural}
   function and prints the same information as the object
   \code{lpmodel.natural}.
}
