% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{summary.lpmodel}
\alias{summary.lpmodel}
\title{Summary of results from \code{lpmodel}}
\usage{
\method{summary}{lpmodel}(x, ...)
}
\arguments{
\item{x}{The \code{lpmodel} object.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{lpmodel}.
}
\description{
This function uses the summary method on the return list of the
   function \code{lpmodel}. This is a wrapper of the \code{print} command.
}
