% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{summary.mincriterion}
\alias{summary.mincriterion}
\title{Summary of results from \code{\link[lpinfer]{mincriterion}}}
\usage{
summary.mincriterion(x, ...)
}
\arguments{
\item{x}{The output objects returned from
\code{\link[lpinfer]{mincriterion}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function uses the summary method on the return list of the
   function \code{\link[lpinfer]{mincriterion}}.
}
