% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{summary.qpci}
\alias{summary.qpci}
\title{Summary of results from \code{qpci}}
\usage{
\method{summary}{qpci}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{qpci}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{qpci}.
}
\description{
This function uses the print method on the return list of the
   function \code{qpci}.
}
