% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{summary.subsample}
\alias{summary.subsample}
\title{Summary of results from \code{subsample}}
\usage{
summary.subsample(x, ...)
}
\arguments{
\item{x}{Object returned from \code{subsample}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{subsample}.
}
\description{
This function uses the print method on the return list of the
   function \code{subsample}. This is a wrapper of the \code{print} command.
}
