% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{summary_bisection_print}
\alias{summary_bisection_print}
\title{Print results in constructing bounds in bisection method}
\usage{
summary_bisection_print(df_bis)
}
\arguments{
\item{df_bis}{Data frame storing the information from the bisection method.}
}
\value{
Nothing is returned in this function.
}
\description{
This function is used inside the \code{summary.qpci} to print
   the results in each step of the bisection method.
}
