% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{beta.bs}
\alias{beta.bs}
\title{Evaluate the bootstrap betas}
\usage{
beta.bs(data, lpmodel, seed.list, df.error)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object used in the test.}

\item{seed.list}{The list of RNG states and the corresponding iterations.}
}
\value{
Returns the following objects:
  \item{beta.obs.bs}{The bootstrap estimates of \eqn{\beta_{\rm obs}}.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{error.id}{The list of problematic IDs.}
  \item{R.eval}{The number of bootstrap replications that have been
    conducted.}
  \item{R.succ}{The number of successful bootstrap replications.}
}
\description{
This function computes the bootstrap estimates of
  \eqn{\beta_{\rm obs}}.
}
