% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{beta.product}
\alias{beta.product}
\title{Computes vector products}
\usage{
beta.product(beta, beta.obs.hat, pbar, progress, eval.count)
}
\arguments{
\item{beta}{A bootstrap estimator \code{beta.obs}.}

\item{beta.obs.hat}{The sample estimator
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel} (and \code{data} if applicable).}

\item{pbar}{The progress bar object.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Returns an \eqn{n \times n} matrix.
    \item{beta.prod}{An \eqn{n \times n} matrix.}
}
\description{
This function computes the product of the two vectors. This is
  used in the \code{\link[lpinfer]{sigma.summation}} function that computes
  the asymptotic variance estimator.
}
\details{
Denote \eqn{\bm{\beta}} and \eqn{\hat{\bm{\beta}}_{\rm obs}} as
  the \eqn{n \times 1} vectors \code{beta} and \code{beta.obs.hat}
  respectively. This function returns the \eqn{n \times n} matrix
  \eqn{\widetilde{\bm{\beta}}\widetilde{\bm{\beta}}'} where
  \eqn{\widetilde{\bm{\beta}}} is defined as
  \eqn{\widetilde{\bm{\beta}} \equiv \bm{\beta} -
  \widehat{\bm{\beta}}_{\rm obs}}.
}
