% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{beta.star.qp}
\alias{beta.star.qp}
\title{Computes the starred components of \eqn{\widehat{\bm{\beta}}}}
\usage{
beta.star.qp(
  data,
  lpmodel,
  beta.tgt,
  weight.mat,
  beta.obs.hat,
  beta.sigma,
  solver
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{weight.mat}{The weighting matrix for the \code{\link[lpinfer]{fsst}}
procedure.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}
}
\value{
Returns the following objects:
  \item{beta.star}{The vector \eqn{\widehat{\bm{\beta}}_n^\star}.}
  \item{x}{The optimal point.}
  \item{status}{The status of the optimization problem.}
}
\description{
This function computes the vector
  \eqn{\widehat{\bm{\beta}}_n^\star} for the case where \eqn{d<p} in the
  \code{\link[lpinfer]{fsst}} procedure.
}
\details{
This corresponding to solving the following quadratic program
\deqn{
  \min_{x \in \mathbf{R}^d} \,\,
  \left(\hat{\beta}_{{\rm obs}, n} - A_{\rm obs} x\right)' \hat{\Xi}
  \left(\hat{\beta}_{{\rm obs}, n} - A_{\rm obs} x\right)
  \quad \mathrm{s.t.
  } \quad
  A_{\rm shp} x = \beta_{\rm shp}
  \quad \mathrm{ and } \quad
  A_{\rm tgt} x = \beta_{\rm tgt}
}
in the \code{\link[lpinfer]{fsst}} procedure.
}
