% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{beta.star.qp}
\alias{beta.star.qp}
\title{Computing \eqn{\widehat{\bm{\beta}}_n^\star}}
\usage{
beta.star.qp(
  data,
  lpmodel,
  beta.tgt,
  weight.matrix,
  beta.obs.hat,
  beta.sigma,
  solver
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{weight.matrix}{The option used in the weighting matrix.}

\item{beta.obs.hat}{Estimator of \eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}}
based on the given information in \code{lpmodel}.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the vector \eqn{\widehat{\bm{\beta}}_n^\star}.
  \item{beta.star}{The vector \eqn{\widehat{\bm{\beta}}_n^\star}.}
}
\description{
This function computes the vector
  \eqn{\widehat{\bm{\beta}}_n^\star} for the case where \eqn{d<p}.
}
\details{
This corresponding to solving problem (3) of Torgovitsky (2020).
  Three options for \code{weight.matrix} are available:
  \itemize{
    \item{\code{identity} --- identity matrix}
    \item{\code{diag} --- the diagonal matrix that takes the diagonal
       elements of the inverse of the variance matrix}
    \item{\code{avar} --- inverse of the variance matrix}
}
}
