% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{beta_bs}
\alias{beta_bs}
\title{Computes the test statistics via bootstrapping}
\usage{
beta_bs(
  df1,
  bs_seed,
  bs_size,
  bs_num,
  J,
  s_star,
  A_obs,
  A_tgt,
  beta_obs,
  beta_bs_bar,
  beta_tgt,
  tau
)
}
\arguments{
\item{df1}{The dataframe of the observed values, i.e. corresponds to the
case of \eqn{D = 1}.}

\item{bs_seed}{The starting value of the seed in the bootstrap.}

\item{bs_size}{The size of each bootstrap.}

\item{bs_num}{The total number of bootstraps.}

\item{J}{The number of distinct nonzero values in vector \eqn{\bm{y}}.}

\item{s_star}{The value of s_star in the cone-tightening procedure.}

\item{A_obs}{The observed matrix in the quadratic program.}

\item{A_tgt}{The "target matrix" in the quadratic program.}

\item{beta_obs}{The observed value of the vector beta.}

\item{beta_bs_bar}{The tau-tightened recentered bootstrap estimate.}

\item{beta_tgt}{The value of t in the null hypothesis.}

\item{tau}{The value of tau used in the quadratic program.}
}
\value{
Returns the list of test statistics obtained from bootstrap, i.e.
   \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.
}
\description{
This function computes the test statistics via bootstrapping.
}
