% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{beta_bs}
\alias{beta_bs}
\title{Computes the bootstrap test statistics}
\usage{
beta_bs(
  df,
  bs_seed,
  bs_num,
  J,
  s_star,
  A_obs,
  A_tgt,
  func_obs,
  beta_obs_hat,
  beta_tgt,
  tau,
  n,
  solver
)
}
\arguments{
\item{df}{The dataframe that contains the sample data.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_num}{The total number of bootstraps \eqn{B}.}

\item{J}{The number of distinct nonzero values in vector \eqn{\bm{y}}.}

\item{s_star}{The value of 
\eqn{\hat{s}^\ast \equiv A_{\mathrm{obs}}\hat{\bm{x}}_n^\ast} 
in the cone-tightening procedure.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta_obs_hat}{The value of \eqn{\hat{\beta}_{\mathrm{obs}}} based on
the function supplied by the user.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the value 
of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{tau}{The value of tau to be used in the linear program.}

\item{n}{The number of observations in the dataframe.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the list of estimates from bootstrap:
  \item{T_bs}{A list of bootstrap test statistics 
     \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.}
 \item{beta_bs_bar_set}{A list of \eqn{\tau_n}-tightened recentered 
    bootstrap estimates \eqn{\bar{\beta}^\ast_{\mathrm{obs},n,b}}}
}
\description{
This function computes the bootstrap test statistics.
}
