% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{bisec.eval}
\alias{bisec.eval}
\title{Evaluation of test statistic and check if the point has been evaluated}
\usage{
bisec.eval(f, farg, pt, df_ci, rngstate)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing
procedure.}

\item{pt}{Point to be evaluated in the bisection method.}

\item{df_ci}{Data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{rngstate}{The current RNG state obtained from \code{.Random.seed}.}
}
\value{
Returns the \eqn{p}-value of the point considered and an updated
   data frame that contains the points and the \eqn{p}-values.
   \item{pval}{\eqn{p}-value of the point.}
   \item{df_ci}{Updated data frame that consists of the points that have
      been tested in constructing the confidence intervals.}
}
\description{
This function checks if the \eqn{p}-value for the point
   considered has already been evaluated in previous iterations or provided
   by the user. The function will compute the \eqn{p}-value if it has been
   evaluated. Otherwise, it will use the previous data.
}
