% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bs.assign}
\alias{bs.assign}
\title{Auxiliary function in the bootstrap replications}
\usage{
bs.assign(
  R,
  R.eval,
  R.succ,
  maxR,
  any.list,
  lpmodel,
  data = NULL,
  m,
  replace = TRUE
)
}
\arguments{
\item{R}{Number of bootstrap replications.}

\item{R.eval}{Number of bootstrap replications that has been evaluated/}

\item{R.succ}{Number of successful bootstrap replications.}

\item{maxR}{Maximum number of bootstrap replications.}

\item{any.list}{An object returned from the \code{lpmodel.anylist}
function, which indicates whether at least one of the components in
\code{lpmodel} is a \code{list}.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{data}{Data used in the tests.}

\item{m}{Size of each subsample.}

\item{replace}{Boolean variable to indicate whether the function samples
the data with or without replacement.}
}
\value{
Returns the following three objects.
  \item{i0}{Starting index to be evaluated.}
  \item{i1}{Last index to be evaluated.}
  \item{bs.list}{List of bootstrap replications in the \code{lpmodel}
    object passed from the user.}
}
\description{
This function is used in the bootstrap replications of the
  tests. It returns the indices to be considered in the
  bootstrap replications and the list to be passed to the
  \code{lpmodel.anylist} function.
}
