% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.Ab}
\alias{check.Ab}
\title{Check function: constraint matrix and the corresponding rhs vector}
\usage{
check.Ab(A, b, Aname, bname)
}
\arguments{
\item{A}{Constraint matrix.}

\item{b}{Corresponding rhs vector.}

\item{Aname}{Name of the constraint matrix.}

\item{bname}{Name of the corresponding rhs vector.}
}
\value{
Returns the updated matrix and rhs vector or print a
   stop message.
   \item{A}{Updated constraint matrix.}
   \item{b}{Updated rhs vector.}
}
\description{
This function checks the constraint matrix \eqn{\bm{A}} and
   the corresponding vector \eqn{\bm{\beta}}.
}
