% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.betatgt.lp}
\alias{check.betatgt.lp}
\title{Construct the linear program for the function \code{check.betatgt}}
\usage{
check.betatgt.lp(data, lpmodel, beta.tgt, modelsense, solver)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{modelsense}{String that indicates whether the program is a 
maximization or minimization problem. If it is "max", then it is referring
to a maximization problem. Otherewise, it is referring to a minimization
problem.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the linear program.
  \item{objval}{Objective value of the linear program. It is either the
  logical upper bound or the logical lower bound.}
}
\description{
This function checks whether the parameter \code{beta.tgt}
  specified is within the logical bound. If it is not the logical
  bound, then reject immediately.
}
