% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.betatgt.lp}
\alias{check.betatgt.lp}
\title{Construct the linear program for the function \code{check.betatgt}}
\usage{
check.betatgt.lp(data, lpmodel, beta.tgt, modelsense, solver)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{Value of beta to be tested.}

\item{modelsense}{String that indicates whether the program is a
maximization or minimization problem. If it is "max", then it is referring
to a maximization problem. Otherwise, it is referring to a minimization
problem.}

\item{solver}{A linear or quadratic programming solver. The exact solver
that is supported depends on the test chosen.}
}
\value{
Returns the solution to the linear program.
  \item{objval}{Objective value of the linear program. It is either the
  logical upper bound or the logical lower bound.}
}
\description{
This function solves the linear program that is used to find
  the logical bounds in \code{check.betatgt} function.
}
