% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.betatgt}
\alias{check.betatgt}
\title{Check function: check if \code{beta.tgt} is within the logical bound}
\usage{
check.betatgt(data, lpmodel, beta.tgt, solver)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{Value of beta to be tested.}

\item{solver}{A linear or quadratic programming solver. The exact solver
that is supported depends on the test chosen.}
}
\value{
Returns an indicator to show whether the \code{beta.tgt} is within
  the logical bound or not.
  \item{beta.tgt.logi}{Equals 1 if it is within the logical bound.
  Equals 0 otherwise.}
}
\description{
This function checks whether the parameter \code{beta.tgt}
  specified is within the logical bounds. If it is not within the logical
  bound, then reject immediately.
}
