% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.dataframe}
\alias{check.dataframe}
\title{Check function: data frame}
\usage{
check.dataframe(data)
}
\arguments{
\item{data}{A \code{data} object.}
}
\value{
Returns the object as a \code{data.frame}.
}
\description{
This function coerce the data provided from the user as a
  \code{data.frame}.
}
