% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.dataframe}
\alias{check.dataframe}
\title{Check function: data frame}
\usage{
check.dataframe(data, name.var)
}
\arguments{
\item{data}{Data frame to be checked.}

\item{name.var}{Name of the variable.}
}
\value{
Returns the object in the format of \code{matrix}.
}
\description{
This function checks whether the data is a \code{data.frame}
  or \code{data.table} or \code{matrix} or \code{tibble}. If yes, the 
  object will be returned as an object in the \code{matrix} class. If not, 
  an error message is displayed.
}
