% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.errormsg}
\alias{check.errormsg}
\title{General error for checking the objects}
\usage{
check.errormsg(name.var, needs.to.be)
}
\arguments{
\item{name.var}{The name of the object.}

\item{needs.to.be}{What the object is supposed to be.}
}
\value{
Nothing is returned.
}
\description{
This is a general function to produce the error messages in
  the check functions.
}
