% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.lpmodel}
\alias{check.lpmodel}
\title{Check function: check lpmodel}
\usage{
check.lpmodel(
  data,
  lpmodel,
  name.var,
  A.tgt.cat,
  A.obs.cat,
  A.shp.cat,
  beta.obs.cat,
  beta.shp.cat,
  R
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{name.var}{Name of the \code{lpmodel} object.}

\item{A.tgt.cat}{Category of the \code{A.tgt} object.}

\item{A.obs.cat}{Category of the \code{A.obs} object.}

\item{A.shp.cat}{Category of the \code{A.shp} object.}

\item{beta.obs.cat}{Category of the \code{beta.obs} object.}

\item{beta.shp.cat}{Category of the \code{beta.shp} object.}

\item{R}{Number of bootstraps chosen by the users.}
}
\value{
Returns the updated \code{lpmodel} object.
   \item{lpmodel}{Updated \code{lpmodel} object.}
}
\description{
This function checks if the object \code{lpmodel} is in the
   correct format.
}
\details{
In each of the testing procedures, there are five possible
categories of the each of the object:
  \itemize{
    \item{Category '\code{not_used}': This refers to the case where the 
      object is not used in the function}
    \item{Category '\code{matrix}': This refers to the case where the
      object has to be a matrix}
    \item{Category '\code{function_mat}': This refers to the case where 
      the object has to be a function that produces a matrix.}
    \item{Category '\code{list}': This refers to the case where the
      object is a list}
    \item{Category '\code{function_obs_var}': This refers to the case 
      where the object is a function that produces a list that contains 
      a vector and a matrix. This is typically the case for \code{beta.obs}
      that the function produces an estimator and the asymptotic variance.}
  }
  Each object can belong to one of more categories.
}
