% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.lpmodel}
\alias{check.lpmodel}
\title{Check function: \code{lpmodel}}
\usage{
check.lpmodel(
  data,
  lpmodel,
  name.var,
  A.tgt.cat,
  A.obs.cat,
  A.shp.cat,
  beta.obs.cat,
  beta.shp.cat,
  R,
  is.estbounds = FALSE
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{name.var}{Name of the \code{lpmodel} object.}

\item{A.tgt.cat}{Category of the \code{A.tgt} object.}

\item{A.obs.cat}{Category of the \code{A.obs} object.}

\item{A.shp.cat}{Category of the \code{A.shp} object.}

\item{beta.obs.cat}{Category of the \code{beta.obs} object.}

\item{beta.shp.cat}{Category of the \code{beta.shp} object.}

\item{R}{Number of bootstrap replications.}

\item{is.estbounds}{Boolean variable that indicates whether the test
function being called is \code{estbounds}. If this function is being
called, then it means that when \code{data} is \code{NULL}, then each
component of the \code{lpmodel} object is not required to be a
\code{list}. Otherwise, the \code{lpmodel} object needs to contain the
bootstrap estimates if \code{data} is \code{NULL}.}
}
\value{
Returns the updated \code{lpmodel} object.
   \item{lpmodel}{Updated \code{lpmodel} object.}
}
\description{
This function checks if the object \code{lpmodel} is in the
   correct format.
}
\details{
In each of the testing procedures, there are five possible
categories for the each of the object:
\itemize{
  \item{\code{not_used}: This refers to the case where the object is not
   used in the function.}
  \item{\code{matrix}: This refers to the case where the object has to be
   a matrix.}
 \item{\code{function_mat}: This refers to the case where
   the object has to be a function that produces a matrix.}
 \item{\code{list}: This refers to the case where the object is a list.}
 \item{\code{function_obs_var}: This refers to the case where
   the object is a function that produces a list that contains a matrix
   and a vector. This is typically the case for \code{beta.obs} when
   the testing procedure requires both the observed value of
   \code{beta.obs} and the estimator of the asymptotic variance.}
}
 Each object can belong to one of more categories.
}
