% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.lpobjects}
\alias{check.lpobjects}
\title{Check function: check matrices and vectors in lpmodel}
\usage{
check.lpobjects(data, mat, mat.name, mat.cat, R)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{mat}{The matrix object in \code{lpmodel}.}

\item{mat.name}{Name of the matrix object in \code{lpmodel}.}

\item{mat.cat}{Category of the matrix object.}

\item{R}{Number of bootstraps chosen by the users.}
}
\value{
Returns two objects:
   \item{mat}{The updated object in \code{lpmodel}.}
   \item{sample}{A sample object that is used to compare the dimension.}
}
\description{
This function checks if the matrix objects in \code{lpmodel}
   are in the correct format.
}
\details{
See the details for the function `\code{check.lpmodel}` for the 
  details on the strings for each category.
}
