% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.matrix}
\alias{check.matrix}
\title{Check function: check matrix}
\usage{
check.matrix(mat, mat.name, mat.cat, inside.list)
}
\arguments{
\item{mat}{The matrix object in \code{lpmodel}.}

\item{mat.name}{Name of the matrix object in \code{lpmodel}.}

\item{inside.list}{Indicator variables of whether the object that is 
being checked is inside a list or not.}
}
\description{
This function checks if the matrix objects in \code{lpmodel} 
   are in the correct format.
}
