% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.nonnegaetive}
\alias{check.nonnegaetive}
\title{Check function: Nonnegative number}
\usage{
check.nonnegaetive(x, name.var)
}
\arguments{
\item{x}{Variable to be checked.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the class of the variable
  is \code{numeric}, has length 1 and is a nonnegative number. If not, an
  error message is displayed.
}
