% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.numeric}
\alias{check.numeric}
\title{Check function: numeric}
\usage{
check.numeric(x, name.var)
}
\arguments{
\item{x}{Variable to be checked.}
}
\value{
Nothing is returned.
}
\description{
This function checks whether the class of the variable
  is \code{numeric} and has length 1.
}
