% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.vector}
\alias{check.vector}
\title{Check function: check vector}
\usage{
check.vector(vec, vec.name, inside.list)
}
\arguments{
\item{vec}{The vector object in \code{lpmodel}.}

\item{vec.name}{Name of the vector object in \code{lpmodel}.}

\item{inside.list}{Indicator variables of whether the object that is
being checked is inside a list or not.}
}
\description{
This function checks if the matrix objects in \code{lpmodel}
   are in the correct format.
}
