% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.bs.fn}
\alias{chorussell.bs.fn}
\title{Carries out one bootstrap replication for the Cho-Russell procedure}
\usage{
chorussell.bs.fn(
  x,
  data,
  lpmodel,
  beta.tgt,
  kappa,
  norm,
  n,
  estimate,
  solver,
  pbar,
  eval.count,
  n.bs,
  any.list,
  progress
)
}
\arguments{
\item{x}{This is either the list of indices that represent the bootstrap
replications, or the list of bootstrap components of the \code{lpmodel}
object passed from the user.}

\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{estimate}{A boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{pbar}{The progress bar object.}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}

\item{n.bs}{The total number of replications to be conducted in this
procedure.}

\item{any.list}{This object contains information on whether any components
in the \code{lpmodel} object is a \code{list}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns a list of output that are obtained from the Cho-Russell
  procedure:
  \item{ub}{The bootstrap estimate of the upper bound.}
  \item{ub}{The bootstrap estimate of the lower bound.}
  \item{kappa.error}{The \code{kappa} parameter that leads to an error (if
    applicable).}
  \item{msg}{An error message (if applicable).}
}
\description{
This function carries out one bootstrap replication of the
  Cho-Russell procedure. This function is used in the
  \code{\link[lpinfer]{chorussell.bs}} function via the \code{future_map}
  function.
}
