% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.bs}
\alias{chorussell.bs}
\title{Bootstrap procedure for the Cho-Russell procedure}
\usage{
chorussell.bs(
  data,
  lpmodel,
  beta.tgt,
  R,
  maxR,
  kappa,
  norm,
  n,
  estimate,
  solver,
  progress
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test that consists of
the following objects:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstrap replications.}

\item{maxR}{Maximum number of bootstrap replications to be considered in
case there are some errors.}

\item{kappa}{Tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{estimate}{Boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns a list of output that are obtained from the Cho-Russell
  procedure:
  \item{ub.bs}{List of upper bounds from bootstrap data.}
  \item{lb.bs}{List of lower bounds from bootstrap data.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.eval}{Number of bootstrap replications that have been conducted.}
  \item{R.succ}{Number of successful bootstrap replications.}
}
\description{
This function carries out the bootstrap procedure of the
  Cho-Russell procedure. This function supports parallel programming via the
  \code{future.apply} package.
}
