% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.check}
\alias{chorussell.check}
\title{Checks and updates the input in the Cho-Russell procedure}
\usage{
chorussell.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  Rmulti,
  kappa,
  norm,
  n,
  estimate,
  solver,
  ci,
  alpha,
  tol,
  progress
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test that consists of
the following objects:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstrap replications.}

\item{Rmulti}{Multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{kappa}{Tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{estimate}{Boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{alpha}{Significance level.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the updated parameters back to the function
  \code{chorussell}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{seed}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{test.logical}}
      \item{\code{n}}
   }
}
\description{
This function checks and updates the input of the user. If
   there is any invalid input, this function will be terminated and
   generates appropriate error messages.
}
