% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell.pt}
\alias{chorussell.pt}
\title{Checks if \code{beta.tgt} is inside the \eqn{(1-\alpha)}-confidence interval}
\usage{
chorussell.pt(cr.lp.return, beta.tgt)
}
\arguments{
\item{cr.lp.return}{List of objects returned from
\code{\link[lpinfer]{chorussell}}.}

\item{beta.tgt}{The value to be tested.}
}
\value{
Returns the decision.
  \item{decision}{A boolean variable that equals to \code{TRUE} if
    \code{beta.tgt} is inside the \eqn{(1-\alpha)}-confidence interval, and
    equals to \code{FALSE} otherwise.}
}
\description{
This function checks whether \code{beta.tgt} is inside the
  \eqn{(1-\alpha)}-confidence interval from the
  \code{\link[lpinfer]{chorussell}} procedure.
}
