% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell}
\alias{chorussell}
\title{Conducts inference using the Cho-Russell procedure}
\usage{
chorussell(
  data = NULL,
  lpmodel,
  beta.tgt,
  n = NULL,
  R = 100,
  Rmulti = 1.25,
  kappa = 1e-05,
  norm = 2,
  estimate = TRUE,
  solver = NULL,
  ci = FALSE,
  alpha = 0.05,
  tol = 1e-04,
  progress = TRUE,
  remove.const = TRUE
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test that consists of
the following objects:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{R}{Number of bootstrap replications.}

\item{Rmulti}{Multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{kappa}{Tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{estimate}{Boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{alpha}{Significance level.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{remove.const}{Determine whether the constraints are to be removed.}
}
\value{
Returns the following list of output:
  \item{ub}{Upper bound from original data.}
  \item{lb}{Lower bound from original data.}
  \item{ub.bs}{List of upper bounds from bootstrap data.}
  \item{lb.bs}{List of lower bounds from bootstrap data.}
  \item{test.logical}{Indicator variable for whether the computation has
    been conducted. If \code{test.logical} is 1, it refers to the case
    where \code{beta.tgt} is inside the logical bound. If
    \code{test.logical} is 0, it refers to the case where
    \code{beta.tgt} is outside the logical bound.}
  \item{logical.lb}{Logical lower bound.}
  \item{logical.ub}{Logical upper bound.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{Number of successful bootstrap replications.}
  \item{ci}{Boolean variable that indicates whether a \eqn{p}-value or a
    \eqn{(1-\alpha)}-confidence interval is returned.}
  \item{pval}{\eqn{p}-value (if \code{ci} is set as \code{FALSE}).}
  \item{c.ub}{Upper bound of the \eqn{(1-\alpha)}-confidence interval.}
  \item{c.lb}{Lower bound of the \eqn{(1-\alpha)}-confidence interval.}
  \item{alpha}{Significance level.}
  \item{iter}{Total number of iterations (if \code{ci} is \code{FALSE}.)}
  \item{unique}{Whether the solution is unique.}
}
\description{
This function conducts inference using the procedure
  proposed by Cho and Russell (2019).
}
\details{
See the details section of \code{\link[lpinfer]{estbounds}} for a
  list of strings acceptable for the option \code{norm}.
}
\examples{
## ========================================================================= ##
##
##  Example for the chorussell function
##
##  This followings illustrate how the function can be used to compute
##  p-values and construct confidence intervals using the missing data
##  problem.
##
## ========================================================================= ##
rm(list = ls())

# ---------------- #
# Part 1: Load required packages
# ---------------- #
library(lpinfer)
library(future)

# ---------------- #
# Part 2: Data and lpmodel preparation
# ---------------- #
source("./example/dgp_missingdata.R")
J <- 5
N <- 1000
data <- missingdata_draw(J = J, n = N, seed = 1, prob.obs = .5)
lpmodel.full <- missingdata_lpm(J = J, info = "full", data = data)
lpmodel.twom <- missingdata_lpm(J = J, info = "mean", data = data)

tau <- sqrt(log(N)/N)
beta.tgt <- .2
kappa <- 1e-5

# ---------------- #
# Step 3: Run chorussell
# ---------------- #
# Example 1: By default, the function is used to find the p-values,
# i.e. ci = FALSE
set.seed(1)
cr1 <- chorussell(data = data,
                  lpmodel = lpmodel.full,
                  beta.tgt = beta.tgt,
                  R = 100,
                  norm = 1,
                  kappa = kappa,
                  solver = "gurobi")

# Example 2: Sample as example 1 except that the two-moments approach is used
set.seed(1)
cr2 <- chorussell(data = data,
                  lpmodel = lpmodel.twom,
                  beta.tgt = beta.tgt,
                  R = 100,
                  norm = 1,
                  kappa = kappa,
                  solver = "gurobi")

# Example 3: Find the 95\% confidence interval with full-information approach
# The default is to set the signficance level as 0.05
set.seed(1)
cr3 <- chorussell(data = data,
                  lpmodel = lpmodel.full,
                  beta.tgt = beta.tgt,
                  R = 100,
                  norm = 1,
                  kappa = kappa,
                  ci = TRUE,
                  solver = "gurobi")

# Example 4: Find the 90\% confidence interval with full-information approach
set.seed(1)
cr4 <- chorussell(data = data,
                  lpmodel = lpmodel.full,
                  beta.tgt = beta.tgt,
                  R = 100,
                  norm = 1,
                  kappa = kappa,
                  ci = TRUE,
                  alpha = .1,
                  solver = "gurobi")

# Example 5: Multiple lambda can be specified
set.seed(1)
cr5 <- chorussell(data = data,
                  lpmodel = lpmodel.full,
                  beta.tgt = beta.tgt,
                  R = 100,
                  norm = 1,
                  kappa = c(1e-10, 1e-5, 1e-1),
                  ci = TRUE,
                  alpha = .1,
                  solver = "gurobi")


# Example 6: Uses parallel programming by the future package
plan(multisession, workers = 8)
set.seed(1)
cr6 <- chorussell(data = data,
                  lpmodel = lpmodel.full,
                  beta.tgt = beta.tgt,
                  R = 100,
                  norm = 1,
                  kappa = c(1e-10, 1e-5),
                  solver = "gurobi")

# Example 7: The confidence interval can also be obtained via the `invertci`
# command although it can be done without using the `invertci` command.
set.seed(1)
farg <- list(data = data,
             lpmodel = lpmodel.full,
             beta.tgt = beta.tgt,
             R = 100,
             norm = 1,
             kappa = kappa,
             solver = "gurobi",
             ci = TRUE,
             progress = FALSE)
cr7 <- invertci(f = chorussell,
                farg = farg,
                alpha = .05,
                max.iter = 100,
                tol = 1e-4)

# Example 8: To construct confidence interval for multiple alphas, this can be
# done by specifying multiple alpha in the `alpha` option
set.seed(1)
cr8 <- invertci(f = chorussell,
                farg = farg,
                alpha = c(.01, .1),
                max.iter = 100,
                tol = 1e-5)
}
