% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{chorussell}
\alias{chorussell}
\title{Conducts inference using the Cho-Russell procedure}
\usage{
chorussell(
  data = NULL,
  lpmodel,
  beta.tgt = NULL,
  n = NULL,
  R = 100,
  Rmulti = 1.25,
  kappa = 0,
  norm = 2,
  estimate = TRUE,
  solver = NULL,
  ci = NULL,
  alpha = 0.05,
  tol = 1e-04,
  progress = TRUE,
  remove.const = TRUE,
  previous.output = NA
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{R}{The number of bootstrap replications.}

\item{Rmulti}{The multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number or a vector of nonnegative numbers.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{estimate}{A boolean variable to indicate whether the bounds should be
estimated or not.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{ci}{A boolean variable that indicates whether a \eqn{p}-value or a
\eqn{(1-\alpha)}-confidence interval is returned. If \code{ci} is
\code{TRUE}, then a confidence interval is returned. Otherwise, a
\eqn{p}-value is returned.}

\item{alpha}{The significance level. This can be a vector.}

\item{tol}{The tolerance level in the bisection procedure.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{remove.const}{A boolean variable. This determine whether the
constraints are to be removed.}

\item{previous.output}{The list of outputs obtained from the previous
evaluation of the procedure. This is only used in the
\code{\link[lpinfer]{fsst}} test and when \code{previous.out} contains
the studentized matrix \code{omega.i}. The rest of the arguments
are ignored.}
}
\value{
Returns the following list of output:
  \item{ub}{The upper bound from original data.}
  \item{lb}{The lower bound from original data.}
  \item{ub.bs}{The list of upper bounds from bootstrap data.}
  \item{lb.bs}{The list of lower bounds from bootstrap data.}
  \item{test.logical}{An indicator variable for whether the computation has
    been conducted. If \code{test.logical} is 1, it refers to the case
    where \code{beta.tgt} is inside the logical bound. If
    \code{test.logical} is 0, it refers to the case where
    \code{beta.tgt} is outside the logical bound.}
  \item{logical.lb}{The logical lower bound.}
  \item{logical.ub}{The logical upper bound.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{The number of successful bootstrap replications.}
  \item{ci}{A boolean variable that indicates whether a \eqn{p}-value or a
    \eqn{(1-\alpha)}-confidence interval is returned.}
  \item{pval}{\eqn{p}-value (if \code{ci} is set as \code{FALSE}).}
  \item{c.ub}{The upper bound of the \eqn{(1-\alpha)}-confidence interval.}
  \item{c.lb}{The lower bound of the \eqn{(1-\alpha)}-confidence interval.}
  \item{alpha}{The significance level.}
  \item{iter}{The total number of iterations (if \code{ci} is \code{FALSE}.)}
  \item{unique}{A boolean variable showing whether the solution is unique.}
}
\description{
This function conducts inference using the procedure
  proposed by Cho and Russell (2019).
}
\details{
\itemize{
 \item{See the details section of the \code{\link[lpinfer]{estbounds}}
    function for a list of strings acceptable for the option \code{norm}.}
 \item{The following components are required in the \code{lpmodel} for the
   Cho-Russell procedure:
   \itemize{
     \item{\code{A.tgt}}
     \item{\code{A.obs}}
     \item{\code{A.shp}}
     \item{\code{beta.obs}}
     \item{\code{beta.shp}}
   }
 \item{The input \code{beta.tgt} is not required when \code{ci = TRUE}.}
 }
}
}
\section{Example}{

\preformatted{
  source("./example/dgp_missingdata.R") # Change directory if necessary
  J <- 5
  N <- 1000
  data <- missingdata_draw(J = J, n = N, seed = 1, prob.obs = .5)
  lpm <- missingdata_lpm(J = J, info = "full", data = data)
  chorussell(data = data,
             lpmodel = lpm,
             beta.tgt = .2,
             R = 100,
             norm = 2,
             ci = TRUE,
             solver = "gurobi")
}
}

\section{More examples}{

  More examples can be found in the \code{chorussell_example.R} file
  under the \code{example} subdirectory of the installation directory for
  the \code{lpinfer} package.
}

