% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{ci.bisection}
\alias{ci.bisection}
\title{Bisection method for constructing confidence intervals}
\usage{
ci.bisection(
  f,
  farg,
  alpha,
  b0,
  b1,
  tol,
  max.iter,
  df_ci,
  progress,
  type,
  dp,
  rngstate,
  para.match
)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing
procedure.}

\item{alpha}{Significance level of the test.}

\item{b0}{Logical lower or upper bound for the confidence interval.}

\item{b1}{Maximum possible lower bound or minimum possible upper bound for
the confidence interval.}

\item{tol}{Tolerance level in the bisection method.}

\item{max.iter}{Maximum number of iterations in the bisection method.}

\item{df_ci}{Data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the procedure of constructing confidence interval. If
it is set as \code{TRUE}, the messages are displayed throughout the
procedure. Otherwise, the messages will not be displayed.}

\item{type}{Type of the confidence interval. Set \code{type} as 1 for the
upper bound of the confidence interval and set \code{type} as -1 for the
lower bound of the confidence interval.}

\item{dp}{Number of decimal places to be displayed for the \eqn{p}-values
and confidence intervals in the messages if \code{progress} is set
as \code{TRUE}.}

\item{rngstate}{The current RNG state obtained from \code{.Random.seed}.}

\item{para.match}{The list of parameters to match.}
}
\value{
Return the solution of the bisection method and the updated
   data frame.
   \item{soln}{Solution to the bisection method.}
   \item{df_ci}{Data frame that consists of the points and the
      corresponding \eqn{p}-values that have been tested in constructing the
      confidence intervals.}
   \item{df_bis}{Data frame storing the information for the bisection
      method in each iteration.}
}
\description{
This function constructs the two-sided confidence interval
   of a given testing procedure using the bisection method.
}
