% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{consolidate.invertci}
\alias{consolidate.invertci}
\title{Function to consolidate the print the summary table for the interations}
\usage{
consolidate.invertci(df, msg)
}
\value{
Returns the consolidate table.
}
\description{
This function is used to consolidate the summary table for
  display via the \code{summary} command.
}
