% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{cplexapi.optim}
\alias{cplexapi.optim}
\title{LP and QP solver by cplexAPI}
\usage{
cplexapi.optim(Af, bf, nf, A, rhs, sense, modelsense, lb, weight = NULL)
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the quadratic and linear
   programs using the `\code{cplexAPI}' package.
}
