% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{cplexapi_optim}
\alias{cplexapi_optim}
\title{cplexAPI solver for quadratic and linear programs}
\usage{
cplexapi_optim(obj2, obj1, obj0, A, rhs, sense, modelsense, lb)
}
\value{
Returns the optimal objective value and the corresponding argument
  to the quadratic or linear program.
}
\description{
This function computes the solution to the quadratic and linear
   programs using the `\code{cplexAPI}' package.
}
