% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R, R/dkqs_objfunction_separate_obj210_Ab.R
\name{cplexapi_optim}
\alias{cplexapi_optim}
\title{LP and QP solver by cplexAPI}
\usage{
cplexapi_optim(
  obj2,
  obj1,
  obj0,
  A,
  rhs,
  sense,
  modelsense,
  lb,
  A_obs,
  beta_hat_obs,
  n
)

cplexapi_optim(
  obj2,
  obj1,
  obj0,
  A,
  rhs,
  sense,
  modelsense,
  lb,
  A_obs,
  beta_hat_obs,
  n
)
}
\arguments{
\item{n}{The number of observations in the dataframe.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}

Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the quadratic and linear
   programs using the `\code{cplexAPI}' package.

This function computes the solution to the quadratic and linear
   programs using the `\code{cplexAPI}' package.
}
