% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs.bs}
\alias{dkqs.bs}
\title{Bootstrap procedure for the DKQS test}
\usage{
dkqs.bs(
  data,
  lpmodel,
  beta.tgt,
  R,
  maxR,
  s.star.list,
  tau.list,
  solver,
  progress,
  seed,
  n
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstrap replications.}

\item{maxR}{Maximum number of bootstrap replications to be considered in
case there are some errors.}

\item{s.star.list}{The list of values of
\eqn{\hat{\bm{s}}^\star \equiv \bm{A}_{\mathrm{obs}}\hat{\bm{x}}_n^\star}
in the cone-tightening procedure for each \eqn{\tau}.}

\item{tau.list}{The list of feasible parameters \eqn{\tau}.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{seed}{The \code{.Random.seed} obtained in the beginning of the code.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}
}
\value{
Returns the list of estimates from bootstrap:
  \item{T.bs}{A list of bootstrap test statistics
     \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.}
 \item{beta.bs.bar.list}{A list of \eqn{\tau_n}-tightened recentered
    bootstrap estimates \eqn{\bar{\bm{\beta}}^\star_{\mathrm{obs},n,b}}.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{Number of successful bootstrap replications.}
}
\description{
This function carries out the bootstrap procedure of the
  DKQS test. This function supports parallel programming via the
  \code{future.apply} package.
}
