% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs}
\alias{dkqs}
\title{Conducts inference using the DKQS procedure}
\usage{
dkqs(
  data = NULL,
  lpmodel,
  beta.tgt,
  R = 100,
  Rmulti = 1.25,
  tau = NULL,
  n = NULL,
  solver = NULL,
  progress = TRUE
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstrap replications.}

\item{Rmulti}{Multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{tau}{Value of the tuning parameter \eqn{\tau} in the DKQS
procedure.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the following list of outputs:
  \item{pval}{A table of \eqn{p}-values for each \eqn{\tau}.}
  \item{tau.feasible}{The list of \eqn{\tau} that are feasible.}
  \item{tau.infeasible}{The list of \eqn{\tau} that are infeasible.}
  \item{tau.max}{Maximum value of the feasible \eqn{\tau} for the problem.}
  \item{T.n}{Test statistic \eqn{T_n}.}
  \item{T.bs}{The list of bootstrap test statistics
     \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}} for each \eqn{\tau}.}
  \item{beta.bs.bar}{The list of \eqn{\tau}-tightened re-centered bootstrap
     estimators \eqn{\bar{\beta}^\ast_{\mathrm{obs},n,b}}.}
  \item{lb0}{Logical lower bound of the problem for each \eqn{\tau}.}
  \item{ub0}{Logical upper bound of the problem for each \eqn{\tau}.}
  \item{solver}{Solver used.}
  \item{cv.table}{Table of critical values.}
  \item{call}{The function that has been called.}
  \item{test.logical}{Indicator variable for whether the computation has
    been conducted. If \code{test.logical} is 1, it refers to the case
    where \code{beta.tgt} is inside the logical bounds. If
    \code{test.logical} is 0, it refers to the case where
    \code{beta.tgt} is outside the logical bounds.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.succ}{Number of successful bootstrap replications.}
}
\description{
This module conducts inference using the cone-tightening
  procedure proposed by Deb, Kitamura, Quah and Stoye (2018).
}
\details{
If the value of the test statistic \eqn{T_n} is zero, the
   bootstrap procedure will be skipped and the \eqn{p}-value is zero.
}
