% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs_cone}
\alias{dkqs_cone}
\title{Compute p-value of a quadratic program}
\usage{
dkqs_cone(
  df,
  A_obs,
  A_tgt,
  func_obs,
  beta_tgt,
  bs_seed = 1,
  bs_num = 100,
  p_sig = 2,
  tau_input = 0.5
)
}
\arguments{
\item{df}{The dataframe that contains the sample data.}

\item{A_obs}{The observed matrix in the linear program.}

\item{A_tgt}{The "target matrix" in the linear program.}

\item{func_obs}{The function that generates the required 
\eqn{\beta_{\mathrm{obs}}}.}

\item{beta_tgt}{The value of \eqn{t} in the null hypothesis.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_num}{The total number of bootstraps.}

\item{p_sig}{The number of decimal places in the \eqn{p}-value.}

\item{tau_input}{The value of tau chosen by the user.}
}
\value{
Returns the \eqn{p}-value, the value of tau used \eqn{\tau^\ast}, 
  test statistic \eqn{T_n(\tau_n)} and the list of bootstrap test statistics 
  \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}}.
}
\description{
This function runs the procedure suggested by Torgovitsky (2019)
   using the cone-tightening procedure proposed by Deb, Kitamura, Quah and
   Stoye (2018).
}
