% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs_cone}
\alias{dkqs_cone}
\title{Computes the p-value of a quadratic program}
\usage{
dkqs_cone(
  df,
  A_obs,
  A_tgt,
  func_obs,
  beta_tgt,
  bs_seed = 1,
  bs_num = 100,
  p_sig = 2,
  tau_input = 0.5,
  lpsolver = NULL,
  qpsolver = NULL
)
}
\arguments{
\item{df}{The dataframe that contains the sample data.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the value 
of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_num}{The total number of bootstraps \eqn{B}.}

\item{p_sig}{The number of decimal places in the \eqn{p}-value.}

\item{tau_input}{The value of tau chosen by the user.}

\item{lpsolver}{The name of the linear programming solver package to 
be used to obtain the solutions to the linear programs. The linear 
programming solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}' and `\code{lpsolveAPI}'.}

\item{qpsolver}{The name of the quadratic programming solver package to 
be used to obtain the solutions to the quadratic programs. The quadratic 
programming solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}' and `\code{osqp}'.}
}
\value{
Returns the \eqn{p}-value, the value of tau used \eqn{\tau^\ast}, 
  test statistic \eqn{T_n(\tau_n)}, the list of bootstrap test statistics 
  \eqn{\{\overline{T}_{n,b}(\tau_n)\}^B_{b=1}} and the list of tau-tightened
  re-centered bootstrap estimators \eqn{\bar{\beta}^\ast_{\mathrm{obs},n,b}}.
}
\description{
This module conducts inference in quadratic programs using the 
   procedure suggested by Torgovitsky (2019) that incorporates the 
   cone-tightening procedure proposed by Deb, Kitamura, Quah and
   Stoye (2018).
}
