% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs_cone_check}
\alias{dkqs_cone_check}
\title{Checks and update the input}
\usage{
dkqs_cone_check(
  df,
  A_obs,
  A_tgt,
  func_obs,
  beta_tgt,
  bs_seed,
  bs_num,
  p_sig,
  tau_input,
  solver
)
}
\arguments{
\item{df}{The data being used in the inference.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_num}{The total number of bootstraps \eqn{B} to be conducted.}

\item{p_sig}{The number of decimal places in the \eqn{p}-value.}

\item{tau_input}{The value of tau chosen by the user.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the list of updated parameters as follows:
  \item{df}{Upated data in class \code{data.frame}}
  \item{A_obs}{Update "observed" matrix in class \code{matrix}.}
  \item{A_tgt}{Update "target" matrix in class \code{matrix}.}
  \item{beta_obs_tgt}{Obtain \eqn{\widehat{\bm{\beta}}_{\mathrm{tgt}}} 
     that is obtained from the function \code{func_obs}.}
  \item{solver}{Update name of solver in lower case.}
}
\description{
This function checks and updates the input of the user. If 
   there is any invalid input, this function will be terminated and 
   generates appropriate error messages.
}
