% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{dkqs_cone_errormsg}
\alias{dkqs_cone_errormsg}
\title{Error messages generator}
\usage{
dkqs_cone_errormsg(
  df,
  A_obs,
  A_tgt,
  beta_obs,
  beta_tgt,
  bs_seed,
  bs_size,
  bs_num,
  p_sig,
  tau_input
)
}
\arguments{
\item{df}{The dataframe that contains the sample data.}

\item{A_obs}{The observed matrix in the linear program.}

\item{A_tgt}{The "target matrix" in the linear program.}

\item{beta_obs}{The observed value of the vector beta.}

\item{beta_tgt}{The value of t in the null hypothesis.}

\item{bs_seed}{The starting value of the seed in bootstrap.}

\item{bs_size}{The size of each bootstrap sample.}

\item{bs_num}{The total number of bootstraps.}

\item{p_sig}{The number of decimal places in the \eqn{p}-value.}

\item{tau_input}{The value of tau chosen by the user.}
}
\description{
This function generates the error messages and stops the program
   if there is any invalid input.
}
