% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds.check}
\alias{estbounds.check}
\title{Checks and updates the input in \code{estbounds}}
\usage{
estbounds.check(data, lpmodel, kappa, norm, solver, estimate)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{estimate}{A boolean variable to indicate whether the bounds should be
estimated or not.}
}
\value{
Returns the updated parameters and objects back to the function
\code{estbounds}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
      \item{\code{kappa}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{estbounds}} function. If there is any invalid input,
   the function will be terminated and error messages will be printed.
}
