% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds.check}
\alias{estbounds.check}
\title{Checks and updates the input of the function \code{estbounds}}
\usage{
estbounds.check(data, lpmodel, kappa, norm, solver, estimate, progress)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{A.shp}}
  \item{\code{beta.obs}}
  \item{\code{beta.shp}}
}}

\item{kappa}{Parameter used in the second step of the two-step procedure
for obtaining the solution subject to the shape constraints. It can be
any nonnegative number.}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{estimate}{Boolean variable to indicate whether the estimated
problem should be considered.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
Returns the updated parameters and objects back to the function
\code{estbounds}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
      \item{\code{kappa}}
   }
}
\description{
This function checks and updates the input from the user for
   the function \code{estbounds}. If there is any invalid input, this
   function will terminate the procedure and generate appropriate error
   messages.
}
