% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds.original}
\alias{estbounds.original}
\title{Computes the true bounds with shape constraints}
\usage{
estbounds.original(data, lpmodel, original.sense, solver)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test that consists of
the following objects:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{original.sense}{Sense of the constraints for the true bounds.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}
}
\value{
Returns the solution to the linear program.
 \item{objval}{Optimal objective value.}
 \item{x}{Optimal point.}
}
\description{
This function computes the true bounds subject to the shape
   constraints without approximation.
}
