% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds.original}
\alias{estbounds.original}
\title{Computes the true bounds with shape contraints}
\usage{
estbounds.original(data, lpmodel, original.sense, solver)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear 
programming problems. The list of objects required in the \code{dkqs} 
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{original.sense}{Sense of the contraints to compute the true bound.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the linear program.
 \item{objval}{Optimal value calculated from the optimizer.}
 \item{x}{Optimal point calculated from the optimizer.}
}
\description{
The function computes the true bound subject to the shape 
   constraints without approximation.
}
