% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds}
\alias{estbounds}
\title{Estimate bounds with shape restrictions}
\usage{
estbounds(
  data = NULL,
  lpmodel,
  kappa = 1e-05,
  norm = 2,
  estimate = TRUE,
  solver = NULL,
  progress = TRUE
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{A.shp}}
  \item{\code{beta.obs}}
  \item{\code{beta.shp}}
}}

\item{kappa}{Parameter used in the second step of the two-step procedure
for obtaining the solution subject to the shape constraints. It can be
any nonnegative number.}

\item{norm}{Norm used in the optimization problem.}

\item{estimate}{Boolean variable to indicate whether the estimated
problem should be considered.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
Returns the bounds subject to the shape constraints.
  \item{ub}{Upper bound with shape constraints}
  \item{lb}{Lower bound with shape constraints}
  \item{est}{Indicator of whether estimation is involved in the
  estimation}
  \item{call}{The function that has been called.}
  \item{norm}{Norm used in the optimization problem.}
  \item{solver}{Name of the solver used.}
}
\description{
This function computes the bound of the linear program
   subject to shape constraints. This function also offers an option
   to estimate the shape constraints using a two-step procedure and
   some tolerance level. \eqn{\ell^1}-norm and \eqn{\ell^2}-norm are
   supported in the estimation procedure.
}
