% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds}
\alias{estbounds}
\title{Estimate bounds with shape restrictions}
\usage{
estbounds(
  data,
  func_obs,
  A_obs,
  A_tgt,
  A_shp_eq,
  A_shp_ineq,
  beta_shp_eq,
  beta_shp_ineq,
  kappa = 1e-05,
  norm = 2,
  solver = "gurobi",
  estimate = TRUE,
  progress = TRUE
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{A_obs}{The "observed matrix" in the inference \eqn{A_{\mathrm{obs}}}.}

\item{A_tgt}{The "target matrix" in the inference \eqn{A_{\mathrm{tgt}}}.}

\item{A_shp_eq}{Matrix representing equality shape constraints.}

\item{A_shp_ineq}{Matrix representing inequality shape constraints.}

\item{beta_shp_eq}{RHS vector in equality shape constraints.}

\item{beta_shp_ineq}{RHS vector in inequality shape constraints.}

\item{kappa}{Parameter used in the second step of the two-step procedure 
for obtaining the solution subject to the shape constraints.}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{estimate}{Boolean variable to indicate whether the estimated 
problem should be considered.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the inference procedure. If it is set as \code{TRUE}, the 
messages are displayed throughout the procedure. Otherwise, the messages
will not be displayed.}
}
\value{
Returns the bounds subject to the shape constraints.
  \item{ub}{Upper bound with shape constraints}
  \item{lb}{Lower bound with shape constraints}
  \item{est}{Indicator of whether estimation is involved in the 
  estimation}
  \item{call}{The function that has been called.}
  \item{norm}{Norm used in the optimization problem.}
}
\description{
This function computes the bound of the linear program 
   subject to shape constraints. This function also offers an option
   to estimate the shape constraints using a two-step procedure and 
   some tolerance level. \eqn{\ell^1}-norm and \eqn{\ell^2}-norm are 
   supported in the estimation procedure.
}
