% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{estbounds2.L1}
\alias{estbounds2.L1}
\title{Estimates the bounds with shape constraints (stage 2 with 1-norm)}
\usage{
estbounds2.L1(data, firststepsoln, lpmodel, modelsense, kappa, solver)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{firststepsoln}{The list of solutions to the first step problem.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{kappa}{The tuning parameter used in the second step of the two-step
procedure for obtaining the bounds subject to the shape constraints.
It can be any nonnegative number.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}
}
\value{
Returns the solution to the second step of the two-step procedure.
 \item{objval}{The optimal value.}
 \item{x}{The optimal point.}
 \item{status}{The status of the optimization problem.}
}
\description{
This function evaluates the solution to stage 2 of the
   two-step procedure to obtain the estimated bound with the 1-norm.
}
