% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.bs.fn}
\alias{fsst.beta.bs.fn}
\title{Computes one bootstrap estimates for \code{beta.obs}.}
\usage{
fsst.beta.bs.fn(x, data, lpmodel, pbar, progress, eval.count, n.bs, any.list)
}
\arguments{
\item{x}{This is either the list of indices that represent the bootstrap
replications, or the list of bootstrap components of the \code{lpmodel}
object passed from the user.}

\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{pbar}{The progress bar object.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}

\item{n.bs}{The total number of replications to be conducted in this
procedure.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{beta}{A bootstrap estimator of \eqn{\hat{\beta}_{n,b}}.}
  \item{msg}{An error message (if applicable).}
}
\description{
This function carries out one bootstrap replication for
  getting \code{beta.obs} in the \code{\link[lpinfer]{fsst}} procedure.
  This function is used in the \code{\link[lpinfer]{fsst.beta.bs}} function
  via the \code{future_map} function.
}
