% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.bs}
\alias{fsst.beta.bs}
\title{Computing the bootstrap estimates of \code{beta.obs}}
\usage{
fsst.beta.bs(
  n,
  data,
  beta.obs.hat,
  lpmodel,
  R,
  maxR,
  progress,
  df.error,
  iseq,
  seed,
  eval.count
)
}
\arguments{
\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{data}{Data used in the tests.}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel}.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{R}{Number of bootstrap replications.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{df.error}{Table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{iseq}{The list of indices or betas to iterate over.}

\item{seed}{The \code{.Random.seed} obtained in the beginning of the code.}

\item{eval.count}{Count for the number of times the \code{future_lapply}
function has been called. If this object is zero, it means that the
\code{future_lapply} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_lapply} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Returns the bootstrap estimators.
  \item{beta.obs.bs}{List of bootstrap betas.}
  \item{df.error}{Updated table showing the id of the bootstrap
      replication(s) with error(s) and the corresponding error message(s).}
  \item{R.eval}{Number of bootstrap replications that has been conducted.}
  \item{R.succ}{Number of successful bootstrap replications.}
}
\description{
This function computes the bootstrap estimates of
  \code{beta.obs}.
}
