% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.star.bs.fn}
\alias{fsst.beta.star.bs.fn}
\title{Computes one bootstrap estimates of \code{beta.star} and \code{x.star}}
\usage{
fsst.beta.star.bs.fn(
  beta.obs.bs,
  data,
  lpmodel,
  beta.tgt,
  weight.mat,
  sigma.beta.obs,
  solver,
  pbar,
  progress,
  n.bs,
  eval.count
)
}
\arguments{
\item{beta.obs.bs}{Bootstrap estimates of \code{beta.obs}.}

\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{weight.mat}{Weighting matrix.}

\item{sigma.beta.obs}{Estimator of the asymptotic variance for
\code{beta.obs}.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{pbar}{Progress bar object.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{n.bs}{Total number of replications to be conducted in this procedure.}

\item{eval.count}{Count for the number of times the \code{future_lapply}
function has been called. If this object is zero, it means that the
\code{future_lapply} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_lapply} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{beta.star}{Bootstrap estimates of \code{beta.star}.}
  \item{x.star}{Bootstrap estimates of \code{x.star}.}
  \item{msg}{Error message (if applicable).}
}
\description{
This function computes one bootstrap estimate for
  \code{beta.star} and \code{x.star}. This function is used
  in the \code{fsst.beta.star.bs} function via the \code{future_lapply}
  command.
}
