% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.beta.star.bs}
\alias{fsst.beta.star.bs}
\title{Computes the bootstrap estimates of the starred version of \code{beta.obs}}
\usage{
fsst.beta.star.bs(
  data,
  lpmodel,
  beta.n,
  beta.n.bs,
  beta.tgt,
  weight.mat,
  beta.obs.hat,
  beta.obs.bs,
  R,
  sigma.beta.obs,
  solver,
  df.error,
  p,
  d,
  progress,
  eval.count
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.n}{The sample \eqn{\widehat{\bm{\beta}}_n} vector that is defined
as \eqn{\widehat{\bm{\beta}}_n \equiv (\widehat{\bm{\beta}}_{{\rm obs},n},
\bm{\beta}_{{\rm shp},n}, \bm{\beta}_{{\rm tgt}})'}.}

\item{beta.n.bs}{The bootstrap estimates of \code{beta.n}.}

\item{beta.tgt}{The value to be tested.}

\item{weight.mat}{The weighting matrix for the \code{\link[lpinfer]{fsst}}
procedure.}

\item{beta.obs.hat}{The sample estimator
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel} (and \code{data} if applicable).}

\item{beta.obs.bs}{The bootstrap estimates of \code{beta.obs}.}

\item{R}{The number of bootstrap replications.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{df.error}{A table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{p}{The length of the beta vector.}

\item{d}{The number of columns of the \code{A} matrix.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Return the following list of objects:
  \item{beta.star}{This corresponds to \eqn{\widehat{\bm{\beta}}^\star}.}
  \item{beta.star.bs}{This corresponds to the bootstrap estimates of
    \eqn{\widehat{\bm{\beta}}^\star}.}
  \item{x.star}{This corresponds to \eqn{\bm{x}^\star}, which is the
    solution to the quadratic program \code{\link[lpinfer]{beta.star.qp}}.}
  \item{x.star.bs}{This corresponds to the bootstrap estimates of
    \eqn{\bm{x}^\star}.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{The number of new errors.}
  \item{error.id}{The indices of the bootstrap replications that give
    errors.}
}
\description{
This function computes the bootstrap estimates of
  \eqn{\widehat{\bm{\beta}}^\star_n}.
}
