% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.check}
\alias{fsst.check}
\title{Checks and updates the input in \code{fsst}}
\usage{
fsst.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  Rmulti,
  lambda,
  rho,
  n,
  weight.matrix,
  solver,
  progress,
  sqrtm.method,
  sqrtm.tol,
  previous.output
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{R}{The number of bootstrap replications.}

\item{Rmulti}{The multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{lambda}{Parameter used to obtain the restricted estimator
\eqn{\widehat{\bm{\beta}}^r_n}. A data-driven parameter \code{lambda} can
be included if \code{NA} is included as part of the vector for
\code{lambda}. For instance, if \code{lambda} is set as \code{c(0.1, NA)},
then both 0.1 and the data-driven \code{lambda} will be applied in the
\code{\link[lpinfer]{fsst}} test. The default is to use the
data-driven \code{lambda}.}

\item{rho}{Parameter used in the studentization of matrices.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{weight.matrix}{The option used in the weighting matrix. There are
three options available:
\itemize{
  \item{\code{identity} --- identity matrix}
  \item{\code{diag} --- the diagonal matrix that takes the diagonal
     elements of the inverse of the variance matrix}
  \item{\code{avar} --- inverse of the variance matrix}
}}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{sqrtm.method}{The method used to obtain the matrix square root in
the \code{\link[lpinfer]{fsst}} procedure. This has to be a function that
takes one argument that accepts a square matrix of size k x k and returns
a square matrix of size k x k, where k can be the length of the
\eqn{\beta(P)} vector, or the \code{beta.obs} component of the
\code{lpinfer} object.}

\item{sqrtm.tol}{The absolute tolerance used to check whether the matrix
square root is correct. This is done by checking whether the Frobenius
norm is smaller than the tolerance level, i.e., when \eqn{A} is the
give matrix, \eqn{B} is the matrix square root obtained from the
given \code{sqrtm.method} function, and \eqn{\epsilon} is the
tolerance level, the FSST test checks whether
\eqn{||A - BB||_F < \epsilon}. If this does not hold, the FSST test will
use the \code{\link[expm]{sqrtm}} function from the \code{expm} package
to obtain the matrix square root.}

\item{previous.output}{The list of outputs obtained from the previous
evaluation of the procedure. This is only used in the
\code{\link[lpinfer]{fsst}} test and when \code{previous.out} contains
the studentized matrix \code{omega.i}. The rest of the arguments
are ignored.}
}
\value{
Returns the updated parameters back to the function \code{fsst}.
  The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{test.logical}}
      \item{\code{logical.lb}}
      \item{\code{logical.ub}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{fsst}} function. If there is any invalid input,
   the function will be terminated and error messages will be printed.
}
