% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.check}
\alias{fsst.check}
\title{Checks and updates the input}
\usage{
fsst.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  lambda,
  rho,
  n,
  weight.matrix,
  solver,
  cores,
  progress
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstraps chosen by the users.}

\item{lambda}{Parameter used in the cone bootstrap sattistics}

\item{rho}{Parameter used in the studentization of matrices}

\item{weight.matrix}{The option used in the weighting matrix.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop.
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
Returns the list of updated parameters as follows:
  \item{data}{Upated data in class \code{data.frame}}
  \item{lpmodel}{A list of linear programming objects in this
     `\code{lpinfer}` package.}
  \item{solver}{Solver to be used.}
  \item{solver.name}{Updated name of solver in lower case.}
  \item{cores}{Updated number of cores to be used in the parallelization
     of the for-loop in the bootstrap procedure.}
}
\description{
This function checks and updates the input of the user. If
   there is any invalid input, this function will be terminated and
   generates appropriate error messages.
}
