% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.check}
\alias{fsst.check}
\title{Checks and updates the input in \code{fsst}}
\usage{
fsst.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  Rmulti,
  lambda,
  rho,
  n,
  weight.matrix,
  solver,
  progress
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstrap replications.}

\item{Rmulti}{Multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{lambda}{Parameter used to obtain the restricted estimator
\eqn{\widehat{\bm{\beta}}^r_n}. A data-driven parameter \code{lambda} can
be included if \code{NA} is included as part of the vector for
\code{lambda}. For instance, if \code{lambda} is set as \code{c(0.1, NA)},
then both 0.1 and the data-driven \code{lambda} will be applied in the
FSST procedure. The default is to use the data-driven approach for
\code{lambda}.}

\item{rho}{Parameter used in the studentization of matrices.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{weight.matrix}{The option used in the weighting matrix. There are
three options available:
\itemize{
  \item{\code{identity} --- identity matrix}
  \item{\code{diag} --- the diagonal matrix that takes the diagonal
     elements of the inverse of the variance matrix}
  \item{\code{avar} --- inverse of the variance matrix}
}}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the updated parameters back to the function \code{fsst}.
  The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{seed}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{test.logical}}
   }
}
\description{
This function checks and updates the input of the user. If
   there is any invalid input, this function will be terminated and
   generates appropriate error messages.
}
