% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.cone.lp}
\alias{fsst.cone.lp}
\title{Computes the solution to the cone problem}
\usage{
fsst.cone.lp(n, omega.i, beta.n, beta.star, lpmodel, indicator, solver)
}
\arguments{
\item{omega.i}{The matrix \eqn{\widehat{\bm{\Omega}}^i_n}, i.e. the 
regularized matrix for 
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{lpmodel}{A list of objects that are used in inference of linear 
programming problems. The list of objects required in the \code{dkqs} 
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{indicator}{A binary variable that equals to 1 for \eqn{d \geq p} and 
equals to 0 for \eqn{d < p}.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the cone problem.
}
