% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvtable.R
\name{fsst.cv.table}
\alias{fsst.cv.table}
\title{Wrapper for the '\code{fsst}' procedure}
\usage{
fsst.cv.table(
  param,
  param.name,
  cone.n.list,
  range.n.list,
  cone.bs.list,
  range.bs.list,
  levels = c(0.99, 0.95, 0.9),
  digits = 5
)
}
\arguments{
\item{param}{Tuning parameter of the testing procedure.}

\item{param.name}{Name of the tuning parameter.}

\item{cone.n.list}{A list of sample cone test statistics.}

\item{range.n.list}{A list of sample range test statistics.}

\item{cone.bs.list}{A vector or a list of bootstrap cone test statistics.}

\item{range.bs.list}{A vector or a list of bootstrap range test statistics.}

\item{levels}{Confidence levels.}

\item{digits}{Number of digits in the test statistics.}
}
\value{
Returns a table of critical values.
  \code{cv.table}{Table of critical values.}
}
\description{
This is a wrapper of the `\code{fsst}` procedure in order to
  produce a table that consists of the critical values for the test
  statistics, cone component and the range component.
}
