% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.lambda}
\alias{fsst.lambda}
\title{Data-driven choice of \code{lambda} in the \code{\link[lpinfer]{fsst}}
procedure}
\usage{
fsst.lambda(
  n,
  omega.i,
  beta.n,
  beta.star,
  lpmodel,
  R.succ,
  beta.star.list,
  solver,
  progress,
  df.error,
  p,
  d,
  eval.count
)
}
\arguments{
\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{omega.i}{The matrix \eqn{\widehat{\bm{\Omega}}^i_n}, i.e. the
regularized matrix for
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{beta.n}{The sample \eqn{\widehat{\bm{\beta}}_n} vector that is defined
as \eqn{\widehat{\bm{\beta}}_n \equiv (\widehat{\bm{\beta}}_{{\rm obs},n},
\bm{\beta}_{{\rm shp},n}, \bm{\beta}_{{\rm tgt}})'}.}

\item{beta.star}{The starred version of the \code{beta.n} vector, i.e.
\eqn{\widehat{\bm{\beta}}^\star_n}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{R.succ}{The number of successful bootstrap replications.}

\item{beta.star.list}{This corresponds to the bootstrap estimates of
\code{beta.star}.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{df.error}{A table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{p}{The length of the vector \eqn{\bm{\beta}}.}

\item{d}{The number of columns of the \eqn{\bm{A}} matrix.}

\item{eval.count}{The count for the number of times the \code{future_lapply}
function has been called. If this object is zero, it means that the
\code{future_lapply} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_lapply} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Returns the data-driven \code{lambda} and the error messages.
  \item{lambda}{The data-driven \code{lambda}}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{The number of new errors.}
  \item{error.id}{The indices of the bootstrap replications that give
    errors.}
  \item{R.succ}{The number of successful bootstrap replications.}
}
\description{
This function provides a data-driven choice of \code{lambda}
  in the \code{\link[lpinfer]{fsst}} procedure.
}
\details{
\eqn{\alpha_n} is set as \eqn{1} if the number of observations in
  the data set is less than 16.
}
