% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.pval}
\alias{fsst.pval}
\title{Calculates the \eqn{p}-value for the \code{\link[lpinfer]{fsst}} procedure}
\usage{
fsst.pval(range.n, cone.n, range.n.list, cone.n.list, R, alpha = 0.05)
}
\arguments{
\item{range.n}{The range component of the sample test statistic.}

\item{cone.n}{The cone component of the sample test statistic.}

\item{range.n.list}{The bootstrap estimates of the range component in the
test statistics.}

\item{cone.n.list}{The bootstrap estimates of the cone component in the
test statistics.}

\item{R}{The number of bootstrap replications.}

\item{alpha}{The significance level.}
}
\value{
Returns the \eqn{p}-value and the decision.
  \item{pval}{The \eqn{p}-value.}
  \item{decision}{The decision to reject or not.}
}
\description{
This function computes the \eqn{p}-value of the test based on
   the bootstrap estimates for the \code{\link[lpinfer]{fsst}} procedure.
}
