% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.pval}
\alias{fsst.pval}
\title{Auxiliary function to calculate the p-value of `\code{fsst}`}
\usage{
fsst.pval(range.n, cone.n, range.n.list, cone.n.list, R, alpha = 0.05)
}
\arguments{
\item{range.n}{The point estimate for the `Range` test statistics.}

\item{cone.n}{The point estimate for the `Cone` test statistics.}

\item{range.n.list}{The bootstrap test estimates for the `Range` test 
statistics.}

\item{cone.n.list}{The bootstrap test estimates for the `Cone` test 
statistics.}

\item{R}{Number of bootstraps chosen by the users.}
}
\value{
Returns the \eqn{p}-value and the decision.
  \item{pval}{\eqn{p}-value.}
  \item{decision}{Decision to reject or not.}
}
\description{
This function computes the \eqn{p}-value of the test based on
   the bootstrap estimates.
}
