% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs.fn}
\alias{fsst.range.bs.fn}
\title{Computes one bootstrap estimate for the range component for the test
statistic}
\usage{
fsst.range.bs.fn(beta.x.star, n, lpmodel, beta.obs.hat, x.star, weight.mat)
}
\arguments{
\item{beta.x.star}{A list that combines the bootstrap estimates of
\code{beta.star} and \code{x.star}.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel}.}

\item{x.star}{Optimal value from \code{beta.star.qp}.}

\item{weight.mat}{Weighting matrix.}
}
\value{
Return the following list of objects:
  \item{Ts}{Range test statistic.}
  \item{msg}{Error message (if applicable).}
}
\description{
This function computes the one bootstrap estimate for the
  range component in the test statistic.
}
