% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs.fn}
\alias{fsst.range.bs.fn}
\title{Computes one bootstrap estimate for the range component of the test
statistic}
\usage{
fsst.range.bs.fn(
  beta.x.star,
  n,
  lpmodel,
  beta.obs.hat,
  x.star,
  weight.mat.root,
  pbar,
  progress,
  n.bs,
  eval.count
)
}
\arguments{
\item{beta.x.star}{A list that combines the bootstrap estimates of
\code{beta.star} and \code{x.star}.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.obs.hat}{The sample estimator
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel} (and \code{data} if applicable).}

\item{x.star}{The optimal value from \code{beta.star.qp}.}

\item{weight.mat.root}{The matrix square root of \code{weight.mat}.}

\item{pbar}{The progress bar object.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{n.bs}{The total number of replications to be conducted in this
procedure.}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Return the following list of objects:
  \item{Ts}{The bootstrap range test statistic.}
  \item{msg}{An error message (if applicable).}
}
\description{
This function computes the one bootstrap estimate for the
  range component of the test statistic.
}
