% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs}
\alias{fsst.range.bs}
\title{Computing bootstrap components of `Range`}
\usage{
fsst.range.bs(
  n,
  omega.e,
  beta.n,
  beta.star,
  R,
  beta.n.bs,
  beta.star.bs,
  solver,
  cores,
  progress
)
}
\arguments{
\item{omega.e}{The matrix \eqn{\widehat{\bm{\Omega}}^e_n}, i.e. the 
regularized matrix for 
\eqn{\widehat{\bm{\Sigma}}^{\beta^\star}_{n,\bar{\rho}}}.}

\item{R}{Number of bootstraps chosen by the users.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop. 
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{progress}{The boolean variable for whether the result messages 
should be displayed in the inference procedure. If it is set as 
\code{TRUE}, the messages are displayed throughout the procedure. 
Otherwise, the messages will not be displayed.}
}
\value{
A list of bootstrap Range statistics.
  \item{range.n.list}{A list of bootstrap Range statistics 
    \eqn{\{\mathrm{Range}_{n,b}\}^B_{b=1}}}
}
\description{
This function computes the bootstrap components of the 
  `Range` test statistics.
}
