% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs}
\alias{fsst.range.bs}
\title{Bootstrap procedure of computing the range component}
\usage{
fsst.range.bs(
  n,
  lpmodel,
  beta.obs.hat,
  beta.obs.bs,
  x.star,
  x.star.bs,
  weight.mat,
  R,
  progress,
  df.error
)
}
\arguments{
\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the FSST test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }}

\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel}.}

\item{beta.obs.bs}{Bootstrap estimates of \code{beta.obs}.}

\item{x.star}{Optimal value from \code{beta.star.qp}.}

\item{x.star.bs}{Bootstrap estimates of \code{x.star}.}

\item{weight.mat}{Weighting matrix.}

\item{R}{Number of bootstrap replications.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{df.error}{Table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{p}{Length of the vector \eqn{\bm{\beta}}.}

\item{d}{Number of columns of the \eqn{\bm{A}} matrix.}
}
\value{
Return the following list of objects:
  \item{range.n.list}{List of bootstrap range components.}
  \item{df.error}{Table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{Number of new errors.}
  \item{error.id}{Indices of the bootstrap replications that give errors.}
}
\description{
This function computes the bootstrap estimates of the range
  components.
}
