% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range.bs}
\alias{fsst.range.bs}
\title{Bootstrap procedure of computing the range component}
\usage{
fsst.range.bs(
  n,
  lpmodel,
  beta.obs.hat,
  beta.obs.bs,
  x.star,
  x.star.bs,
  weight.mat,
  R,
  progress,
  df.error,
  eval.count
)
}
\arguments{
\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{beta.obs.hat}{The sample estimator
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel} (and \code{data} if applicable).}

\item{beta.obs.bs}{The bootstrap estimates of \code{beta.obs}.}

\item{x.star}{The optimal value from \code{beta.star.qp}.}

\item{x.star.bs}{The bootstrap estimates of \code{x.star}.}

\item{weight.mat}{The weighting matrix for the \code{\link[lpinfer]{fsst}}
procedure.}

\item{R}{The number of bootstrap replications.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{df.error}{A table showing the id of the bootstrap replication(s)
with error(s) and the corresponding error message(s).}

\item{eval.count}{The count for the number of times the \code{future_map}
function has been called. If this object is zero, it means that the
\code{future_map} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_map} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}
}
\value{
Return the following list of objects:
  \item{range.n.list}{The list of bootstrap range components.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{new.error}{The number of new errors.}
  \item{error.id}{The indices of the bootstrap replications that give
    errors.}
}
\description{
This function computes the bootstrap estimates of the range
  components.
}
