% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.range}
\alias{fsst.range}
\title{Computes the statistic to the range problem in the FSST procedure}
\usage{
fsst.range(n, beta.obs.hat, x.star, lpmodel, weight.mat)
}
\arguments{
\item{beta.obs.hat}{Estimator of
\eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}} based on the given
information in \code{lpmodel}.}

\item{x.star}{Optimal value from `\code{beta.star.qp}`.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{weight.mat}{Weighting matrix}
}
\value{
Returns the optimal point and optimal value.
 \item{range}{The optimal value of the Range component.}
}
\description{
This function computes the statistic of the range problem in
   the FSST procedure.
}
