% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{fsst.weight.matrix}
\alias{fsst.weight.matrix}
\title{Computes the weighting matrix in the \code{\link[lpinfer]{fsst}} procedure}
\usage{
fsst.weight.matrix(weight.matrix, beta.obs.hat, beta.sigma)
}
\arguments{
\item{weight.matrix}{The option used in the weighting matrix. There are
three options available:
\itemize{
  \item{\code{identity} --- identity matrix}
  \item{\code{diag} --- the diagonal matrix that takes the diagonal
     elements of the inverse of the variance matrix}
  \item{\code{avar} --- inverse of the variance matrix}
}}

\item{beta.obs.hat}{The sample estimator
\eqn{\widehat{\bm{\beta}}_{\mathrm{obs}, n}} based on the given
information in \code{lpmodel} (and \code{data} if applicable).}

\item{beta.sigma}{The variance estimator of \eqn{\bm{\beta}_{\rm obs}}.}
}
\value{
The weighting matrix is returned.
   \item{weight.max}{The weighting matrix.}
}
\description{
This function returns the weighting matrix in the
  \code{\link[lpinfer]{fsst}} procedure.
}
