% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{full.beta.bs}
\alias{full.beta.bs}
\title{Construct the full beta vector}
\usage{
full.beta.bs(lpmodel, beta.tgt, beta.obs.bs, R)
}
\arguments{
\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{beta.obs.bs}{List of bootstrap components of \code{beta.obs}.}

\item{R}{Number of bootstraps chosen by the users.}
}
\value{
List of full beta vectors
  \item{beta.bs}{List of full beta vector
  \eqn{\{\widehat{\bm{\beta}_{n,b}}}\}^B_{b=1}}
}
\description{
This function concatenate the \code{beta.obs}, \code{beta.shp}
  and \code{beta.tgt} components to form the full \code{beta} vector.
}
