% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R, R/dkqs_objfunction_separate_obj210_Ab.R
\name{gurobi_optim}
\alias{gurobi_optim}
\title{LP and QP solver by Gurobi}
\usage{
gurobi_optim(
  obj2,
  obj1,
  obj0,
  A,
  rhs,
  sense,
  modelsense,
  lb,
  A_obs,
  beta_hat_obs,
  n
)

gurobi_optim(
  obj2,
  obj1,
  obj0,
  A,
  rhs,
  sense,
  modelsense,
  lb,
  A_obs,
  beta_hat_obs,
  n
)
}
\arguments{
\item{obj2}{The matrix that corresponds to the second-order term in the
quadratic program. This input will be set to \code{NULL} if the problem
considered is a linear program.}

\item{obj1}{The vector that corresponds to the first-order term in the
quadratic or linear program.}

\item{obj0}{The constant term in the quadratic or linear program.}

\item{A}{The constraint matrix.}

\item{rhs}{The RHS vector for the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{lb}{The lower lound vector.}

\item{n}{The number of observations in the dataframe.}

\item{Af}{The matrix that is involved in the objective function.}

\item{bf}{The vector that is involved in the objective function.}

\item{nf}{The number of observations in the dataframe.}
}
\value{
Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}

Returns the optimal point and optimal value.
 \item{x}{Optimal point calculated from the optimizer.}
 \item{objval}{Optimal value calculated from the optimizer.}
}
\description{
This function computes the solution to the quadratic or linear
   program using the `\code{Gurobi}' package.

This function computes the solution to the quadratic or linear
   program using the `\code{Gurobi}' package.
}
