% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{infeasible.msg.betatgt}
\alias{infeasible.msg.betatgt}
\title{General message for infeasible \code{beta.tgt}}
\usage{
infeasible.msg.betatgt()
}
\value{
Returns a string of the message.
  \item{msg.explain}{The message to explain why the \eqn{p}-value is zero.}
  \item{msg.pval}{The message indicating that the \eqn{p}-value is zero.}
}
\description{
This function prints the message to inform the user that
  the \eqn{p}-value is directly set to 0 because they have specified a
  \code{beta.tgt} parameter that is infeasible, i.e. outside the logical
  bounds of the program.
}
