% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{infeasible.pval.msg}
\alias{infeasible.pval.msg}
\title{Wrapper for \code{infeasible.msg.betatgt}}
\usage{
infeasible.pval.msg()
}
\value{
Nothing is returned.
}
\description{
This function is a wrapper for the
  \code{infeasible.msg.betatgt} function to print that \eqn{p}-value is
  zero in the \code{print} or \code{summary} messages.
}
