% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{invertci.check}
\alias{invertci.check}
\title{Checks and updates the input of the function \code{\link[lpinfer]{invertci}}}
\usage{
invertci.check(
  f,
  farg,
  alpha,
  init.lb,
  init.ub,
  tol,
  max.iter,
  pvals,
  progress
)
}
\arguments{
\item{f}{The function that represents a testing procedure.}

\item{farg}{The list of arguments to be passed to the function of testing
procedure.}

\item{alpha}{The significance level(s). This can be a vector.}

\item{init.lb}{The initial brackets to search for the lower bound. This is
not required if the \code{\link[lpinfer]{chorussell}} is used.}

\item{init.ub}{The initial brackets to search for the upper bound. This is
not required if the \code{\link[lpinfer]{chorussell}} is used.}

\item{tol}{The tolerance level in the bisection method.}

\item{max.iter}{The maximum number of iterations in the bisection method.}

\item{pvals}{The data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the procedure of constructing confidence interval. If
it is set as \code{TRUE}, the messages are displayed throughout the
procedure. Otherwise, the messages will not be displayed.}
}
\value{
Returns the updated parameters back to the function
\code{subsample}. The following information are updated:
   \itemize{
      \item{\code{pvals}}
      \item{\code{lb0}}
      \item{\code{lb1}}
      \item{\code{ub0}}
      \item{\code{ub1}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{invertci}} function. If there is any invalid input,
   the function will be terminated and error messages will be printed.
}
