% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{invertci}
\alias{invertci}
\title{Constructs confidence interval}
\usage{
invertci(
  f,
  farg,
  alpha = 0.05,
  lb0 = NULL,
  lb1 = NULL,
  ub0 = NULL,
  ub1 = NULL,
  tol = 1e-04,
  max.iter = 20,
  df_ci = NULL,
  progress = FALSE
)
}
\arguments{
\item{f}{Function that represents a testing procedure.}

\item{farg}{List of arguments to be passed to the function of testing 
procedure.}

\item{alpha}{Significance level of the test.}

\item{lb0}{Logical lower bound for the confidence interval.}

\item{lb1}{Maximum possible lower bound for the confidence interval.}

\item{ub0}{Logical upper bound for the confidence interval.}

\item{ub1}{Minimum possible upper bound for the confidence interval.}

\item{tol}{Tolerance level in the bisection method.}

\item{max.iter}{Maximum number of iterations in the bisection method.}

\item{df_ci}{Data frame that consists the points and the corresponding
\eqn{p}-values that have been tested in the previous iterations.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the procedure of constructing confidence interval. If 
it is set as \code{TRUE}, the messages are displayed throughout the 
procedure. Otherwise, the messages will not be displayed.}
}
\value{
Returns the confidence interval and a data frame that contains the 
   points being tested in the procedure.
   \item{up}{Upper bound of the confidence interval.}
   \item{low}{Lower bound of the confidence interval.}
   \item{df_ci}{data frame that consists of the points and the corresponding
      \eqn{p}-values that have been tested in constructing the confidence 
      intervals.}
   \item{df_ub}{Data frame storing the information for the bisection 
      method in each iteration when evaluating the upper bound of the 
      confidence interval.}
   \item{df_lb}{Data frame storing the information for the bisection 
      method in each iteration when evaluating the lower bound of the 
      confidence interval.}
   \item{tol}{Tolerance level in the bisection method.}
   \item{iter}{Total number of iterations taken.}
   \item{call}{The function that has been called.}
}
\description{
This function constructs the confidence interval using the
   bisection method.
}
\details{
The number of decimal places displayed in the messages (if 
   \code{progress} is set as \code{TRUE}) is equal to the number of decimal
   places in the variable \code{tol}.
}
