% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{limsolve.optim}
\alias{limsolve.optim}
\title{LP and QP solver by \code{limSolve}}
\usage{
limsolve.optim(Af, bf, nf, A, rhs, sense, modelsense, lb, weight = NULL, ...)
}
\arguments{
\item{Af}{The matrix that is involved in the objective function.}

\item{bf}{The vector that is involved in the objective function.}

\item{nf}{The number of observations in the data frame.}

\item{A}{The constraint matrix.}

\item{rhs}{The rhs vector for the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{lb}{The lower bound vector.}

\item{weight}{Weighting matrix.}

\item{...}{List of options to be passed to the Gurobi solver. This part is
optional.}
}
\value{
Returns the optimal point and optimal value.
 \item{objval}{Optimal objective value.}
 \item{x}{Optimal point.}
 \item{status}{Status of the optimization problem.}
}
\description{
This function computes the solution to linear and quadratic
   programs using the \code{limSolve} package.
}
