% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{limsolve_optim}
\alias{limsolve_optim}
\title{limSolve function for linear and quadratic programs}
\usage{
limsolve_optim(
  obj2,
  obj1,
  obj0,
  A,
  rhs,
  sense,
  modelsense,
  lb,
  A_obs,
  beta_hat_obs,
  n
)
}
\arguments{
\item{obj2}{The matrix that corresponds to the second-order term in the
quadratic program. This input will be set to \code{NULL} if the problem
considered is a linear program.}

\item{obj1}{The vector that corresponds to the first-order term in the
quadratic or linear program.}

\item{obj0}{The constant term in the quadratic or linear program.}

\item{A}{The constraint matrix.}

\item{rhs}{The RHS vector for the linear constraints.}

\item{sense}{The sense of the linear constraints.}

\item{modelsense}{The indicator of whether the model is to max or min an
objective function.}

\item{lb}{The lower lound vector.}

\item{n}{The number of observations in the dataframe.}
}
\value{
Returns the optimal objective value and the corresponding argument
  to the quadratic or linear program.
}
\description{
This function computes the solution to linear and quadratic 
   programs using the `\code{limSolve}' package.
}
