% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{lpm.print}
\alias{lpm.print}
\title{Print the `\code{lpmodel}` or `\code{lpmodel.natural}` object}
\usage{
lpm.print(x, lpm.string, data = NULL, ...)
}
\arguments{
\item{x}{The `\code{lpmodel}` object or `\code{lpmodel.natural}` object.}
}
\value{
Print the summary of the objects in `\code{lpmodel}` or 
   `\code{lpmodel.natural}`.
}
\description{
This function prints objects that are contained in the
   list of `\code{lpmodel}` or `\code{lpmodel.natural}`.
}
