% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{mincriterion.check}
\alias{mincriterion.check}
\title{Checks and updates the input in \code{mincriterion}}
\usage{
mincriterion.check(data, lpmodel, norm, solver)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}
}
\value{
Returns the updated parameters back to the function
\code{mincriterion}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
   }
}
\description{
This function checks and updates the input from the user in the
   \code{\link[lpinfer]{mincriterion}} function. If there is any invalid
   input, the function will be terminated and error messages will be printed.
}
