% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{mincriterion}
\alias{mincriterion}
\title{First-stage estimation procedure for \code{\link[lpinfer]{estbounds}}}
\usage{
mincriterion(data = NULL, lpmodel, norm = 2, solver = NULL)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{The \code{lpmodel} object.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}
}
\value{
Returns the solution to the first step of the two-step procedure
   and argument for the linear program.
 \item{objval}{The optimal value.}
 \item{x}{The optimal point.}
 \item{larg}{The arguments for the estimation program.}
 \item{norm}{The norm used in the estimation problem.}
 \item{solver}{The solver used in the estimation problem}
 \item{call}{The details of the function that has been called.}
}
\description{
This function evaluates the solution to stage 1 of the
   two-step procedure to obtain the estimated bounds in the
   \code{\link[lpinfer]{estbounds}} procedure. This function can
   be used to evaluate both the estimation problem with the 1-norm or
   the 2-norm.
}
\section{Example}{

\preformatted{
  source("./example/dgp_missingdata.R") # Change directory if necessary
  J <- 5
  N <- 1000
  data <- missingdata_draw(J = J, n = N, seed = 1, prob.obs = .5)
  lpm <- missingdata_lpm(J = J, info = "full", data = data)
  mincriterion(data = data,
               lpmodel = lpm,
               norm = 2,
               solver = "gurobi")
}
}

\section{More examples}{

  More examples can be found in the \code{mincriterion_example.R} file
  under the \code{example} subdirectory of the installation directory for
  the \code{lpinfer} package.
}

