% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{mincriterion}
\alias{mincriterion}
\title{First-stage of the estimation procedure for \code{estbounds}}
\usage{
mincriterion(data, lpmodel, norm, solver)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{A.shp}}
  \item{\code{beta.obs}}
  \item{\code{beta.shp}}
}}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}
}
\value{
Returns the solution to the first step of the two-step procedure
   and argument for the linear program.
 \item{objval}{Optimal value calculated from the optimizer.}
 \item{x}{Optimal point calculated from the optimizer.}
 \item{larg}{Arguments for the estimation program.}
 \item{norm}{Norm used in the estimation problem.}
 \item{solver}{The solver used in the estimation problem}
 \item{call}{The details of the function that has been called.}
}
\description{
This function evaluates the solution to stage 1 of the
   two-step procedure obtaining the estimated bound. This function can
   be used to evaluate both the estimation problem with the 1-norm or
   the 2-norm.
}
