% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{post.bs}
\alias{post.bs}
\title{Auxiliary function for the post-bootstrap procedure}
\usage{
post.bs(
  test.return,
  i0,
  i1,
  R.eval,
  T.list = NULL,
  beta.list = NULL,
  error.list = NULL,
  list.param = NULL,
  error.param = NULL
)
}
\arguments{
\item{test.return}{List of information returned from the test function.}

\item{i0}{Starting index to be evaluated.}

\item{i1}{Last index to be evaluated.}

\item{R.eval}{Number of bootstrap replications that has been evaluated.}

\item{T.list}{List of bootstrap test statistics.}

\item{beta.list}{List of the bootstrap replications of the \code{beta.obs}
component.}

\item{error.list}{List of error messages.}

\item{error.param}{List of parameters that lead to errors (if applicable).}

\item{param.list}{List of parameters (if applicable).}
}
\value{
Returns the following three objects.
  \item{R.succ}{Number of successful bootstrap replications.}
  \item{R.eval}{Number of bootstrap replications that has been evaluated.}
  \item{T.list}{List of test statistics.}
  \item{beta.list}{List of the bootstrap replications of the
    \code{beta.obs} component.}
  \item{error.list}{List of error messages (if applicable).}
  \item{error.param}{List of parameters that lead to errors (if
    applicable).}
}
\description{
This function is used after the bootstrap replications to
  extract the relevant and updated information.
}
