% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorussell.R
\name{print.chorussell}
\alias{print.chorussell}
\title{Print results from \code{chorussell}}
\usage{
\method{print}{chorussell}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{chorussell}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function either prints the \eqn{p}-values or a
  \eqn{(1-\alpha)} confidence interval from \code{chorussell}.
}
