% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{print.dkqs}
\alias{print.dkqs}
\title{Print results from \code{\link[lpinfer]{dkqs}}}
\usage{
print.dkqs(x, ...)
}
\arguments{
\item{x}{An output object returned from \code{\link[lpinfer]{dkqs}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints the \eqn{p}-values from
  \code{\link[lpinfer]{dkqs}}.
}
