% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{print.estbounds}
\alias{print.estbounds}
\title{Print results from \code{estbounds}}
\usage{
\method{print}{estbounds}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{estbounds}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned. This function prints results from
   \code{estbounds}.
}
\description{
This function uses the print method on the return list of the
   function \code{estbounds}.
}
