% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{print.fsst}
\alias{print.fsst}
\title{Print results from \code{\link[lpinfer]{fsst}}}
\usage{
\method{print}{fsst}(x, ...)
}
\arguments{
\item{x}{The output objects returned from \code{\link[lpinfer]{fsst}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints the \eqn{p}-values from
  \code{\link[lpinfer]{fsst}}.
}
