% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{print.invertci}
\alias{print.invertci}
\title{Print results from \code{\link[lpinfer]{invertci}}}
\usage{
print.invertci(x, ...)
}
\arguments{
\item{x}{The output object returned from \code{\link[lpinfer]{invertci}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned
}
\description{
This function prints the results from
  \code{\link[lpinfer]{invertci}}.
}
