% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{print.invertci_multiple}
\alias{print.invertci_multiple}
\title{Print results from \code{invertci} with vector-valued alpha}
\usage{
\method{print}{invertci_multiple}(x, alphas = NULL, ...)
}
\arguments{
\item{x}{Object returned from \code{invertci}.}

\item{...}{Additional arguments.}
}
\value{
Print the basic set of results from \code{invertci}.
}
\description{
This function uses the print method on the return list of the
   function \code{invertci}.
}
