% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{print.invertci_single}
\alias{print.invertci_single}
\title{Print results from \code{\link[lpinfer]{invertci}} with a single
significance level}
\usage{
\method{print}{invertci_single}(x, ...)
}
\arguments{
\item{x}{The output object returned from \code{\link[lpinfer]{invertci}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned
}
\description{
This function prints the results from
  \code{\link[lpinfer]{invertci}}.
}
