% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{print.lpmodel}
\alias{print.lpmodel}
\title{Print the \code{lpmodel} object}
\usage{
print.lpmodel(x, data = NULL, ...)
}
\arguments{
\item{x}{The \code{lpmodel} object.}
}
\value{
Print the summary of the objects in \code{lpmodel}.
}
\description{
This function prints objects that are contained in the 
   list of \code{lpmodel}.
}
