% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{print.lpmodel}
\alias{print.lpmodel}
\title{Print the \code{lpmodel} object}
\usage{
print.lpmodel(x, data = NULL, ...)
}
\arguments{
\item{x}{An \code{lpmodel} object.}

\item{data}{An \code{data frame} or a \code{matrix}.}
}
\value{
Nothing is returned.
}
\description{
This function is a wrapper of the
  \code{\link[lpinfer]{lpm.print}} function and prints the details of the
  components in the \code{lpmodel} object.
}
