% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{print.mincriterion}
\alias{print.mincriterion}
\title{Print results from \code{\link[lpinfer]{mincriterion}}}
\usage{
\method{print}{mincriterion}(x, ...)
}
\arguments{
\item{x}{The output objects returned from
\code{\link[lpinfer]{mincriterion}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{\link[lpinfer]{mincriterion}}.
}
