% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{print.subsample}
\alias{print.subsample}
\title{Print results from \code{subsample}}
\usage{
print.subsample(x, ...)
}
\arguments{
\item{x}{Object returned from \code{subsample}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function uses the print method on the return list of the
   function \code{subsample}.
}
