% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{prog_cone}
\alias{prog_cone}
\title{Formulating and solving quadratic programs}
\usage{
prog_cone(A_obs, A_tgt, beta_obs_hat, beta_tgt, tau, problem, n)
}
\arguments{
\item{A_obs}{The observed matrix in the linear program.}

\item{A_tgt}{The "target matrix" in the linear program.}

\item{beta_tgt}{The value of \eqn{t} in the null hypothesis.}

\item{tau}{The RHS vector fo the linear constraints.}

\item{problem}{The sense of the linear constraints.}
}
\value{
Returns the solution to the quadratic program.
}
\description{
This function formulates and solves the quadratic programs (5)
   and (6) in Torgovitsky (2019).
}
