% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{pval}
\alias{pval}
\title{Auxiliary function to calculate the p-value}
\usage{
pval(T.bs, T.n, alpha = 0.05)
}
\arguments{
\item{T.n}{The test statistics obtained from quadratic program (5).}

\item{alpha}{The significance level.}

\item{T_bs}{The test statistics obtained from bootstrap.}
}
\value{
Returns the \eqn{p}-value:
  \item{p}{\eqn{p}-value.}
  \item{decision}{Decision to reject or not.}
}
\description{
This function computes the \eqn{p}-value of the test based on
   the bootstrap estimates.
}
