% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{sigma.est}
\alias{sigma.est}
\title{Computing bootstrap and variance estimators}
\usage{
\method{sigma}{est}(n, data, beta.obs.hat, lpmodel, R, progress)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{beta.obs.hat}{Estimator of \eqn{\widehat{\bm{\beta}_{\mathrm{obs}, n}}}
based on the given information in \code{lpmodel}.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{R}{Number of bootstraps chosen by the users.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
Returns the bootstrap estimators and the estimator of the
  asymptotic variance.
    \item{sigma.mat}{Estimator of the asympototic variance}
    \item{beta.obs.bs}{List of bootstrap betas}
}
\description{
Based on the function \code{beta.obs} from the object
  \code{lpmodel}, this function computes the bootstrap estimates and
  the asympototic variance estimator of the boostrap estimator. The
  asymptotic variance estimator is computed as
    \deqn{\frac{n}{B} \sum^B_{i=1} \left(\widehat{\bm{\beta}}_b -
    \widehat{\bm{\beta}}\right)  \left(\widehat{\bm{\beta}}_b -
    \widehat{\bm{\beta}}\right)'}
  for some suitable estimators \eqn{\widehat{\bm{\beta}}} and the
  corresponding bootstrap estimators
  \eqn{\{\widehat{\bm{\beta}}_b\}^B_{b=1}}.
}
