% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{sigma.summation}
\alias{sigma.summation}
\title{Computing the variance estimator}
\usage{
\method{sigma}{summation}(n, beta.bs.list)
}
\arguments{
\item{beta.bs.list}{List of bootstrap estimators
\eqn{\{\widehat{\bm{\beta}}_b\}^B_{b=1}}.}
}
\value{
Returns the bootstrap estimators and the estimator of the
  asymptotic variance.
    \item{sigma.mat}{Estimator of the asympototic variance}
}
\description{
Based on the bootstrap estimators
  \eqn{\{\widehat{\bm{\beta}}_b\}^B_{b=1}}, this function computes the
  asympototic variance estimator of the boostrap estimator, i.e.
    \deqn{\frac{n}{B} \sum^B_{i=1} \left(\widehat{\bm{\beta}}_b -
    \widehat{\bm{\beta}}\right)  \left(\widehat{\bm{\beta}}_b -
    \widehat{\bm{\beta}}\right)'.}
}
