% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsf.R
\name{standard.form}
\alias{standard.form}
\title{Obtain standard form of linear program}
\usage{
standard.form(A, b, sense, lb = NULL, obj = NULL)
}
\arguments{
\item{A}{Matrix that represents the constraints}

\item{b}{RHS vector that represents the constraints}

\item{sense}{Sense of the constraints}

\item{lb}{Lower bound of the vectors}

\item{obj}{Objective function}
}
\value{
Returns the bounds subject to the shape constraints.
  \item{A}{Matrix representing the constraints in standard form}
  \item{b}{RHS vector representing the constraints in standard form}
  \item{lb}{Lower bound of the variables}
  \item{obj}{Objective function}
}
\description{
This function is mainly a wrapper of the \code{simplex}
   function from the \code{boot} package. It returns the standard form
   of the constraints matrix in a linear program.
}
