% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.bs.fn}
\alias{subsample.bs.fn}
\title{Carries out one bootstrap replication for the subsampling test}
\usage{
subsample.bs.fn(
  x,
  data,
  lpmodel,
  beta.tgt,
  norm,
  m,
  solver,
  replace,
  n,
  pbar,
  eval.count,
  n.bs,
  progress,
  omega.hat
)
}
\arguments{
\item{x}{This is either the list of indices that represent the bootstrap
replications, or the list of bootstrap components of the \code{lpmodel}
object passed from the user.}

\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{m}{Size of each subsample.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{replace}{Boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{pbar}{Progress bar object.}

\item{eval.count}{Count for the number of times the \code{future_lapply}
function has been called. If this object is zero, it means that the
\code{future_lapply} function is being called for the first time in this
subprocedure. Otherwise, it means that the \code{future_lapply} function
has been called for more than once. This situation typically refers to the
situations where there are some errors in the first time of the
replications.}

\item{n.bs}{Total number of replications to be conducted in this procedure.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{omega.hat}{Estimator of the asymptotic variance.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{Ts}{Bootstrap test statistic.}
  \item{beta}{Bootstrap estimate of \code{beta.obs}.}
  \item{msg}{Error message (if applicable).}
}
\description{
This function carries out the one bootstrap replication of the
  subsampling test. This function is used in the \code{subsample.bs}
  function via the \code{future_lapply} command.
}
