% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.bs}
\alias{subsample.bs}
\title{Bootstrap procedure for the \code{\link[lpinfer]{subsample}} procedure}
\usage{
subsample.bs(
  data,
  i1,
  lpmodel,
  beta.tgt,
  norm,
  solver,
  replace,
  progress,
  m,
  n,
  omega.hat,
  df.error,
  eval.count,
  error.id,
  seed.list
)
}
\arguments{
\item{data}{An \code{data frame} or a \code{matrix}.}

\item{lpmodel}{An \code{lpmodel} object.}

\item{beta.tgt}{The value to be tested.}

\item{norm}{The norm used in the optimization problem. It can be either a
1-norm or a 2-norm. See the details section of
\code{\link[lpinfer]{estbounds}} for a list of accepted strings.}

\item{solver}{The name of the linear and quadratic programming solver that
is used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{replace}{A boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}

\item{m}{The size of each subsample.}

\item{n}{The sample size. This is only required if \code{data} is omitted
in the input.}

\item{omega.hat}{The estimator of the asymptotic variance.}

\item{error.id}{The list of ID that corresponds to problematic bootstrap
draws.}

\item{seed.list}{The list of RNG states and the corresponding iterations.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{T.sub}{The bootstrap test statistics from the subsampling procedure.}
  \item{beta.sub}{The bootstrap estimators for the \code{beta} component.}
  \item{df.error}{A table showing the id of the bootstrap replication(s)
    with error(s) and the corresponding error message(s).}
  \item{R.eval}{The number of bootstrap replications that have been
    conducted.}
  \item{R.succ}{The number of successful bootstrap replications.}
}
\description{
This function carries out the bootstrap procedure of the
  \code{\link[lpinfer]{subsample}} procedure This function supports
  parallel programming via the \code{furrr} package.
}
