% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.check}
\alias{subsample.check}
\title{Checks and updates the input in \code{subsample}}
\usage{
subsample.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  Rmulti,
  solver,
  norm,
  phi,
  n,
  replace,
  progress
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the DKQS test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{beta.obs}}
 }}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstrap replications.}

\item{Rmulti}{Multiplier for the number of bootstrap replications. The
product of \code{Rmulti} and \code{R} refers to the maximum
number of bootstrap replications to be conducted if there are errors.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{phi}{Tuning parameter for the subsampling test. The size of each
subsample is \eqn{n^\phi} where \eqn{\phi \in [0,1]}.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{replace}{Boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{progress}{The boolean variable for whether the progress bars should
be displayed. If it is set as \code{TRUE}, the progress bars will be
displayed while the code is running.}
}
\value{
Returns the updated parameters and objects back to the function
\code{subsample}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{norm}}
      \item{\code{test.logical}}
      \item{\code{logical.lb}}
      \item{\code{logical.ub}}
   }
}
\description{
This function checks and updates the input of the user. If
   there is any invalid input, this function will be terminated and
   generates appropriate error messages. This function is mainly a wrapper
   of the selected functions from the \code{checks} files to conduct the
   checks and updates.
}
