% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.check}
\alias{subsample.check}
\title{Checks and updates the input from \code{subsample}}
\usage{
subsample.check(
  data,
  lpmodel,
  beta.tgt,
  R,
  solver,
  cores,
  norm,
  phi,
  replace,
  progress
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
programming problems. The list of objects required in the \code{dkqs}
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstraps chosen by the users.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop.
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{norm}{Norm used in the optimization problem.}

\item{phi}{Power for the sample. \eqn{n^\phi} represents the size
of each subsample.}

\item{replace}{Boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
Returns the updated parameters and objects back to the function
\code{subsample}. The following information are updated:
   \itemize{
      \item{\code{data}}
      \item{\code{lpmodel}}
      \item{\code{solver}}
      \item{\code{solver.name}}
      \item{\code{cores}}
      \item{\code{norm}}
      \item{\code{test.logical}}
   }
}
\description{
This function checks and updates the input of the user. If
   there is any invalid input, this function will be terminated and
   generates appropriate error messages. This function is mainly a wrapper
   of the selected functions from the \code{checks} files to conduct the
   checks and updates.
}
