% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.onecore}
\alias{subsample.onecore}
\title{Subsampling procedure without parallel programming}
\usage{
subsample.onecore(data, R, lpmodel, beta.tgt, norm, solver, progress, m)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{R}{Number of bootstraps chosen by the users.}

\item{lpmodel}{A list of objects that are used in inference of linear 
programming problems. The list of objects required in the \code{dkqs} 
procedure are:
\itemize{
  \item{\code{A.tgt}}
  \item{\code{A.obs}}
  \item{\code{beta.obs}}
}}

\item{beta.tgt}{Value of beta to be tested.}

\item{norm}{Norm used in the optimization problem.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{progress}{The boolean variable for whether the result messages 
should be displayed in the inference procedure. If it is set as 
\code{TRUE}, the messages are displayed throughout the procedure. 
Otherwise, the messages will not be displayed.}

\item{m}{Size of each subsample.}
}
\value{
Returns a list of output that are obtained from the subsampling
  procedure:
  \item{T.sub}{List of test statistic from the subsampling procedure.}
  \item{pb}{Progress bar object.}
}
\description{
This function computes the list of test statistics that are 
  obtained from the subsampling procedure without using parallel 
  programming.
}
