% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample.prob}
\alias{subsample.prob}
\title{Formulates and solves the subsampling problem}
\usage{
subsample.prob(
  data,
  lpmodel,
  beta.tgt,
  norm,
  solver,
  n,
  beta.obs.hat,
  omega.hat
)
}
\arguments{
\item{data}{Data used in the tests.}

\item{lpmodel}{The \code{lpmodel} object used in the test. The following
components are required in the \code{lpmodel} for the subsampling test:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }
matrix of the estimator \eqn{\hat{\bm{\beta}}_{\mathrm{obs}}}.}

\item{beta.tgt}{Value of beta to be tested.}

\item{norm}{Norm used in the optimization problem. It can be either a 1-norm
or a 2-norm.}

\item{solver}{Name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this package are \code{cplexAPI}, \code{gurobi},
\code{limSolve} and \code{Rcplex}.}

\item{n}{Sample size (only required if \code{data} is omitted in the input).}

\item{beta.obs.hat}{The value of sample \eqn{\hat{\bm{\beta}}_{\mathrm{obs}}}
from the \code{lpmodel} object.}

\item{omega.hat}{Estimator of the asymptotic variance.}
}
\value{
Returns the following list of outputs:
  \item{status}{Status of the optimization problem.}
  \item{x}{Optimal point.}
  \item{objval}{Optimal objective value.}
  \item{larg}{List of arguments passed to the optimizer.}
  \item{beta}{The beta vector \eqn{\widehat{\bm{\beta}}_{\mathrm{obs}}}
    used in the optimization problem that is obtained from the
    \code{beta.obs} component of the \code{lpmodel} object.}
  \item{omega}{The Omega matrix \eqn{\widehat{\bm{\Omega}}_n} used in the
    optimization problem that is obtained from the \code{beta.obs}
    component of the \code{lpmodel} object.}
}
\description{
This function formulates and solves the linear or quadratic
  program in the subsampling procedure. If the user chooses a 1-norm, this
  function solves a linear program. If the user chooses a 2-norm, this
  function solves a quadratic program.
}
