% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Computes the \eqn{p}-value of the subsampling procedure}
\usage{
subsample(
  data,
  A_obs,
  func_obs,
  func_var,
  A_shp,
  beta_shp,
  A_tgt,
  beta_tgt,
  R = 100,
  solver = NULL,
  cores = 8,
  norm = 2,
  phi = 2/3,
  alpha = 0.05,
  progress = FALSE
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{A_obs}{The "observed matrix" in the inference \eqn{A_{\mathrm{obs}}}.}

\item{func_obs}{The function that generates the required 
\eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{func_var}{Function that generates the asymptotic variance 
matrix of the estimator \eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{A_tgt}{The "target matrix" in the inference \eqn{A_{\mathrm{tgt}}}.}

\item{beta_tgt}{The value of \eqn{\hat{\beta}_{\mathrm{tgt}}} (i.e. the 
value of \eqn{t} in the missing data problem) in the null hypothesis.}

\item{R}{The total number of bootstraps \eqn{B} to be conducted.}

\item{solver}{The name of the linear and quadratic programming solver that 
are used to obtain the solution to linear and quadratic programs. 
The solvers supported by this module are `\code{cplexAPI}', 
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop. 
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{norm}{Norm used in the optimization problem.}

\item{phi}{Power for the sample. \eqn{n^\phi} represents the size
of each subsample.}

\item{alpha}{Significance level.}

\item{progress}{The boolean variable for whether the result messages should
be displayed in the inference procedure. If it is set as \code{TRUE}, the 
messages are displayed throughout the procedure. Otherwise, the messages
will not be displayed.}
}
\value{
Returns a list of output calculated from the function:
  \item{p_val}{\eqn{p}-value.}
  \item{decision}{Decision of the test.}
  \item{alpha}{Significance level.}
  \item{T_n}{Test statistic \eqn{T_n}.}
  \item{T_sub}{The list of test statistics from the subsampling procedure.}
  \item{solver}{Solver used in solving the linear and quadratic programs.}
  \item{cores}{Number of cores used.}
  \item{call}{The function that has been called.}
  \item{norm}{Norm used.}
}
\description{
This function conducts inference and returns the 
  \eqn{p}-value using the subsampling procedure.
}
