% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Computes the \eqn{p}-value of the subsampling procedure}
\usage{
subsample(
  data = NULL,
  lpmodel,
  beta.tgt,
  R = 100,
  solver = NULL,
  cores = 1,
  norm = 2,
  phi = 2/3,
  replace = FALSE,
  progress = FALSE
)
}
\arguments{
\item{data}{The data being used in the inference.}

\item{lpmodel}{A list of objects that are used in inference of linear
 programming problems. The list of objects required in the \code{dkqs}
 procedure are:
 \itemize{
   \item{\code{A.tgt}}
   \item{\code{A.obs}}
   \item{\code{A.shp}}
   \item{\code{beta.obs}}
   \item{\code{beta.shp}}
 }
matrix of the estimator \eqn{\hat{\beta}_{\mathrm{obs}}}.}

\item{beta.tgt}{Value of beta to be tested.}

\item{R}{Number of bootstraps chosen by the users.}

\item{solver}{The name of the linear and quadratic programming solver that
are used to obtain the solution to linear and quadratic programs.
The solvers supported by this module are `\code{cplexAPI}',
`\code{gurobi}', `\code{limSolve}' and `\code{Rcplex}'.}

\item{cores}{Number of cores to be used in the parallelized for-loop.
Parallelized for-loop is used if \code{cores} is set to be an integer
greater than or equal to 2.}

\item{norm}{Norm used in the optimization problem.}

\item{phi}{Power for the sample. \eqn{n^\phi} represents the size
of each subsample.}

\item{replace}{Boolean variable to indicate whether the function samples
the data with or without replacement.}

\item{progress}{The boolean variable for whether the result messages
should be displayed in the inference procedure. If it is set as
\code{TRUE}, the messages are displayed throughout the procedure.
Otherwise, the messages will not be displayed.}
}
\value{
Returns a list of output calculated from the function:
  \item{pval}{\eqn{p}-value.}
  \item{decision}{Decision of the test.}
  \item{T.n}{Test statistic \eqn{T_n}.}
  \item{T.sub}{The list of test statistics from the subsampling procedure.}
  \item{solver}{Solver used in solving the linear and quadratic programs.}
  \item{cores}{Number of cores used.}
  \item{call}{The function that has been called.}
  \item{phi}{The \eqn{\phi} parameter used.}
  \item{norm}{Norm used.}
  \item{subsample.size}{Size of subsample}
}
\description{
This function conducts inference and returns the
  \eqn{p}-value using the subsampling procedure.
}
\details{
There are three possible combinations for the parameters 
\code{phi} and \code{replace}:
\itemize{
  \item{If \code{replace} is set as \code{FALSE}, it refers to the 
    subsampling procedure. In this case, \code{phi} has to be in the 
    interval \eqn{(0, 1)}.}
  \item{If \code{replace} is set as \code{TRUE} and \code{phi} is set as 1,
    then it refers to the bootstrap procedure.}
  \item{If \code{replace} is set as \code{TRUE} and \code{phi} is in the 
    interval \eqn{(0, 1)}, then it refers to the \eqn{m} out of \eqn{n} 
    bootstrap procedure, where \eqn{m} is the size of the subsample and 
    \eqn{n} is the total number of observations.}
}
}
