% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.bisection.print}
\alias{summary.bisection.print}
\title{Print results in constructing bounds in bisection method}
\usage{
summary.bisection.print(df_bis, i)
}
\arguments{
\item{df_bis}{Data frame storing the information from the bisection method.}
}
\value{
Nothing is returned in this function.
}
\description{
This function is used to display the message when constructing
   the bounds and used in \code{summary.invertci} to print the results in
   each step of the bisection method.
}
