% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{summary.dkqs}
\alias{summary.dkqs}
\title{Summary of results from \code{dkqs}}
\usage{
\method{summary}{dkqs}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{dkqs}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{dkqs}.
}
\description{
This function uses the summary method on the return list of the
   function \code{dkqs}. This is a wrapper of the \code{print} command.
}
