% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{summary.estbounds}
\alias{summary.estbounds}
\title{Summary of results from \code{estbounds}}
\usage{
summary.estbounds(x, ...)
}
\arguments{
\item{x}{Object returned from \code{estbounds}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned. This function prints results from 
   \code{estbounds}.
}
\description{
This function uses the summary method on the return list of the
   function \code{estbounds}. This is a wrapper of the function
   \code{print.estbounds}.
}
