% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{summary.estbounds}
\alias{summary.estbounds}
\title{Summary of results from \code{\link[lpinfer]{estbounds}}}
\usage{
\method{summary}{estbounds}(x, ...)
}
\arguments{
\item{x}{The output objects returned from \code{\link[lpinfer]{estbounds}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{\link[lpinfer]{estbounds}}.
}
