% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{summary.fsst}
\alias{summary.fsst}
\title{Summary of results from \code{fsst}}
\usage{
summary.fsst(x, ...)
}
\arguments{
\item{x}{Object returned from \code{fsst}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{fsst}.
}
\description{
This function uses the summary method on the return list of the
   function \code{fsst}. This is a wrapper of the \code{print} command.
}
