% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{summary.fsst}
\alias{summary.fsst}
\title{Summary of results from \code{\link[lpinfer]{fsst}}}
\usage{
summary.fsst(x, ...)
}
\arguments{
\item{x}{The output objects returned from \code{\link[lpinfer]{fsst}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  \code{\link[lpinfer]{fsst}}.
}
