% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsst.R
\name{summary.fsst}
\alias{summary.fsst}
\title{Summary of results from \code{fsst}}
\usage{
summary.fsst(x, ...)
}
\arguments{
\item{x}{Object returned from \code{fsst}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned
}
\description{
This function uses the summary method on the return list of the
   function \code{fsst}. This is a wrapper of the \code{print} command.
}
\details{
The following information are printed
  \itemize{
    \item{Test statistic (Max of Cone and Range)}
    \item{Test statistic (Cone)}
    \item{Test statistic (Range)}
    \item{\eqn{p}-value}
    \item{Solver used}
    \item{Number of cores used}}
}
