% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.invertci}
\alias{summary.invertci}
\title{Summary of results from \code{invertci}}
\usage{
summary.invertci(x, alphas = NULL, ...)
}
\arguments{
\item{x}{Object returned from \code{invertci}.}

\item{alphas}{List of alphas that the user would like to print.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{invertci}.
}
\description{
This function uses the print method on the return list of the
   function \code{invertci}.
}
