% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{summary.invertci}
\alias{summary.invertci}
\title{Summary of results from \code{invertci}}
\usage{
\method{summary}{invertci}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{invertci}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{invertci}.
}
\description{
This function uses the print method on the return list of the
   function \code{invertci}.
}
