% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.invertci_single}
\alias{summary.invertci_single}
\title{Summary of results from \code{\link[lpinfer]{invertci}} for a single
significance level}
\usage{
\method{summary}{invertci_single}(x, alphas, msg.bound, ...)
}
\arguments{
\item{x}{The output object returned from \code{\link[lpinfer]{invertci}}.}

\item{alphas}{A list of alphas that the user would like to print.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function summarizes the results for
  \code{\link[lpinfer]{invertci}}.
}
