% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertci.R
\name{summary.invertci_single}
\alias{summary.invertci_single}
\title{Summary of results from \code{invertci} for a single alpha}
\usage{
summary.invertci_single(x, alphas, ...)
}
\arguments{
\item{x}{Object returned from \code{invertci}.}

\item{...}{Additional arguments.}
}
\value{
Print the summary of the basic set of results from \code{invertci}.
}
\description{
This function uses the print method on the return list of the
   function \code{invertci}.
}
