% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpmodel.R
\name{summary.lpmodel}
\alias{summary.lpmodel}
\title{Summary of the \code{lpmodel} object}
\usage{
\method{summary}{lpmodel}(x, ...)
}
\arguments{
\item{x}{The \code{lpmodel} object.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned
}
\description{
This function is a wrapper of the \code{print.lpmodel}
   function and prints the same information for the object \code{lpmodel}.
}
