% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estbounds.R
\name{summary.mincriterion}
\alias{summary.mincriterion}
\title{Summary of results from \code{mincriterion}}
\usage{
\method{summary}{mincriterion}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{mincriterion}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned. This function prints results from
   \code{mincriterion}.
}
\description{
This function uses the summary method on the return list of the
   function \code{mincriterion}.
}
