% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{summary.subsample}
\alias{summary.subsample}
\title{Summary of results from the \code{\link[lpinfer]{subsample}} procedure}
\usage{
summary.subsample(x, ...)
}
\arguments{
\item{x}{An output object returned from \code{\link[lpinfer]{subsample}}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function prints a summary of the results obtained from
  the \code{\link[lpinfer]{subsample}} procedure.
}
