% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{summary.subsample}
\alias{summary.subsample}
\title{Summary of results from \code{subsample}}
\usage{
\method{summary}{subsample}(x, ...)
}
\arguments{
\item{x}{Object returned from \code{subsample}.}

\item{...}{Additional arguments.}
}
\value{
Nothing is returned.
}
\description{
This function uses the summary method on the return list of
   the function \code{subsample}.
}
