% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{tau_constraints}
\alias{tau_constraints}
\title{Auxiliary function to create the constraints for the linear program for tau}
\usage{
tau_constraints(
  length.tau,
  coeff.tau,
  coeff.x,
  ind.x,
  rhs,
  sense,
  lp.lhs.tau,
  lp.rhs.tau,
  lp.sense.tau
)
}
\arguments{
\item{length.tau}{The number of variables in the constraint.}

\item{coeff.tau}{The coefficient in front of tau in the constraint.}

\item{coeff.x}{The coefficient in front of \eqn{x_i} in the constraint.}

\item{ind.x}{The index of \eqn{x_i}, i.e. the value of \eqn{i}.}

\item{rhs}{The RHS vector of the new constraint.}

\item{sense}{The equality or inequality symbol to be used in the new
constraint.}

\item{lp.lhs.tau}{The constraint matrix to be updated.}

\item{lp.rhs.tau}{The RHS vector of the linear constraints to be updated.}

\item{lp.sense.tau}{The sense vector fo the linear constraints to be 
updated.}
}
\value{
Returns the list of matrices that corresponds to the updated 
  constraints:
  \item{lp.lhs.tau}{Upated constraint matrix.}
  \item{lp.rhs.tau}{Update RHS vector.}
  \item{lp.sense.tau}{Update sense for the constraints.}
}
\description{
This function generates the constraints in the linear program
  for computing the value of tau based on the procedure suggested by Kamat 
  (2018), i.e. linear program (6) of Torgovitsky (2019).
}
