% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkqs.R
\name{tau_constraints}
\alias{tau_constraints}
\title{Auxiliary function to create the constraints for the linear program for tau}
\usage{
tau_constraints(
  length_tau,
  coeff_tau,
  coeff_x,
  ind_x,
  rhs,
  sense,
  lp_lhs_tau,
  lp_rhs_tau,
  lp_sense_tau
)
}
\arguments{
\item{length_tau}{The number of variables in the constraint.}

\item{coeff_tau}{The coefficient in front of tau in the constraint.}

\item{coeff_x}{The coefficient in front of \eqn{x_i} in the constraint.}

\item{ind_x}{The index of \eqn{x_i}, i.e. the value of \eqn{i}.}

\item{rhs}{The RHS vector of the new constraint.}

\item{sense}{The equality or inequality symbol to be used in the new
constraint.}

\item{lp_lhs_tau}{The constraint matrix to be updated.}

\item{lp_rhs_tau}{The RHS vector of the linear constraints to be updated.}

\item{lp_sense_tau}{The sense vector fo the linear constraints to be updated}
}
\value{
Returns the list of matrices that corresponds to the updated 
  constraints:
  \item{lp_lhs_tau}{Upated constraint matrix.}
  \item{lp_rhs_tau}{Update RHS vector.}
  \item{lp_sense_tau}{Update sense for the constraints.}
}
\description{
This function generates the constraints in the linear program
  for computing the value of tau based on the procedure suggested by Kamat 
  (2018), i.e. linear program (6) of Torgovitsky (2019).
}
